Network.AWS.IAM.AddClientIdToOpenIdConnectProvider

Description

Adds a new client ID (also known as audience) to the list of client IDs already registered for the specified IAM OpenID Connect (OIDC) provider resource.

This action is idempotent; it does not fail or return an error if you add an existing client ID to the provider.

Creating a Request

addClientIdToOpenIdConnectProvider

Arguments

:: Text
acitoicpOpenIdConnectProviderARN

-> Text
acitoicpClientId

-> AddClientIdToOpenIdConnectProvider

Creates a value of AddClientIdToOpenIdConnectProvider with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acitoicpOpenIdConnectProviderARN - The Amazon Resource Name (ARN) of the IAM OpenID Connect (OIDC) provider resource to add the client ID to. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders action.

acitoicpClientId - The client ID (also known as audience) to add to the IAM OpenID Connect provider resource.

data AddClientIdToOpenIdConnectProvider

See: addClientIdToOpenIdConnectProvider smart constructor.

Request Lenses

acitoicpOpenIdConnectProviderARN :: Lens' AddClientIdToOpenIdConnectProvider Text

The Amazon Resource Name (ARN) of the IAM OpenID Connect (OIDC) provider resource to add the client ID to. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders action.

acitoicpClientId :: Lens' AddClientIdToOpenIdConnectProvider Text

The client ID (also known as audience) to add to the IAM OpenID Connect provider resource.

Destructuring the Response

addClientIdToOpenIdConnectProviderResponse :: AddClientIdToOpenIdConnectProviderResponse

Creates a value of AddClientIdToOpenIdConnectProviderResponse with the minimum fields required to make a request.

data AddClientIdToOpenIdConnectProviderResponse

See: addClientIdToOpenIdConnectProviderResponse smart constructor.