Network.AWS.IAM.AddRoleToInstanceProfile

Description

Adds the specified IAM role to the specified instance profile. An instance profile can contain only one role, and this limit cannot be increased.

For more information about roles, go to Working with Roles . For more information about instance profiles, go to About Instance Profiles .

Creating a Request

addRoleToInstanceProfile

Arguments

:: Text
artipInstanceProfileName

-> Text
artipRoleName

-> AddRoleToInstanceProfile

Creates a value of AddRoleToInstanceProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

artipInstanceProfileName - The name of the instance profile to update. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

artipRoleName - The name of the role to add. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

data AddRoleToInstanceProfile

See: addRoleToInstanceProfile smart constructor.

Request Lenses

artipInstanceProfileName :: Lens' AddRoleToInstanceProfile Text

The name of the instance profile to update. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

artipRoleName :: Lens' AddRoleToInstanceProfile Text

The name of the role to add. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

addRoleToInstanceProfileResponse :: AddRoleToInstanceProfileResponse

Creates a value of AddRoleToInstanceProfileResponse with the minimum fields required to make a request.

data AddRoleToInstanceProfileResponse

See: addRoleToInstanceProfileResponse smart constructor.