Network.AWS.IAM.AttachGroupPolicy

Description

Attaches the specified managed policy to the specified group.

You use this API to attach a managed policy to a group. To embed an inline policy in a group, use PutGroupPolicy.

For more information about policies, refer to Managed Policies and InlinePolicies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_AttachGroupPolicy.html

Request

data AttachGroupPolicy

Request constructor

attachGroupPolicy

Arguments

:: Text
agpGroupName

-> Text
agpPolicyArn

-> AttachGroupPolicy

AttachGroupPolicy constructor.

The fields accessible through corresponding lenses are:

agpGroupName :: Text

agpPolicyArn :: Text

Request lenses

agpGroupName :: Lens' AttachGroupPolicy Text

The name (friendly name, not ARN) of the group to attach the policy to.

agpPolicyArn :: Lens' AttachGroupPolicy Text

Response

data AttachGroupPolicyResponse

Response constructor

attachGroupPolicyResponse :: AttachGroupPolicyResponse

AttachGroupPolicyResponse constructor.