Network.AWS.IAM.AttachUserPolicy

Description

Attaches the specified managed policy to the specified user.

You use this API to attach a managed policy to a user. To embed an inline policy in a user, use PutUserPolicy .

For more information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

attachUserPolicy

Arguments

:: Text
aupUserName

-> Text
aupPolicyARN

-> AttachUserPolicy

Creates a value of AttachUserPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aupUserName - The name (friendly name, not ARN) of the IAM user to attach the policy to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

aupPolicyARN - The Amazon Resource Name (ARN) of the IAM policy you want to attach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data AttachUserPolicy

See: attachUserPolicy smart constructor.

Request Lenses

aupUserName :: Lens' AttachUserPolicy Text

The name (friendly name, not ARN) of the IAM user to attach the policy to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

aupPolicyARN :: Lens' AttachUserPolicy Text

The Amazon Resource Name (ARN) of the IAM policy you want to attach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

attachUserPolicyResponse :: AttachUserPolicyResponse

Creates a value of AttachUserPolicyResponse with the minimum fields required to make a request.

data AttachUserPolicyResponse

See: attachUserPolicyResponse smart constructor.