Network.AWS.IAM.ChangePassword

Description

Changes the password of the IAM user calling ChangePassword. The root account password is not affected by this action. For information about modifying passwords, see Managing Passwords in the Using IAM guide. https:/iam.amazonaws.com ?Action=ChangePassword &OldPassword=U79}kgds4? &NewPassword=Lb0*1(9xpN &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data ChangePassword

Request constructor

mkChangePassword

Arguments

:: Text
cpOldPassword

-> Text
cpNewPassword

-> ChangePassword

Smart constructor for the minimum required parameters to construct a valid ChangePassword request.

The fields accessible through corresponding lenses are:

OldPassword :: Text

NewPassword :: Text

Request lenses

cpOldPassword :: Lens' ChangePassword Text

The IAM users's current password.

cpNewPassword :: Lens' ChangePassword Text

The new password. The new password must conform to the AWS account's password policy, if one exists.

Response

data ChangePasswordResponse

Response constructor

mkChangePasswordResponse :: ChangePasswordResponse

Smart constructor for the minimum required parameters to construct a valid ChangePasswordResponse response.

This constructor is provided for convenience and testing purposes.