Network.AWS.IAM.CreateAccessKey

Description

Creates a new AWS secret access key and corresponding AWS access key ID for the specified user. The default status for new keys is Active.

If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

For information about limits on the number of keys you can create, see Limitations on IAM Entities in the Using IAM guide.

To ensure the security of your AWS account, the secret access key is accessible only during key and user creation. You must save the key (for example, in a text file) if you want to be able to access it again. If a secret key is lost, you can delete the access keys for the associated user and then create new keys.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateAccessKey.html

Request

data CreateAccessKey

Request constructor

createAccessKey :: CreateAccessKey

CreateAccessKey constructor.

The fields accessible through corresponding lenses are:

cakUserName :: Maybe Text

Request lenses

cakUserName :: Lens' CreateAccessKey (Maybe Text)

The user name that the new key will belong to.

Response

data CreateAccessKeyResponse

Response constructor

createAccessKeyResponse

Arguments

:: AccessKey
cakrAccessKey

-> CreateAccessKeyResponse

CreateAccessKeyResponse constructor.

The fields accessible through corresponding lenses are:

cakrAccessKey :: AccessKey

Response lenses

cakrAccessKey :: Lens' CreateAccessKeyResponse AccessKey

Information about the access key.