Network.AWS.IAM.CreateInstanceProfile

Description

Creates a new instance profile. For information about instance profiles, go to About Instance Profiles.

For information about the number of instance profiles you can create, see Limitations on IAM Entities in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateInstanceProfile.html

Request

data CreateInstanceProfile

Request constructor

createInstanceProfile

Arguments

:: Text
cipInstanceProfileName

-> CreateInstanceProfile

CreateInstanceProfile constructor.

The fields accessible through corresponding lenses are:

cipInstanceProfileName :: Text

cipPath :: Maybe Text

Request lenses

cipInstanceProfileName :: Lens' CreateInstanceProfile Text

The name of the instance profile to create.

cipPath :: Lens' CreateInstanceProfile (Maybe Text)

The path to the instance profile. For more information about paths, see IAMIdentifiers in the Using IAM guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

Response

data CreateInstanceProfileResponse

Response constructor

createInstanceProfileResponse

Arguments

:: InstanceProfile
ciprInstanceProfile

-> CreateInstanceProfileResponse

CreateInstanceProfileResponse constructor.

The fields accessible through corresponding lenses are:

ciprInstanceProfile :: InstanceProfile

Response lenses

ciprInstanceProfile :: Lens' CreateInstanceProfileResponse InstanceProfile

Information about the instance profile.