Network.AWS.IAM.CreateLoginProfile

Description

Creates a password for the specified user, giving the user the ability to access AWS services through the AWS Management Console. For more information about managing passwords, see Managing Passwords in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateLoginProfile.html

Request

data CreateLoginProfile

Request constructor

createLoginProfile

Arguments

:: Text
clpUserName

-> Text
clpPassword

-> CreateLoginProfile

CreateLoginProfile constructor.

The fields accessible through corresponding lenses are:

clpPassword :: Text

clpPasswordResetRequired :: Maybe Bool

clpUserName :: Text

Request lenses

clpPassword :: Lens' CreateLoginProfile Text

The new password for the user.

clpPasswordResetRequired :: Lens' CreateLoginProfile (Maybe Bool)

Specifies whether the user is required to set a new password on next sign-in.

clpUserName :: Lens' CreateLoginProfile Text

The name of the user to create a password for.

Response

data CreateLoginProfileResponse

Response constructor

createLoginProfileResponse

Arguments

:: LoginProfile
clprLoginProfile

-> CreateLoginProfileResponse

CreateLoginProfileResponse constructor.

The fields accessible through corresponding lenses are:

clprLoginProfile :: LoginProfile

Response lenses

clprLoginProfile :: Lens' CreateLoginProfileResponse LoginProfile

The user name and password create date.