Network.AWS.IAM.CreatePolicy

Description

Creates a new managed policy for your AWS account.

This operation creates a policy version with a version identifier of v1 and sets v1 as the policy's default version. For more information about policy versions, see Versioning for Managed Policies in the Using IAM guide.

For more information about managed policies in general, refer to ManagedPolicies and Inline Policies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreatePolicy.html

Request

data CreatePolicy

Request constructor

createPolicy

Arguments

:: Text
cpPolicyName

-> Text
cpPolicyDocument

-> CreatePolicy

CreatePolicy constructor.

The fields accessible through corresponding lenses are:

cpDescription :: Maybe Text

cpPath :: Maybe Text

cpPolicyDocument :: Text

cpPolicyName :: Text

Request lenses

cpDescription :: Lens' CreatePolicy (Maybe Text)

A friendly description of the policy.

Typically used to store information about the permissions defined in the policy. For example, "Grants access to production DynamoDB tables."

The policy description is immutable. After a value is assigned, it cannot be changed.

cpPath :: Lens' CreatePolicy (Maybe Text)

The path for the policy.

For more information about paths, see IAM Identifiers in the Using IAM guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

cpPolicyDocument :: Lens' CreatePolicy Text

The policy document.

The policy must be URL-encoded according to RFC 3986.

cpPolicyName :: Lens' CreatePolicy Text

The name of the policy document.

Response

data CreatePolicyResponse

Response constructor

createPolicyResponse :: CreatePolicyResponse

CreatePolicyResponse constructor.

The fields accessible through corresponding lenses are:

cprPolicy :: Maybe Policy

Response lenses

cprPolicy :: Lens' CreatePolicyResponse (Maybe Policy)

Information about the policy.