Network.AWS.IAM.CreatePolicyVersion

Description

Creates a new version of the specified managed policy. To update a managed policy, you create a new policy version. A managed policy can have up to five versions. If the policy has five versions, you must delete an existing version using DeletePolicyVersion before you create a new version.

Optionally, you can set the new version as the policy's default version. The default version is the operative version; that is, the version that is in effect for the IAM users, groups, and roles that the policy is attached to.

For more information about managed policy versions, see Versioning forManaged Policies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreatePolicyVersion.html

Request

data CreatePolicyVersion

Request constructor

createPolicyVersion

Arguments

:: Text
cpvPolicyArn

-> Text
cpvPolicyDocument

-> CreatePolicyVersion

CreatePolicyVersion constructor.

The fields accessible through corresponding lenses are:

cpvPolicyArn :: Text

cpvPolicyDocument :: Text

cpvSetAsDefault :: Maybe Bool

Request lenses

cpvPolicyArn :: Lens' CreatePolicyVersion Text

cpvPolicyDocument :: Lens' CreatePolicyVersion Text

The policy document.

The policy must be URL-encoded according to RFC 3986.

cpvSetAsDefault :: Lens' CreatePolicyVersion (Maybe Bool)

Specifies whether to set this version as the policy's default version.

When this parameter is true, the new policy version becomes the operative version; that is, the version that is in effect for the IAM users, groups, and roles that the policy is attached to.

For more information about managed policy versions, see Versioning forManaged Policies in the Using IAM guide.

Response

data CreatePolicyVersionResponse

Response constructor

createPolicyVersionResponse :: CreatePolicyVersionResponse

CreatePolicyVersionResponse constructor.

The fields accessible through corresponding lenses are:

cpvrPolicyVersion :: Maybe PolicyVersion

Response lenses

cpvrPolicyVersion :: Lens' CreatePolicyVersionResponse (Maybe PolicyVersion)

Information about the policy version.