Network.AWS.IAM.CreateUser

Description

Creates a new IAM user for your AWS account.

For information about limitations on the number of IAM users you can create, see Limitations on IAM Entities in the IAM User Guide .

Creating a Request

createUser

Arguments

:: Text
cuUserName

-> CreateUser

Creates a value of CreateUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cuPath - The path for the user name. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cuUserName - The name of the user to create. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-. User names are not distinguished by case. For example, you cannot create users named both TESTUSER and "testuser".

data CreateUser

See: createUser smart constructor.

Request Lenses

cuPath :: Lens' CreateUser (Maybe Text)

The path for the user name. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cuUserName :: Lens' CreateUser Text

The name of the user to create. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-. User names are not distinguished by case. For example, you cannot create users named both TESTUSER and "testuser".

Destructuring the Response

createUserResponse

Arguments

:: Int
cursResponseStatus

-> CreateUserResponse

Creates a value of CreateUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cursUser - A structure with details about the new IAM user.

cursResponseStatus - -- | The response status code.

data CreateUserResponse

Contains the response to a successful CreateUser request.

See: createUserResponse smart constructor.

Response Lenses

cursUser :: Lens' CreateUserResponse (Maybe User)

A structure with details about the new IAM user.

cursResponseStatus :: Lens' CreateUserResponse Int

- | The response status code.