Network.AWS.IAM.CreateVirtualMFADevice

Description

Creates a new virtual MFA device for the AWS account. After creating the virtual MFA, use EnableMFADevice to attach the MFA device to an IAM user. For more information about creating and working with virtual MFA devices, go to Using a Virtual MFA Device in the IAM User Guide .

For information about limits on the number of MFA devices you can create, see Limitations on Entities in the IAM User Guide .

Important: The seed information contained in the QR code and the Base32 string should be treated like any other secret access information, such as your AWS access keys or your passwords. After you provision your virtual device, you should ensure that the information is destroyed following secure procedures.

Creating a Request

createVirtualMFADevice

Arguments

:: Text
cvmdVirtualMFADeviceName

-> CreateVirtualMFADevice

Creates a value of CreateVirtualMFADevice with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvmdPath - The path for the virtual MFA device. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cvmdVirtualMFADeviceName - The name of the virtual MFA device. Use with path to uniquely identify a virtual MFA device. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data CreateVirtualMFADevice

See: createVirtualMFADevice smart constructor.

Request Lenses

cvmdPath :: Lens' CreateVirtualMFADevice (Maybe Text)

The path for the virtual MFA device. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cvmdVirtualMFADeviceName :: Lens' CreateVirtualMFADevice Text

The name of the virtual MFA device. Use with path to uniquely identify a virtual MFA device. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

createVirtualMFADeviceResponse

Arguments

:: Int
cvmdrsResponseStatus

-> VirtualMFADevice
cvmdrsVirtualMFADevice

-> CreateVirtualMFADeviceResponse

Creates a value of CreateVirtualMFADeviceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvmdrsResponseStatus - -- | The response status code.

cvmdrsVirtualMFADevice - A structure containing details about the new virtual MFA device.

data CreateVirtualMFADeviceResponse

Contains the response to a successful CreateVirtualMFADevice request.

See: createVirtualMFADeviceResponse smart constructor.

Response Lenses

cvmdrsResponseStatus :: Lens' CreateVirtualMFADeviceResponse Int

- | The response status code.

cvmdrsVirtualMFADevice :: Lens' CreateVirtualMFADeviceResponse VirtualMFADevice

A structure containing details about the new virtual MFA device.