Network.AWS.IAM.DeleteAccessKey

Description

Deletes the access key pair associated with the specified IAM user.

If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

Creating a Request

deleteAccessKey

Arguments

:: AccessKey
dakAccessKeyId

-> DeleteAccessKey

Creates a value of DeleteAccessKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dakUserName - The name of the user whose access key pair you want to delete. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

dakAccessKeyId - The access key ID for the access key ID and secret access key you want to delete. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

data DeleteAccessKey

See: deleteAccessKey smart constructor.

Request Lenses

dakUserName :: Lens' DeleteAccessKey (Maybe Text)

The name of the user whose access key pair you want to delete. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

dakAccessKeyId :: Lens' DeleteAccessKey AccessKey

The access key ID for the access key ID and secret access key you want to delete. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

Destructuring the Response

deleteAccessKeyResponse :: DeleteAccessKeyResponse

Creates a value of DeleteAccessKeyResponse with the minimum fields required to make a request.

data DeleteAccessKeyResponse

See: deleteAccessKeyResponse smart constructor.