Network.AWS.IAM.DeleteAccountPasswordPolicy

Description

Deletes the password policy for the AWS account. There are no parameters.

Creating a Request

deleteAccountPasswordPolicy :: DeleteAccountPasswordPolicy

Creates a value of DeleteAccountPasswordPolicy with the minimum fields required to make a request.

data DeleteAccountPasswordPolicy

See: deleteAccountPasswordPolicy smart constructor.

Destructuring the Response

deleteAccountPasswordPolicyResponse :: DeleteAccountPasswordPolicyResponse

Creates a value of DeleteAccountPasswordPolicyResponse with the minimum fields required to make a request.

data DeleteAccountPasswordPolicyResponse

See: deleteAccountPasswordPolicyResponse smart constructor.