Network.AWS.IAM.DeleteGroupPolicy

Description

Deletes the specified inline policy that is embedded in the specified group.

A group can also have managed policies attached to it. To detach a managed policy from a group, use DetachGroupPolicy. For more information about policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_DeleteGroupPolicy.html

Request

data DeleteGroupPolicy

Request constructor

deleteGroupPolicy

Arguments

:: Text
dgp1GroupName

-> Text
dgp1PolicyName

-> DeleteGroupPolicy

DeleteGroupPolicy constructor.

The fields accessible through corresponding lenses are:

dgp1GroupName :: Text

dgp1PolicyName :: Text

Request lenses

dgp1GroupName :: Lens' DeleteGroupPolicy Text

The name (friendly name, not ARN) identifying the group that the policy is embedded in.

dgp1PolicyName :: Lens' DeleteGroupPolicy Text

The name identifying the policy document to delete.

Response

data DeleteGroupPolicyResponse

Response constructor

deleteGroupPolicyResponse :: DeleteGroupPolicyResponse

DeleteGroupPolicyResponse constructor.