Network.AWS.IAM.DeleteInstanceProfile

Description

Deletes the specified instance profile. The instance profile must not have an associated role.

Make sure you do not have any Amazon EC2 instances running with the instance profile you are about to delete. Deleting a role or instance profile that is associated with a running instance will break any applications running on the instance. For more information about instance profiles, go to About Instance Profiles.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_DeleteInstanceProfile.html

Request

data DeleteInstanceProfile

Request constructor

deleteInstanceProfile

Arguments

:: Text
dipInstanceProfileName

-> DeleteInstanceProfile

DeleteInstanceProfile constructor.

The fields accessible through corresponding lenses are:

dipInstanceProfileName :: Text

Request lenses

dipInstanceProfileName :: Lens' DeleteInstanceProfile Text

The name of the instance profile to delete.

Response

data DeleteInstanceProfileResponse

Response constructor

deleteInstanceProfileResponse :: DeleteInstanceProfileResponse

DeleteInstanceProfileResponse constructor.