Network.AWS.IAM.DeleteOpenIDConnectProvider

Description

Deletes an IAM OpenID Connect identity provider.

Deleting an OIDC provider does not update any roles that reference the provider as a principal in their trust policies. Any attempt to assume a role that references a provider that has been deleted will fail.

This action is idempotent; it does not fail or return an error if you call the action for a provider that was already deleted.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_DeleteOpenIDConnectProvider.html

Request

data DeleteOpenIDConnectProvider

Request constructor

deleteOpenIDConnectProvider

Arguments

:: Text
doidcpOpenIDConnectProviderArn

-> DeleteOpenIDConnectProvider

DeleteOpenIDConnectProvider constructor.

The fields accessible through corresponding lenses are:

doidcpOpenIDConnectProviderArn :: Text

Request lenses

doidcpOpenIDConnectProviderArn :: Lens' DeleteOpenIDConnectProvider Text

The Amazon Resource Name (ARN) of the IAM OpenID Connect provider to delete. You can get a list of OpenID Connect provider ARNs by using the ListOpenIDConnectProviders action.

Response

data DeleteOpenIDConnectProviderResponse

Response constructor

deleteOpenIDConnectProviderResponse :: DeleteOpenIDConnectProviderResponse

DeleteOpenIDConnectProviderResponse constructor.