Network.AWS.IAM.DeleteOpenIdConnectProvider

Description

Deletes an OpenID Connect identity provider (IdP) resource object in IAM.

Deleting an IAM OIDC provider resource does not update any roles that reference the provider as a principal in their trust policies. Any attempt to assume a role that references a deleted provider fails.

This action is idempotent; it does not fail or return an error if you call the action for a provider that does not exist.

Creating a Request

deleteOpenIdConnectProvider

Arguments

:: Text
doicpOpenIdConnectProviderARN

-> DeleteOpenIdConnectProvider

Creates a value of DeleteOpenIdConnectProvider with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

doicpOpenIdConnectProviderARN - The Amazon Resource Name (ARN) of the IAM OpenID Connect provider resource object to delete. You can get a list of OpenID Connect provider resource ARNs by using the ListOpenIDConnectProviders action.

data DeleteOpenIdConnectProvider

See: deleteOpenIdConnectProvider smart constructor.

Request Lenses

doicpOpenIdConnectProviderARN :: Lens' DeleteOpenIdConnectProvider Text

The Amazon Resource Name (ARN) of the IAM OpenID Connect provider resource object to delete. You can get a list of OpenID Connect provider resource ARNs by using the ListOpenIDConnectProviders action.

Destructuring the Response

deleteOpenIdConnectProviderResponse :: DeleteOpenIdConnectProviderResponse

Creates a value of DeleteOpenIdConnectProviderResponse with the minimum fields required to make a request.

data DeleteOpenIdConnectProviderResponse

See: deleteOpenIdConnectProviderResponse smart constructor.