Network.AWS.IAM.DeletePolicy

Description

Deletes the specified managed policy.

Before you can delete a managed policy, you must detach the policy from all users, groups, and roles that it is attached to, and you must delete all of the policy's versions. The following steps describe the process for deleting a managed policy: Detach the policy from all users, groups, and roles that the policy is attached to, using the DetachUserPolicy, DetachGroupPolicy, or DetachRolePolicy APIs. To list all the users, groups, and roles that a policy is attached to, use ListEntitiesForPolicy. Delete all versions of the policy using DeletePolicyVersion. To list the policy's versions, use ListPolicyVersions. You cannot use DeletePolicyVersion to delete the version that is marked as the default version. You delete the policy's default version in the next step of the process. Delete the policy (this automatically deletes the policy's default version) using this API.

For information about managed policies, refer to Managed Policies and InlinePolicies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_DeletePolicy.html

Request

data DeletePolicy

Request constructor

deletePolicy

Arguments

:: Text
dpPolicyArn

-> DeletePolicy

DeletePolicy constructor.

The fields accessible through corresponding lenses are:

dpPolicyArn :: Text

Request lenses

dpPolicyArn :: Lens' DeletePolicy Text

Response

data DeletePolicyResponse

Response constructor

deletePolicyResponse :: DeletePolicyResponse

DeletePolicyResponse constructor.