Network.AWS.IAM.DeletePolicyVersion

Description

Deletes the specified version of the specified managed policy.

You cannot delete the default version of a policy using this API. To delete the default version of a policy, use DeletePolicy. To find out which version of a policy is marked as the default version, use ListPolicyVersions.

For information about versions for managed policies, refer to Versioning forManaged Policies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_DeletePolicyVersion.html

Request

data DeletePolicyVersion

Request constructor

deletePolicyVersion

Arguments

:: Text
dpvPolicyArn

-> Text
dpvVersionId

-> DeletePolicyVersion

DeletePolicyVersion constructor.

The fields accessible through corresponding lenses are:

dpvPolicyArn :: Text

dpvVersionId :: Text

Request lenses

dpvPolicyArn :: Lens' DeletePolicyVersion Text

dpvVersionId :: Lens' DeletePolicyVersion Text

The policy version to delete.

For more information about managed policy versions, see Versioning forManaged Policies in the Using IAM guide.

Response

data DeletePolicyVersionResponse

Response constructor

deletePolicyVersionResponse :: DeletePolicyVersionResponse

DeletePolicyVersionResponse constructor.