Network.AWS.IAM.DeleteRole

Description

Deletes the specified role. The role must not have any policies attached. For more information about roles, go to Working with Roles .

Important: Make sure you do not have any Amazon EC2 instances running with the role you are about to delete. Deleting a role or instance profile that is associated with a running instance will break any applications running on the instance.

Creating a Request

deleteRole

Arguments

:: Text
drRoleName

-> DeleteRole

Creates a value of DeleteRole with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drRoleName - The name of the role to delete. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

data DeleteRole

See: deleteRole smart constructor.

Request Lenses

drRoleName :: Lens' DeleteRole Text

The name of the role to delete. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

deleteRoleResponse :: DeleteRoleResponse

Creates a value of DeleteRoleResponse with the minimum fields required to make a request.

data DeleteRoleResponse

See: deleteRoleResponse smart constructor.