Network.AWS.IAM.DeleteRolePolicy

Description

Deletes the specified inline policy that is embedded in the specified IAM role.

A role can also have managed policies attached to it. To detach a managed policy from a role, use DetachRolePolicy . For more information about policies, refer to Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

deleteRolePolicy

Arguments

:: Text
delRoleName

-> Text
delPolicyName

-> DeleteRolePolicy

Creates a value of DeleteRolePolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delRoleName - The name (friendly name, not ARN) identifying the role that the policy is embedded in. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

delPolicyName - The name of the inline policy to delete from the specified IAM role. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data DeleteRolePolicy

See: deleteRolePolicy smart constructor.

Request Lenses

delRoleName :: Lens' DeleteRolePolicy Text

The name (friendly name, not ARN) identifying the role that the policy is embedded in. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

delPolicyName :: Lens' DeleteRolePolicy Text

The name of the inline policy to delete from the specified IAM role. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

deleteRolePolicyResponse :: DeleteRolePolicyResponse

Creates a value of DeleteRolePolicyResponse with the minimum fields required to make a request.

data DeleteRolePolicyResponse

See: deleteRolePolicyResponse smart constructor.