Network.AWS.IAM.DeleteSAMLProvider

Description

Deletes a SAML provider.

Deleting the provider does not update any roles that reference the SAML provider as a principal in their trust policies. Any attempt to assume a role that references a SAML provider that has been deleted will fail.

This operation requires Signature Version 4.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_DeleteSAMLProvider.html

Request

data DeleteSAMLProvider

Request constructor

deleteSAMLProvider

Arguments

:: Text
dsamlpSAMLProviderArn

-> DeleteSAMLProvider

DeleteSAMLProvider constructor.

The fields accessible through corresponding lenses are:

dsamlpSAMLProviderArn :: Text

Request lenses

dsamlpSAMLProviderArn :: Lens' DeleteSAMLProvider Text

The Amazon Resource Name (ARN) of the SAML provider to delete.

Response

data DeleteSAMLProviderResponse

Response constructor

deleteSAMLProviderResponse :: DeleteSAMLProviderResponse

DeleteSAMLProviderResponse constructor.