Network.AWS.IAM.DeleteSigningCertificate

Description

Deletes the specified signing certificate associated with the specified user.

If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_DeleteSigningCertificate.html

Request

data DeleteSigningCertificate

Request constructor

deleteSigningCertificate

Arguments

:: Text
dscCertificateId

-> DeleteSigningCertificate

DeleteSigningCertificate constructor.

The fields accessible through corresponding lenses are:

dscCertificateId :: Text

dscUserName :: Maybe Text

Request lenses

dscCertificateId :: Lens' DeleteSigningCertificate Text

The ID of the signing certificate to delete.

dscUserName :: Lens' DeleteSigningCertificate (Maybe Text)

The name of the user the signing certificate belongs to.

Response

data DeleteSigningCertificateResponse

Response constructor

deleteSigningCertificateResponse :: DeleteSigningCertificateResponse

DeleteSigningCertificateResponse constructor.