Network.AWS.IAM.DeleteUserPolicy

Description

Deletes the specified inline policy that is embedded in the specified IAM user.

A user can also have managed policies attached to it. To detach a managed policy from a user, use DetachUserPolicy . For more information about policies, refer to Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

deleteUserPolicy

Arguments

:: Text
dupUserName

-> Text
dupPolicyName

-> DeleteUserPolicy

Creates a value of DeleteUserPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dupUserName - The name (friendly name, not ARN) identifying the user that the policy is embedded in. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

dupPolicyName - The name identifying the policy document to delete. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data DeleteUserPolicy

See: deleteUserPolicy smart constructor.

Request Lenses

dupUserName :: Lens' DeleteUserPolicy Text

The name (friendly name, not ARN) identifying the user that the policy is embedded in. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

dupPolicyName :: Lens' DeleteUserPolicy Text

The name identifying the policy document to delete. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

deleteUserPolicyResponse :: DeleteUserPolicyResponse

Creates a value of DeleteUserPolicyResponse with the minimum fields required to make a request.

data DeleteUserPolicyResponse

See: deleteUserPolicyResponse smart constructor.