Network.AWS.IAM.DetachGroupPolicy

Description

Removes the specified managed policy from the specified group.

A group can also have inline policies embedded with it. To delete an inline policy, use the DeleteGroupPolicy API. For information about policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_DetachGroupPolicy.html

Request

data DetachGroupPolicy

Request constructor

detachGroupPolicy

Arguments

:: Text
dgpGroupName

-> Text
dgpPolicyArn

-> DetachGroupPolicy

DetachGroupPolicy constructor.

The fields accessible through corresponding lenses are:

dgpGroupName :: Text

dgpPolicyArn :: Text

Request lenses

dgpGroupName :: Lens' DetachGroupPolicy Text

The name (friendly name, not ARN) of the group to detach the policy from.

dgpPolicyArn :: Lens' DetachGroupPolicy Text

Response

data DetachGroupPolicyResponse

Response constructor

detachGroupPolicyResponse :: DetachGroupPolicyResponse

DetachGroupPolicyResponse constructor.