Network.AWS.IAM.DetachRolePolicy

Description

Removes the specified managed policy from the specified role.

A role can also have inline policies embedded with it. To delete an inline policy, use the DeleteRolePolicy API. For information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

detachRolePolicy

Arguments

:: Text
drpRoleName

-> Text
drpPolicyARN

-> DetachRolePolicy

Creates a value of DetachRolePolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drpRoleName - The name (friendly name, not ARN) of the IAM role to detach the policy from. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

drpPolicyARN - The Amazon Resource Name (ARN) of the IAM policy you want to detach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data DetachRolePolicy

See: detachRolePolicy smart constructor.

Request Lenses

drpRoleName :: Lens' DetachRolePolicy Text

The name (friendly name, not ARN) of the IAM role to detach the policy from. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

drpPolicyARN :: Lens' DetachRolePolicy Text

The Amazon Resource Name (ARN) of the IAM policy you want to detach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

detachRolePolicyResponse :: DetachRolePolicyResponse

Creates a value of DetachRolePolicyResponse with the minimum fields required to make a request.

data DetachRolePolicyResponse

See: detachRolePolicyResponse smart constructor.