Network.AWS.IAM.DetachUserPolicy

Description

Removes the specified managed policy from the specified user.

A user can also have inline policies embedded with it. To delete an inline policy, use the DeleteUserPolicy API. For information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

detachUserPolicy

Arguments

:: Text
dUserName

-> Text
dPolicyARN

-> DetachUserPolicy

Creates a value of DetachUserPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dUserName - The name (friendly name, not ARN) of the IAM user to detach the policy from. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

dPolicyARN - The Amazon Resource Name (ARN) of the IAM policy you want to detach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data DetachUserPolicy

See: detachUserPolicy smart constructor.

Request Lenses

dUserName :: Lens' DetachUserPolicy Text

The name (friendly name, not ARN) of the IAM user to detach the policy from. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

dPolicyARN :: Lens' DetachUserPolicy Text

The Amazon Resource Name (ARN) of the IAM policy you want to detach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

detachUserPolicyResponse :: DetachUserPolicyResponse

Creates a value of DetachUserPolicyResponse with the minimum fields required to make a request.

data DetachUserPolicyResponse

See: detachUserPolicyResponse smart constructor.