Network.AWS.IAM.EnableMFADevice

Description

Enables the specified MFA device and associates it with the specified user name. When enabled, the MFA device is required for every subsequent login by the user name associated with the device.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_EnableMFADevice.html

Request

data EnableMFADevice

Request constructor

enableMFADevice

Arguments

:: Text
emfadUserName

-> Text
emfadSerialNumber

-> Text
emfadAuthenticationCode1

-> Text
emfadAuthenticationCode2

-> EnableMFADevice

EnableMFADevice constructor.

The fields accessible through corresponding lenses are:

emfadAuthenticationCode1 :: Text

emfadAuthenticationCode2 :: Text

emfadSerialNumber :: Text

emfadUserName :: Text

Request lenses

emfadAuthenticationCode1 :: Lens' EnableMFADevice Text

An authentication code emitted by the device.

emfadAuthenticationCode2 :: Lens' EnableMFADevice Text

A subsequent authentication code emitted by the device.

emfadSerialNumber :: Lens' EnableMFADevice Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

emfadUserName :: Lens' EnableMFADevice Text

The name of the user for whom you want to enable the MFA device.

Response

data EnableMFADeviceResponse

Response constructor

enableMFADeviceResponse :: EnableMFADeviceResponse

EnableMFADeviceResponse constructor.