Network.AWS.IAM.GenerateCredentialReport

Description

Generates a credential report for the AWS account. For more information about the credential report, see Getting Credential Reports in the Using IAM guide.

Request

data GenerateCredentialReport

Request constructor

mkGenerateCredentialReport :: GenerateCredentialReport

Smart constructor for the minimum required parameters to construct a valid GenerateCredentialReport request.

Response

data GenerateCredentialReportResponse

Contains the result of a successful invocation of the GenerateCredentialReport action.

Response constructor

mkGenerateCredentialReportResponse :: GenerateCredentialReportResponse

Smart constructor for the minimum required parameters to construct a valid GenerateCredentialReportResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

State :: Maybe ReportStateType

Description :: Maybe Text

Response lenses

gcrrState :: Lens' GenerateCredentialReportResponse (Maybe ReportStateType)

Information about the state of a credential report.

gcrrDescription :: Lens' GenerateCredentialReportResponse (Maybe Text)

Information about the credential report.