Network.AWS.IAM.GetAccessKeyLastUsed

Description

Retrieves information about when the specified access key was last used. The information includes the date and time of last use, along with the AWS service and region that were specified in the last request made with that key.

Creating a Request

getAccessKeyLastUsed

Arguments

:: AccessKey
gakluAccessKeyId

-> GetAccessKeyLastUsed

Creates a value of GetAccessKeyLastUsed with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gakluAccessKeyId - The identifier of an access key. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

data GetAccessKeyLastUsed

See: getAccessKeyLastUsed smart constructor.

Request Lenses

gakluAccessKeyId :: Lens' GetAccessKeyLastUsed AccessKey

The identifier of an access key. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

Destructuring the Response

getAccessKeyLastUsedResponse

Arguments

:: Int
gaklursResponseStatus

-> GetAccessKeyLastUsedResponse

Creates a value of GetAccessKeyLastUsedResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaklursUserName - The name of the AWS IAM user that owns this access key.

gaklursAccessKeyLastUsed - Contains information about the last time the access key was used.

gaklursResponseStatus - -- | The response status code.

data GetAccessKeyLastUsedResponse

Contains the response to a successful GetAccessKeyLastUsed request. It is also returned as a member of the AccessKeyMetaData structure returned by the ListAccessKeys action.

See: getAccessKeyLastUsedResponse smart constructor.

Response Lenses

gaklursUserName :: Lens' GetAccessKeyLastUsedResponse (Maybe Text)

The name of the AWS IAM user that owns this access key.

gaklursAccessKeyLastUsed :: Lens' GetAccessKeyLastUsedResponse (Maybe AccessKeyLastUsed)

Contains information about the last time the access key was used.

gaklursResponseStatus :: Lens' GetAccessKeyLastUsedResponse Int

- | The response status code.