Network.AWS.IAM.GetAccountAuthorizationDetails

Description

Retrieves information about all IAM users, groups, and roles in your account, including their relationships to one another and their policies. Use this API to obtain a snapshot of the configuration of IAM permissions (users, groups, roles, and policies) in your account.

You can optionally filter the results using the Filter parameter. You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetAccountAuthorizationDetails.html

Request

data GetAccountAuthorizationDetails

Request constructor

getAccountAuthorizationDetails :: GetAccountAuthorizationDetails

GetAccountAuthorizationDetails constructor.

The fields accessible through corresponding lenses are:

gaadFilter :: [EntityType]

gaadMarker :: Maybe Text

gaadMaxItems :: Maybe Natural

Request lenses

gaadFilter :: Lens' GetAccountAuthorizationDetails [EntityType]

A list of entity types (user, group, role, local managed policy, or AWS managed policy) for filtering the results.

gaadMarker :: Lens' GetAccountAuthorizationDetails (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

gaadMaxItems :: Lens' GetAccountAuthorizationDetails (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If there are additional items beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data GetAccountAuthorizationDetailsResponse

Response constructor

getAccountAuthorizationDetailsResponse :: GetAccountAuthorizationDetailsResponse

GetAccountAuthorizationDetailsResponse constructor.

The fields accessible through corresponding lenses are:

gaadrGroupDetailList :: [GroupDetail]

gaadrIsTruncated :: Maybe Bool

gaadrMarker :: Maybe Text

gaadrPolicies :: [ManagedPolicyDetail]

gaadrRoleDetailList :: [RoleDetail]

gaadrUserDetailList :: [UserDetail]

Response lenses

gaadrGroupDetailList :: Lens' GetAccountAuthorizationDetailsResponse [GroupDetail]

A list containing information about IAM groups.

gaadrIsTruncated :: Lens' GetAccountAuthorizationDetailsResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items.

gaadrMarker :: Lens' GetAccountAuthorizationDetailsResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

gaadrPolicies :: Lens' GetAccountAuthorizationDetailsResponse [ManagedPolicyDetail]

A list containing information about managed policies.

gaadrRoleDetailList :: Lens' GetAccountAuthorizationDetailsResponse [RoleDetail]

A list containing information about IAM roles.

gaadrUserDetailList :: Lens' GetAccountAuthorizationDetailsResponse [UserDetail]

A list containing information about IAM users.