Network.AWS.IAM.GetAccountSummary

Description

Retrieves account level information about account entity usage and IAM quotas.

For information about limitations on IAM entities, see Limitations on IAMEntities in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetAccountSummary.html

Request

data GetAccountSummary

Request constructor

getAccountSummary :: GetAccountSummary

GetAccountSummary constructor.

Response

data GetAccountSummaryResponse

Response constructor

getAccountSummaryResponse :: GetAccountSummaryResponse

GetAccountSummaryResponse constructor.

The fields accessible through corresponding lenses are:

gasrSummaryMap :: HashMap SummaryKeyType Int

Response lenses

gasrSummaryMap :: Lens' GetAccountSummaryResponse (HashMap SummaryKeyType Int)

A set of key value pairs containing account-level information.

SummaryMap contains the following keys: AccessKeysPerUserQuota - Maximum number of access keys that can be created per user

AccountMFAEnabled - 1 if the root account has an MFA device assigned to it, 0 otherwise

AssumeRolePolicySizeQuota - Maximum allowed size for assume role policy documents (in kilobytes)

GroupPolicySizeQuota - Maximum allowed size for Group policy documents (in kilobytes)

Groups - Number of Groups for the AWS account

GroupsPerUserQuota - Maximum number of groups an IAM user can belong to

GroupsQuota - Maximum groups allowed for the AWS account

InstanceProfiles - Number of instance profiles for the AWS account

InstanceProfilesQuota - Maximum instance profiles allowed for the AWS account

MFADevices - Number of MFA devices, either assigned or unassigned

MFADevicesInUse - Number of MFA devices that have been assigned to an IAM user or to the root account

RolePolicySizeQuota - Maximum allowed size for role policy documents (in kilobytes)

Roles - Number of roles for the AWS account

RolesQuota - Maximum roles allowed for the AWS account

ServerCertificates - Number of server certificates for the AWS account

ServerCertificatesQuota - Maximum server certificates allowed for the AWS account

SigningCertificatesPerUserQuota - Maximum number of X509 certificates allowed for a user

UserPolicySizeQuota - Maximum allowed size for user policy documents (in kilobytes)

Users - Number of users for the AWS account

UsersQuota - Maximum users allowed for the AWS account