Network.AWS.IAM.GetContextKeysForCustomPolicy

Description

Gets a list of all of the context keys referenced in the input policies. The policies are supplied as a list of one or more strings. To get the context keys from policies associated with an IAM user, group, or role, use GetContextKeysForPrincipalPolicy .

Context keys are variables maintained by AWS and its services that provide details about the context of an API query request, and can be evaluated by testing against a value specified in an IAM policy. Use GetContextKeysForCustomPolicy to understand what key names and values you must supply when you call SimulateCustomPolicy . Note that all parameters are shown in unencoded form here for clarity, but must be URL encoded to be included as a part of a real HTML request.

Creating a Request

getContextKeysForCustomPolicy :: GetContextKeysForCustomPolicy

Creates a value of GetContextKeysForCustomPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gckfcpPolicyInputList - A list of policies for which you want the list of context keys referenced in those policies. Each document is specified as a string containing the complete, valid JSON text of an IAM policy. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

data GetContextKeysForCustomPolicy

See: getContextKeysForCustomPolicy smart constructor.

Request Lenses

gckfcpPolicyInputList :: Lens' GetContextKeysForCustomPolicy [Text]

A list of policies for which you want the list of context keys referenced in those policies. Each document is specified as a string containing the complete, valid JSON text of an IAM policy. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

Destructuring the Response

getContextKeysForPolicyResponse :: GetContextKeysForPolicyResponse

Creates a value of GetContextKeysForPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gckfpContextKeyNames - The list of context keys that are referenced in the input policies.

data GetContextKeysForPolicyResponse

Contains the response to a successful GetContextKeysForPrincipalPolicy or GetContextKeysForCustomPolicy request.

See: getContextKeysForPolicyResponse smart constructor.

Response Lenses

gckfpContextKeyNames :: Lens' GetContextKeysForPolicyResponse [Text]

The list of context keys that are referenced in the input policies.