Network.AWS.IAM.GetContextKeysForPrincipalPolicy

Description

Gets a list of all of the context keys referenced in all of the IAM policies attached to the specified IAM entity. The entity can be an IAM user, group, or role. If you specify a user, then the request also includes all of the policies attached to groups that the user is a member of.

You can optionally include a list of one or more additional policies, specified as strings. If you want to include only a list of policies by string, use GetContextKeysForCustomPolicy instead.

Note: This API discloses information about the permissions granted to other users. If you do not want users to see other user's permissions, then consider allowing them to use GetContextKeysForCustomPolicy instead.

Context keys are variables maintained by AWS and its services that provide details about the context of an API query request, and can be evaluated by testing against a value in an IAM policy. Use GetContextKeysForPrincipalPolicy to understand what key names and values you must supply when you call SimulatePrincipalPolicy .

Creating a Request

getContextKeysForPrincipalPolicy

Arguments

:: Text
gckfppPolicySourceARN

-> GetContextKeysForPrincipalPolicy

Creates a value of GetContextKeysForPrincipalPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gckfppPolicyInputList - An optional list of additional policies for which you want the list of context keys that are referenced. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

gckfppPolicySourceARN - The ARN of a user, group, or role whose policies contain the context keys that you want listed. If you specify a user, the list includes context keys that are found in all policies attached to the user as well as to all groups that the user is a member of. If you pick a group or a role, then it includes only those context keys that are found in policies attached to that entity. Note that all parameters are shown in unencoded form here for clarity, but must be URL encoded to be included as a part of a real HTML request. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data GetContextKeysForPrincipalPolicy

See: getContextKeysForPrincipalPolicy smart constructor.

Request Lenses

gckfppPolicyInputList :: Lens' GetContextKeysForPrincipalPolicy [Text]

An optional list of additional policies for which you want the list of context keys that are referenced. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

gckfppPolicySourceARN :: Lens' GetContextKeysForPrincipalPolicy Text

The ARN of a user, group, or role whose policies contain the context keys that you want listed. If you specify a user, the list includes context keys that are found in all policies attached to the user as well as to all groups that the user is a member of. If you pick a group or a role, then it includes only those context keys that are found in policies attached to that entity. Note that all parameters are shown in unencoded form here for clarity, but must be URL encoded to be included as a part of a real HTML request. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

getContextKeysForPolicyResponse :: GetContextKeysForPolicyResponse

Creates a value of GetContextKeysForPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gckfpContextKeyNames - The list of context keys that are referenced in the input policies.

data GetContextKeysForPolicyResponse

Contains the response to a successful GetContextKeysForPrincipalPolicy or GetContextKeysForCustomPolicy request.

See: getContextKeysForPolicyResponse smart constructor.

Response Lenses

gckfpContextKeyNames :: Lens' GetContextKeysForPolicyResponse [Text]

The list of context keys that are referenced in the input policies.