Network.AWS.IAM.GetCredentialReport

Description

Retrieves a credential report for the AWS account. For more information about the credential report, see Getting Credential Reports in the Using IAM guide. https:/iam.amazonaws.com ?Action=GetCredentialReport &Version=2010-05-08 &AUTHPARAMS BASE-64 ENCODED FILE CONTENTS text/csv 2014-08-28T21:42:50Z 29f47818-99f5-11e1-a4c3-27EXAMPLE804.

Request

data GetCredentialReport

Request constructor

getCredentialReport :: GetCredentialReport

Smart constructor for the minimum required parameters to construct a valid GetCredentialReport request.

Response

data GetCredentialReportResponse

Contains the result of a successful invocation of the GetCredentialReport action.

Response constructor

getCredentialReportResponse :: GetCredentialReportResponse

Smart constructor for the minimum required parameters to construct a valid GetCredentialReportResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Content :: Maybe ByteString

ReportFormat :: Maybe ReportFormatType

GeneratedTime :: Maybe ISO8601

Response lenses

gcrrrContent :: Lens' GetCredentialReportResponse (Maybe ByteString)

Contains the credential report. The report is Base64-encoded.

gcrrrReportFormat :: Lens' GetCredentialReportResponse (Maybe ReportFormatType)

The format (MIME type) of the credential report.

gcrrrGeneratedTime :: Lens' GetCredentialReportResponse (Maybe ISO8601)

The time and date when the credential report was created, in ISO 8601 date-time format.