Network.AWS.IAM.GetGroup

Description

Returns a list of users that are in the specified group. You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetGroup.html

Request

data GetGroup

Request constructor

getGroup

Arguments

:: Text
ggGroupName

-> GetGroup

GetGroup constructor.

The fields accessible through corresponding lenses are:

ggGroupName :: Text

ggMarker :: Maybe Text

ggMaxItems :: Maybe Natural

Request lenses

ggGroupName :: Lens' GetGroup Text

The name of the group.

ggMarker :: Lens' GetGroup (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

ggMaxItems :: Lens' GetGroup (Maybe Natural)

Use this only when paginating results to indicate the maximum number of groups you want in the response. If there are additional groups beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data GetGroupResponse

Response constructor

getGroupResponse

Arguments

:: Group
ggrGroup

-> GetGroupResponse

GetGroupResponse constructor.

The fields accessible through corresponding lenses are:

ggrGroup :: Group

ggrIsTruncated :: Maybe Bool

ggrMarker :: Maybe Text

ggrUsers :: [User]

Response lenses

ggrGroup :: Lens' GetGroupResponse Group

Information about the group.

ggrIsTruncated :: Lens' GetGroupResponse (Maybe Bool)

A flag that indicates whether there are more user names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more user names in the list.

ggrMarker :: Lens' GetGroupResponse (Maybe Text)

If IsTruncated is true, then this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

ggrUsers :: Lens' GetGroupResponse [User]

A list of users in the group.