Network.AWS.IAM.GetGroupPolicy

Description

Retrieves the specified inline policy document that is embedded in the specified group.

A group can also have managed policies attached to it. To retrieve a managed policy document that is attached to a group, use GetPolicy to determine the policy's default version, then use GetPolicyVersion to retrieve the policy document.

For more information about policies, refer to Managed Policies and InlinePolicies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetGroupPolicy.html

Request

data GetGroupPolicy

Request constructor

getGroupPolicy

Arguments

:: Text
ggpGroupName

-> Text
ggpPolicyName

-> GetGroupPolicy

GetGroupPolicy constructor.

The fields accessible through corresponding lenses are:

ggpGroupName :: Text

ggpPolicyName :: Text

Request lenses

ggpGroupName :: Lens' GetGroupPolicy Text

The name of the group the policy is associated with.

ggpPolicyName :: Lens' GetGroupPolicy Text

The name of the policy document to get.

Response

data GetGroupPolicyResponse

Response constructor

getGroupPolicyResponse

Arguments

:: Text
ggprGroupName

-> Text
ggprPolicyName

-> Text
ggprPolicyDocument

-> GetGroupPolicyResponse

GetGroupPolicyResponse constructor.

The fields accessible through corresponding lenses are:

ggprGroupName :: Text

ggprPolicyDocument :: Text

ggprPolicyName :: Text

Response lenses

ggprGroupName :: Lens' GetGroupPolicyResponse Text

The group the policy is associated with.

ggprPolicyDocument :: Lens' GetGroupPolicyResponse Text

The policy document.

ggprPolicyName :: Lens' GetGroupPolicyResponse Text

The name of the policy.