Network.AWS.IAM.GetInstanceProfile

Description

Retrieves information about the specified instance profile, including the instance profile's path, GUID, ARN, and role. For more information about instance profiles, go to About Instance Profiles. For more information about ARNs, go to ARNs. https:/iam.amazonaws.com ?Action=GetInstanceProfile &InstanceProfileName=Webserver &Version=2010-05-08 &AUTHPARAMS AIPAD5ARO2C5EXAMPLE3G application_abccomponent_xyz/ arn:aws:iam::123456789012:roleapplication_abccomponent_xyz/S3Access S3Access {Version:"2012-10-17",Statement:[{Effect:Allow,Principal:{Service:["ec2.amazonaws.com"]},Action:["sts:AssumeRole"]}]} 2012-05-09T15:45:35Z AROACVYKSVTSZFEXAMPLE Webserver application_abccomponent_xyz/ arn:aws:iam::123456789012:instance-profileapplication_abccomponent_xyz/Webserver 2012-05-09T16:11:10Z 37289fda-99f2-11e1-a4c3-27EXAMPLE804.

Request

data GetInstanceProfile

Request constructor

getInstanceProfile

Arguments

:: Text
gipInstanceProfileName

-> GetInstanceProfile

Smart constructor for the minimum required parameters to construct a valid GetInstanceProfile request.

The fields accessible through corresponding lenses are:

InstanceProfileName :: Text

Request lenses

gipInstanceProfileName :: Lens' GetInstanceProfile Text

Name of the instance profile to get information about.

Response

data GetInstanceProfileResponse

Contains the result of a successful invocation of the GetInstanceProfile action.

Response constructor

getInstanceProfileResponse

Arguments

:: InstanceProfile
giprInstanceProfile

-> GetInstanceProfileResponse

Smart constructor for the minimum required parameters to construct a valid GetInstanceProfileResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

InstanceProfile :: InstanceProfile

Response lenses

giprInstanceProfile :: Lens' GetInstanceProfileResponse InstanceProfile

Information about the instance profile.