Network.AWS.IAM.GetLoginProfile

Description

Retrieves the user name and password-creation date for the specified user. If the user has not been assigned a password, the action returns a 404 (NoSuchEntity) error.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetLoginProfile.html

Request

data GetLoginProfile

Request constructor

getLoginProfile

Arguments

:: Text
glpUserName

-> GetLoginProfile

GetLoginProfile constructor.

The fields accessible through corresponding lenses are:

glpUserName :: Text

Request lenses

glpUserName :: Lens' GetLoginProfile Text

The name of the user whose login profile you want to retrieve.

Response

data GetLoginProfileResponse

Response constructor

getLoginProfileResponse

Arguments

:: LoginProfile
glprLoginProfile

-> GetLoginProfileResponse

GetLoginProfileResponse constructor.

The fields accessible through corresponding lenses are:

glprLoginProfile :: LoginProfile

Response lenses

glprLoginProfile :: Lens' GetLoginProfileResponse LoginProfile

The user name and password create date for the user.