Network.AWS.IAM.GetOpenIDConnectProvider

Description

Returns information about the specified OpenID Connect provider.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetOpenIDConnectProvider.html

Request

data GetOpenIDConnectProvider

Request constructor

getOpenIDConnectProvider

Arguments

:: Text
goidcpOpenIDConnectProviderArn

-> GetOpenIDConnectProvider

GetOpenIDConnectProvider constructor.

The fields accessible through corresponding lenses are:

goidcpOpenIDConnectProviderArn :: Text

Request lenses

goidcpOpenIDConnectProviderArn :: Lens' GetOpenIDConnectProvider Text

The Amazon Resource Name (ARN) of the IAM OpenID Connect (OIDC) provider to get information for. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders action.

Response

data GetOpenIDConnectProviderResponse

Response constructor

getOpenIDConnectProviderResponse :: GetOpenIDConnectProviderResponse

GetOpenIDConnectProviderResponse constructor.

The fields accessible through corresponding lenses are:

goidcprClientIDList :: [Text]

goidcprCreateDate :: Maybe UTCTime

goidcprThumbprintList :: [Text]

goidcprUrl :: Maybe Text

Response lenses

goidcprClientIDList :: Lens' GetOpenIDConnectProviderResponse [Text]

A list of client IDs (also known as audiences) that are associated with the specified IAM OpenID Connect provider. For more information, see CreateOpenIDConnectProvider.

goidcprCreateDate :: Lens' GetOpenIDConnectProviderResponse (Maybe UTCTime)

The date and time when the IAM OpenID Connect provider entity was created in the AWS account.

goidcprThumbprintList :: Lens' GetOpenIDConnectProviderResponse [Text]

A list of certificate thumbprints that are associated with the specified IAM OpenID Connect provider. For more information, see CreateOpenIDConnectProvider .

goidcprUrl :: Lens' GetOpenIDConnectProviderResponse (Maybe Text)

The URL that the IAM OpenID Connect provider is associated with. For more information, see CreateOpenIDConnectProvider.