Network.AWS.IAM.GetPolicy

Description

Retrieves information about the specified managed policy, including the policy's default version and the total number of users, groups, and roles that the policy is attached to. For a list of the specific users, groups, and roles that the policy is attached to, use the ListEntitiesForPolicy API. This API returns metadata about the policy. To retrieve the policy document for a specific version of the policy, use GetPolicyVersion.

This API retrieves information about managed policies. To retrieve information about an inline policy that is embedded with a user, group, or role, use the GetUserPolicy, GetGroupPolicy, or GetRolePolicy API.

For more information about policies, refer to Managed Policies and InlinePolicies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetPolicy.html

Request

data GetPolicy

Request constructor

getPolicy

Arguments

:: Text
gpPolicyArn

-> GetPolicy

GetPolicy constructor.

The fields accessible through corresponding lenses are:

gpPolicyArn :: Text

Request lenses

gpPolicyArn :: Lens' GetPolicy Text

Response

data GetPolicyResponse

Response constructor

getPolicyResponse :: GetPolicyResponse

GetPolicyResponse constructor.

The fields accessible through corresponding lenses are:

gprPolicy :: Maybe Policy

Response lenses

gprPolicy :: Lens' GetPolicyResponse (Maybe Policy)

Information about the policy.