Network.AWS.IAM.GetPolicyVersion

Description

Retrieves information about the specified version of the specified managed policy, including the policy document.

To list the available versions for a policy, use ListPolicyVersions .

This API retrieves information about managed policies. To retrieve information about an inline policy that is embedded in a user, group, or role, use the GetUserPolicy , GetGroupPolicy , or GetRolePolicy API.

For more information about the types of policies, see Managed Policies and Inline Policies in the IAM User Guide .

For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

Creating a Request

getPolicyVersion

Arguments

:: Text
gpvPolicyARN

-> Text
gpvVersionId

-> GetPolicyVersion

Creates a value of GetPolicyVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpvPolicyARN - The Amazon Resource Name (ARN) of the managed policy that you want information about. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

gpvVersionId - Identifies the policy version to retrieve. This parameter allows (per its regex pattern ) a string of characters that consists of the lowercase letter v followed by one or two digits, and optionally followed by a period . and a string of letters and digits.

data GetPolicyVersion

See: getPolicyVersion smart constructor.

Request Lenses

gpvPolicyARN :: Lens' GetPolicyVersion Text

The Amazon Resource Name (ARN) of the managed policy that you want information about. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

gpvVersionId :: Lens' GetPolicyVersion Text

Identifies the policy version to retrieve. This parameter allows (per its regex pattern ) a string of characters that consists of the lowercase letter v followed by one or two digits, and optionally followed by a period . and a string of letters and digits.

Destructuring the Response

getPolicyVersionResponse

Arguments

:: Int
gpvrsResponseStatus

-> GetPolicyVersionResponse

Creates a value of GetPolicyVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpvrsPolicyVersion - A structure containing details about the policy version.

gpvrsResponseStatus - -- | The response status code.

data GetPolicyVersionResponse

Contains the response to a successful GetPolicyVersion request.

See: getPolicyVersionResponse smart constructor.

Response Lenses

gpvrsPolicyVersion :: Lens' GetPolicyVersionResponse (Maybe PolicyVersion)

A structure containing details about the policy version.

gpvrsResponseStatus :: Lens' GetPolicyVersionResponse Int

- | The response status code.