Network.AWS.IAM.GetRole

Description

Retrieves information about the specified role, including the role's path, GUID, ARN, and the policy granting permission to assume the role. For more information about ARNs, go to ARNs. For more information about roles, go to Working with Roles.

The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http://www.faqs.org/rfcs/rfc3986.html.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetRole.html

Request

data GetRole

Request constructor

getRole

Arguments

:: Text
grRoleName

-> GetRole

GetRole constructor.

The fields accessible through corresponding lenses are:

grRoleName :: Text

Request lenses

grRoleName :: Lens' GetRole Text

The name of the role to get information about.

Response

data GetRoleResponse

Response constructor

getRoleResponse

Arguments

:: Role
grrRole

-> GetRoleResponse

GetRoleResponse constructor.

The fields accessible through corresponding lenses are:

grrRole :: Role

Response lenses

grrRole :: Lens' GetRoleResponse Role

Information about the role.