Network.AWS.IAM.GetRolePolicy

Description

Retrieves the specified inline policy document that is embedded with the specified role.

A role can also have managed policies attached to it. To retrieve a managed policy document that is attached to a role, use GetPolicy to determine the policy's default version, then use GetPolicyVersion to retrieve the policy document.

For more information about policies, refer to Managed Policies and InlinePolicies in the Using IAM guide.

For more information about roles, go to Using Roles to Delegate Permissionsand Federate Identities.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetRolePolicy.html

Request

data GetRolePolicy

Request constructor

getRolePolicy

Arguments

:: Text
grpRoleName

-> Text
grpPolicyName

-> GetRolePolicy

GetRolePolicy constructor.

The fields accessible through corresponding lenses are:

grpPolicyName :: Text

grpRoleName :: Text

Request lenses

grpPolicyName :: Lens' GetRolePolicy Text

The name of the policy document to get.

grpRoleName :: Lens' GetRolePolicy Text

The name of the role associated with the policy.

Response

data GetRolePolicyResponse

Response constructor

getRolePolicyResponse

Arguments

:: Text
grprRoleName

-> Text
grprPolicyName

-> Text
grprPolicyDocument

-> GetRolePolicyResponse

GetRolePolicyResponse constructor.

The fields accessible through corresponding lenses are:

grprPolicyDocument :: Text

grprPolicyName :: Text

grprRoleName :: Text

Response lenses

grprPolicyDocument :: Lens' GetRolePolicyResponse Text

The policy document.

grprPolicyName :: Lens' GetRolePolicyResponse Text

The name of the policy.

grprRoleName :: Lens' GetRolePolicyResponse Text

The role the policy is associated with.