Network.AWS.IAM.GetSAMLProvider

Description

Returns the SAML provider metadocument that was uploaded when the provider was created or updated. This operation requires Signature Version 4. https:/iam.amazonaws.com ?Action=GetSAMLProvider &Name=arn:aws:iam::123456789012:saml-metadata/MyUniversity &Version=2010-05-08 &AUTHPARAMS 2012-05-09T16:27:11Z 2015-12-31T211:59:59Z Pd9fexDssTkRgGNqs...DxptfEs== 29f47818-99f5-11e1-a4c3-27EXAMPLE804.

Request

data GetSAMLProvider

Request constructor

getSAMLProvider

Arguments

:: Text
gsamlpSAMLProviderArn

-> GetSAMLProvider

Smart constructor for the minimum required parameters to construct a valid GetSAMLProvider request.

The fields accessible through corresponding lenses are:

SAMLProviderArn :: Text

Request lenses

gsamlpSAMLProviderArn :: Lens' GetSAMLProvider Text

The Amazon Resource Name (ARN) of the SAML provider to get information about.

Response

data GetSAMLProviderResponse

Contains the result of a successful invocation of the GetSAMLProvider action.

Response constructor

getSAMLProviderResponse :: GetSAMLProviderResponse

Smart constructor for the minimum required parameters to construct a valid GetSAMLProviderResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SAMLMetadataDocument :: Maybe Text

CreateDate :: Maybe ISO8601

ValidUntil :: Maybe ISO8601

Response lenses

gsamlprSAMLMetadataDocument :: Lens' GetSAMLProviderResponse (Maybe Text)

The XML metadata document that includes information about an identity provider.

gsamlprCreateDate :: Lens' GetSAMLProviderResponse (Maybe ISO8601)

The date and time when the SAML provider was created.

gsamlprValidUntil :: Lens' GetSAMLProviderResponse (Maybe ISO8601)

The expiration date and time for the SAML provider.