Network.AWS.IAM.GetServerCertificate

Description

Retrieves information about the specified server certificate stored in IAM.

For more information about working with server certificates, including a list of AWS services that can use the server certificates that you manage with IAM, go to Working with Server Certificates in the IAM User Guide .

Creating a Request

getServerCertificate

Arguments

:: Text
gscServerCertificateName

-> GetServerCertificate

Creates a value of GetServerCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gscServerCertificateName - The name of the server certificate you want to retrieve information about. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data GetServerCertificate

See: getServerCertificate smart constructor.

Request Lenses

gscServerCertificateName :: Lens' GetServerCertificate Text

The name of the server certificate you want to retrieve information about. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

getServerCertificateResponse

Arguments

:: Int
gscrsResponseStatus

-> ServerCertificate
gscrsServerCertificate

-> GetServerCertificateResponse

Creates a value of GetServerCertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gscrsResponseStatus - -- | The response status code.

gscrsServerCertificate - A structure containing details about the server certificate.

data GetServerCertificateResponse

Contains the response to a successful GetServerCertificate request.

See: getServerCertificateResponse smart constructor.

Response Lenses

gscrsResponseStatus :: Lens' GetServerCertificateResponse Int

- | The response status code.

gscrsServerCertificate :: Lens' GetServerCertificateResponse ServerCertificate

A structure containing details about the server certificate.