Network.AWS.IAM.GetUserPolicy

Description

Retrieves the specified inline policy document that is embedded in the specified user.

A user can also have managed policies attached to it. To retrieve a managed policy document that is attached to a user, use GetPolicy to determine the policy's default version, then use GetPolicyVersion to retrieve the policy document.

For more information about policies, refer to Managed Policies and InlinePolicies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetUserPolicy.html

Request

data GetUserPolicy

Request constructor

getUserPolicy

Arguments

:: Text
gupUserName

-> Text
gupPolicyName

-> GetUserPolicy

GetUserPolicy constructor.

The fields accessible through corresponding lenses are:

gupPolicyName :: Text

gupUserName :: Text

Request lenses

gupPolicyName :: Lens' GetUserPolicy Text

The name of the policy document to get.

gupUserName :: Lens' GetUserPolicy Text

The name of the user who the policy is associated with.

Response

data GetUserPolicyResponse

Response constructor

getUserPolicyResponse

Arguments

:: Text
guprUserName

-> Text
guprPolicyName

-> Text
guprPolicyDocument

-> GetUserPolicyResponse

GetUserPolicyResponse constructor.

The fields accessible through corresponding lenses are:

guprPolicyDocument :: Text

guprPolicyName :: Text

guprUserName :: Text

Response lenses

guprPolicyDocument :: Lens' GetUserPolicyResponse Text

The policy document.

guprPolicyName :: Lens' GetUserPolicyResponse Text

The name of the policy.

guprUserName :: Lens' GetUserPolicyResponse Text

The user the policy is associated with.