Network.AWS.IAM.ListAccessKeys

Description

Returns information about the access key IDs associated with the specified user. If there are none, the action returns an empty list. Although each user is limited to a small number of keys, you can still paginate the results using the MaxItems and Marker parameters. If the UserName field is not specified, the UserName is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListAccessKeys.html

Request

data ListAccessKeys

Request constructor

listAccessKeys :: ListAccessKeys

ListAccessKeys constructor.

The fields accessible through corresponding lenses are:

lakMarker :: Maybe Text

lakMaxItems :: Maybe Natural

lakUserName :: Maybe Text

Request lenses

lakMarker :: Lens' ListAccessKeys (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lakMaxItems :: Lens' ListAccessKeys (Maybe Natural)

Use this parameter only when paginating results to indicate the maximum number of keys you want in the response. If there are additional keys beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lakUserName :: Lens' ListAccessKeys (Maybe Text)

The name of the user.

Response

data ListAccessKeysResponse

Response constructor

listAccessKeysResponse :: ListAccessKeysResponse

ListAccessKeysResponse constructor.

The fields accessible through corresponding lenses are:

lakrAccessKeyMetadata :: [AccessKeyMetadata]

lakrIsTruncated :: Maybe Bool

lakrMarker :: Maybe Text

Response lenses

lakrAccessKeyMetadata :: Lens' ListAccessKeysResponse [AccessKeyMetadata]

A list of access key metadata.

lakrIsTruncated :: Lens' ListAccessKeysResponse (Maybe Bool)

A flag that indicates whether there are more keys to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more keys in the list.

lakrMarker :: Lens' ListAccessKeysResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.