Network.AWS.IAM.ListAccountAliases

Description

Lists the account aliases associated with the account. For information about using an AWS account alias, see Using an Alias for Your AWS Account ID in the Using IAM guide.

You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListAccountAliases.html

Request

data ListAccountAliases

Request constructor

listAccountAliases :: ListAccountAliases

ListAccountAliases constructor.

The fields accessible through corresponding lenses are:

laaMarker :: Maybe Text

laaMaxItems :: Maybe Natural

Request lenses

laaMarker :: Lens' ListAccountAliases (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

laaMaxItems :: Lens' ListAccountAliases (Maybe Natural)

Use this only when paginating results to indicate the maximum number of account aliases you want in the response. If there are additional account aliases beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListAccountAliasesResponse

Response constructor

listAccountAliasesResponse :: ListAccountAliasesResponse

ListAccountAliasesResponse constructor.

The fields accessible through corresponding lenses are:

laarAccountAliases :: [Text]

laarIsTruncated :: Maybe Bool

laarMarker :: Maybe Text

Response lenses

laarAccountAliases :: Lens' ListAccountAliasesResponse [Text]

A list of aliases associated with the account.

laarIsTruncated :: Lens' ListAccountAliasesResponse (Maybe Bool)

A flag that indicates whether there are more account aliases to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more account aliases in the list.

laarMarker :: Lens' ListAccountAliasesResponse (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.