Network.AWS.IAM.ListAttachedRolePolicies

Description

Lists all managed policies that are attached to the specified role.

A role can also have inline policies embedded with it. To list the inline policies for a role, use the ListRolePolicies API. For information about policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

You can paginate the results using the MaxItems and Marker parameters. You can use the PathPrefix parameter to limit the list of policies to only those matching the specified path prefix. If there are no policies attached to the specified role (or none that match the specified path prefix), the action returns an empty list.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListAttachedRolePolicies.html

Request

data ListAttachedRolePolicies

Request constructor

listAttachedRolePolicies

Arguments

:: Text
larpRoleName

-> ListAttachedRolePolicies

ListAttachedRolePolicies constructor.

The fields accessible through corresponding lenses are:

larpMarker :: Maybe Text

larpMaxItems :: Maybe Natural

larpPathPrefix :: Maybe Text

larpRoleName :: Text

Request lenses

larpMarker :: Lens' ListAttachedRolePolicies (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

larpMaxItems :: Lens' ListAttachedRolePolicies (Maybe Natural)

Use this only when paginating results to indicate the maximum number of policies you want in the response. If there are additional policies beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

larpPathPrefix :: Lens' ListAttachedRolePolicies (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all policies.

larpRoleName :: Lens' ListAttachedRolePolicies Text

The name (friendly name, not ARN) of the role to list attached policies for.

Response

data ListAttachedRolePoliciesResponse

Response constructor

listAttachedRolePoliciesResponse :: ListAttachedRolePoliciesResponse

ListAttachedRolePoliciesResponse constructor.

The fields accessible through corresponding lenses are:

larprAttachedPolicies :: [AttachedPolicy]

larprIsTruncated :: Maybe Bool

larprMarker :: Maybe Text

Response lenses

larprAttachedPolicies :: Lens' ListAttachedRolePoliciesResponse [AttachedPolicy]

A list of the attached policies.

larprIsTruncated :: Lens' ListAttachedRolePoliciesResponse (Maybe Bool)

A flag that indicates whether there are more policies to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policies in the list.

larprMarker :: Lens' ListAttachedRolePoliciesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.