Network.AWS.IAM.ListAttachedUserPolicies

Description

Lists all managed policies that are attached to the specified user.

A user can also have inline policies embedded with it. To list the inline policies for a user, use the ListUserPolicies API. For information about policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

You can paginate the results using the MaxItems and Marker parameters. You can use the PathPrefix parameter to limit the list of policies to only those matching the specified path prefix. If there are no policies attached to the specified group (or none that match the specified path prefix), the action returns an empty list.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListAttachedUserPolicies.html

Request

data ListAttachedUserPolicies

Request constructor

listAttachedUserPolicies

Arguments

:: Text
laupUserName

-> ListAttachedUserPolicies

ListAttachedUserPolicies constructor.

The fields accessible through corresponding lenses are:

laupMarker :: Maybe Text

laupMaxItems :: Maybe Natural

laupPathPrefix :: Maybe Text

laupUserName :: Text

Request lenses

laupMarker :: Lens' ListAttachedUserPolicies (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

laupMaxItems :: Lens' ListAttachedUserPolicies (Maybe Natural)

Use this only when paginating results to indicate the maximum number of policies you want in the response. If there are additional policies beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

laupPathPrefix :: Lens' ListAttachedUserPolicies (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all policies.

laupUserName :: Lens' ListAttachedUserPolicies Text

The name (friendly name, not ARN) of the user to list attached policies for.

Response

data ListAttachedUserPoliciesResponse

Response constructor

listAttachedUserPoliciesResponse :: ListAttachedUserPoliciesResponse

ListAttachedUserPoliciesResponse constructor.

The fields accessible through corresponding lenses are:

lauprAttachedPolicies :: [AttachedPolicy]

lauprIsTruncated :: Maybe Bool

lauprMarker :: Maybe Text

Response lenses

lauprAttachedPolicies :: Lens' ListAttachedUserPoliciesResponse [AttachedPolicy]

A list of the attached policies.

lauprIsTruncated :: Lens' ListAttachedUserPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more policies to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policies in the list.

lauprMarker :: Lens' ListAttachedUserPoliciesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.