Network.AWS.IAM.ListEntitiesForPolicy

Description

Lists all users, groups, and roles that the specified managed policy is attached to.

You can use the optional EntityFilter parameter to limit the results to a particular type of entity (users, groups, or roles). For example, to list only the roles that are attached to the specified policy, set EntityFilter to Role.

You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListEntitiesForPolicy.html

Request

data ListEntitiesForPolicy

Request constructor

listEntitiesForPolicy

Arguments

:: Text
lefpPolicyArn

-> ListEntitiesForPolicy

ListEntitiesForPolicy constructor.

The fields accessible through corresponding lenses are:

lefpEntityFilter :: Maybe EntityType

lefpMarker :: Maybe Text

lefpMaxItems :: Maybe Natural

lefpPathPrefix :: Maybe Text

lefpPolicyArn :: Text

Request lenses

lefpEntityFilter :: Lens' ListEntitiesForPolicy (Maybe EntityType)

The entity type to use for filtering the results.

For example, when EntityFilter is Role, only the roles that are attached to the specified policy are returned. This parameter is optional. If it is not included, all attached entities (users, groups, and roles) are returned.

lefpMarker :: Lens' ListEntitiesForPolicy (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lefpMaxItems :: Lens' ListEntitiesForPolicy (Maybe Natural)

Use this only when paginating results to indicate the maximum number of entities you want in the response. If there are additional entities beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lefpPathPrefix :: Lens' ListEntitiesForPolicy (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all entities.

lefpPolicyArn :: Lens' ListEntitiesForPolicy Text

Response

data ListEntitiesForPolicyResponse

Response constructor

listEntitiesForPolicyResponse :: ListEntitiesForPolicyResponse

ListEntitiesForPolicyResponse constructor.

The fields accessible through corresponding lenses are:

lefprIsTruncated :: Maybe Bool

lefprMarker :: Maybe Text

lefprPolicyGroups :: [PolicyGroup]

lefprPolicyRoles :: [PolicyRole]

lefprPolicyUsers :: [PolicyUser]

Response lenses

lefprIsTruncated :: Lens' ListEntitiesForPolicyResponse (Maybe Bool)

A flag that indicates whether there are more entities to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more entities in the list.

lefprMarker :: Lens' ListEntitiesForPolicyResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lefprPolicyGroups :: Lens' ListEntitiesForPolicyResponse [PolicyGroup]

A list of groups that the policy is attached to.

lefprPolicyRoles :: Lens' ListEntitiesForPolicyResponse [PolicyRole]

A list of roles that the policy is attached to.

lefprPolicyUsers :: Lens' ListEntitiesForPolicyResponse [PolicyUser]

A list of users that the policy is attached to.