Network.AWS.IAM.ListGroupPolicies

Description

Lists the names of the policies associated with the specified group. If there are none, the action returns an empty list.

You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListGroupPolicies.html

Request

data ListGroupPolicies

Request constructor

listGroupPolicies

Arguments

:: Text
lgpGroupName

-> ListGroupPolicies

ListGroupPolicies constructor.

The fields accessible through corresponding lenses are:

lgpGroupName :: Text

lgpMarker :: Maybe Text

lgpMaxItems :: Maybe Natural

Request lenses

lgpGroupName :: Lens' ListGroupPolicies Text

The name of the group to list policies for.

lgpMarker :: Lens' ListGroupPolicies (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lgpMaxItems :: Lens' ListGroupPolicies (Maybe Natural)

Use this only when paginating results to indicate the maximum number of policy names you want in the response. If there are additional policy names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListGroupPoliciesResponse

Response constructor

listGroupPoliciesResponse :: ListGroupPoliciesResponse

ListGroupPoliciesResponse constructor.

The fields accessible through corresponding lenses are:

lgprIsTruncated :: Maybe Bool

lgprMarker :: Maybe Text

lgprPolicyNames :: [Text]

Response lenses

lgprIsTruncated :: Lens' ListGroupPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more policy names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policy names in the list.

lgprMarker :: Lens' ListGroupPoliciesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lgprPolicyNames :: Lens' ListGroupPoliciesResponse [Text]

A list of policy names.