Network.AWS.IAM.ListGroups

Description

Lists the groups that have the specified path prefix. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListGroups &PathPrefix=division_abcsubdivision_xyz/ &Version=2010-05-08 &AUTHPARAMS division_abcsubdivision_xyz/ Admins AGPACKCEVSQ6C2EXAMPLE arn:aws:iam::123456789012:group/Admins division_abcsubdivision_xyzproduct_1234engineering/ Test AGP2MAB8DPLSRHEXAMPLE arn:aws:iam::123456789012:group division_abcsubdivision_xyzproduct_1234engineering/Test division_abcsubdivision_xyzproduct_1234 Managers AGPIODR4TAW7CSEXAMPLE arn:aws:iam::123456789012 :groupdivision_abcsubdivision_xyzproduct_1234Managers false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data ListGroups

Request constructor

listGroups :: ListGroups

Smart constructor for the minimum required parameters to construct a valid ListGroups request.

The fields accessible through corresponding lenses are:

PathPrefix :: Maybe Text

Marker :: Maybe Text

MaxItems :: Maybe Integer

Request lenses

lgPathPrefix :: Lens' ListGroups (Maybe Text)

The path prefix for filtering the results. For example: division_abcsubdivision_xyz/, which would get all groups whose path starts with division_abcsubdivision_xyz/. This parameter is optional. If it is not included, it defaults to a slash (/), listing all groups.

lgMarker :: Lens' ListGroups (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lgMaxItems :: Lens' ListGroups (Maybe Integer)

Use this only when paginating results to indicate the maximum number of groups you want in the response. If there are additional groups beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListGroupsResponse

Contains the result of a successful invocation of the ListGroups action.

Response constructor

listGroupsResponse

Arguments

:: [Group]
lgrGroups

-> Bool
lgrIsTruncated

-> ListGroupsResponse

Smart constructor for the minimum required parameters to construct a valid ListGroupsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Groups :: [Group]

IsTruncated :: Bool

Marker :: Maybe Text

Response lenses

lgrGroups :: Lens' ListGroupsResponse [Group]

A list of groups.

lgrIsTruncated :: Lens' ListGroupsResponse Bool

A flag that indicates whether there are more groups to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more groups in the list.

lgrMarker :: Lens' ListGroupsResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.