Network.AWS.IAM.ListGroupsForUser

Description

Lists the groups the specified user belongs to.

You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListGroupsForUser.html

Request

data ListGroupsForUser

Request constructor

listGroupsForUser

Arguments

:: Text
lgfuUserName

-> ListGroupsForUser

ListGroupsForUser constructor.

The fields accessible through corresponding lenses are:

lgfuMarker :: Maybe Text

lgfuMaxItems :: Maybe Natural

lgfuUserName :: Text

Request lenses

lgfuMarker :: Lens' ListGroupsForUser (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lgfuMaxItems :: Lens' ListGroupsForUser (Maybe Natural)

Use this only when paginating results to indicate the maximum number of groups you want in the response. If there are additional groups beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lgfuUserName :: Lens' ListGroupsForUser Text

The name of the user to list groups for.

Response

data ListGroupsForUserResponse

Response constructor

listGroupsForUserResponse :: ListGroupsForUserResponse

ListGroupsForUserResponse constructor.

The fields accessible through corresponding lenses are:

lgfurGroups :: [Group]

lgfurIsTruncated :: Maybe Bool

lgfurMarker :: Maybe Text

Response lenses

lgfurGroups :: Lens' ListGroupsForUserResponse [Group]

A list of groups.

lgfurIsTruncated :: Lens' ListGroupsForUserResponse (Maybe Bool)

A flag that indicates whether there are more groups to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more groups in the list.

lgfurMarker :: Lens' ListGroupsForUserResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.