Network.AWS.IAM.ListInstanceProfiles

Description

Lists the instance profiles that have the specified path prefix. If there are none, the action returns an empty list. For more information about instance profiles, go to About Instance Profiles.

You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListInstanceProfiles.html

Request

data ListInstanceProfiles

Request constructor

listInstanceProfiles :: ListInstanceProfiles

ListInstanceProfiles constructor.

The fields accessible through corresponding lenses are:

lipMarker :: Maybe Text

lipMaxItems :: Maybe Natural

lipPathPrefix :: Maybe Text

Request lenses

lipMarker :: Lens' ListInstanceProfiles (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lipMaxItems :: Lens' ListInstanceProfiles (Maybe Natural)

Use this parameter only when paginating results to indicate the maximum number of instance profiles you want in the response. If there are additional instance profiles beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lipPathPrefix :: Lens' ListInstanceProfiles (Maybe Text)

The path prefix for filtering the results. For example, the prefix '/application_abc/component_xyz/' gets all instance profiles whose path starts with '/application_abc/component_xyz/'.

This parameter is optional. If it is not included, it defaults to a slash (/), listing all instance profiles.

Response

data ListInstanceProfilesResponse

Response constructor

listInstanceProfilesResponse :: ListInstanceProfilesResponse

ListInstanceProfilesResponse constructor.

The fields accessible through corresponding lenses are:

liprInstanceProfiles :: [InstanceProfile]

liprIsTruncated :: Maybe Bool

liprMarker :: Maybe Text

Response lenses

liprInstanceProfiles :: Lens' ListInstanceProfilesResponse [InstanceProfile]

A list of instance profiles.

liprIsTruncated :: Lens' ListInstanceProfilesResponse (Maybe Bool)

A flag that indicates whether there are more instance profiles to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more instance profiles in the list.

liprMarker :: Lens' ListInstanceProfilesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.