Network.AWS.IAM.ListInstanceProfilesForRole

Description

Lists the instance profiles that have the specified associated role. If there are none, the action returns an empty list. For more information about instance profiles, go to About Instance Profiles. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListInstanceProfilesForRole &MaxItems=100 &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS false AIPACZLS2EYYXMEXAMPLE application_abccomponent_xyz/ arn:aws:iam::123456789012:roleapplication_abccomponent_xyz/S3Access S3Access {Version:"2012-10-17",Statement:[{Effect:Allow,Principal:{Service:["ec2.amazonaws.com"]},Action:["sts:AssumeRole"]}]} 2012-05-09T15:45:35Z AROACVSVTSZYK3EXAMPLE Webserver application_abccomponent_xyz/ arn:aws:iam::123456789012:instance-profileapplication_abccomponent_xyz/Webserver 2012-05-09T16:27:11Z 6a8c3992-99f4-11e1-a4c3-27EXAMPLE804.

Request

data ListInstanceProfilesForRole

Request constructor

listInstanceProfilesForRole

Arguments

:: Text
lipfrRoleName

-> ListInstanceProfilesForRole

Smart constructor for the minimum required parameters to construct a valid ListInstanceProfilesForRole request.

The fields accessible through corresponding lenses are:

RoleName :: Text

Marker :: Maybe Text

MaxItems :: Maybe Integer

Request lenses

lipfrRoleName :: Lens' ListInstanceProfilesForRole Text

The name of the role to list instance profiles for.

lipfrMarker :: Lens' ListInstanceProfilesForRole (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lipfrMaxItems :: Lens' ListInstanceProfilesForRole (Maybe Integer)

Use this parameter only when paginating results to indicate the maximum number of user names you want in the response. If there are additional user names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListInstanceProfilesForRoleResponse

Contains the result of a successful invocation of the ListInstanceProfilesForRole action.

Response constructor

listInstanceProfilesForRoleResponse

Arguments

:: [InstanceProfile]
lipfrrInstanceProfiles

-> Bool
lipfrrIsTruncated

-> ListInstanceProfilesForRoleResponse

Smart constructor for the minimum required parameters to construct a valid ListInstanceProfilesForRoleResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

InstanceProfiles :: [InstanceProfile]

IsTruncated :: Bool

Marker :: Maybe Text

Response lenses

lipfrrInstanceProfiles :: Lens' ListInstanceProfilesForRoleResponse [InstanceProfile]

A list of instance profiles.

lipfrrIsTruncated :: Lens' ListInstanceProfilesForRoleResponse Bool

A flag that indicates whether there are more instance profiles to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more instance profiles in the list.

lipfrrMarker :: Lens' ListInstanceProfilesForRoleResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.