Network.AWS.IAM.ListMFADevices

Description

Lists the MFA devices. If the request includes the user name, then this action lists all the MFA devices associated with the specified user name. If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListMFADevices &UserName=Bob &AUTHPARAMS Bob R1234 false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data ListMFADevices

Request constructor

listMFADevices :: ListMFADevices

Smart constructor for the minimum required parameters to construct a valid ListMFADevices request.

The fields accessible through corresponding lenses are:

UserName :: Maybe Text

Marker :: Maybe Text

MaxItems :: Maybe Integer

Request lenses

lmfadUserName :: Lens' ListMFADevices (Maybe Text)

Name of the user whose MFA devices you want to list.

lmfadMarker :: Lens' ListMFADevices (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lmfadMaxItems :: Lens' ListMFADevices (Maybe Integer)

Use this only when paginating results to indicate the maximum number of MFA devices you want in the response. If there are additional MFA devices beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListMFADevicesResponse

Contains the result of a successful invocation of the ListMFADevices action.

Response constructor

listMFADevicesResponse

Arguments

:: [MFADevice]
lmfadrMFADevices

-> Bool
lmfadrIsTruncated

-> ListMFADevicesResponse

Smart constructor for the minimum required parameters to construct a valid ListMFADevicesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

MFADevices :: [MFADevice]

IsTruncated :: Bool

Marker :: Maybe Text

Response lenses

lmfadrMFADevices :: Lens' ListMFADevicesResponse [MFADevice]

A list of MFA devices.

lmfadrIsTruncated :: Lens' ListMFADevicesResponse Bool

A flag that indicates whether there are more MFA devices to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more MFA devices in the list.

lmfadrMarker :: Lens' ListMFADevicesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.