Network.AWS.IAM.ListOpenIdConnectProviders

Description

Lists information about the IAM OpenID Connect (OIDC) provider resource objects defined in the AWS account.

Creating a Request

listOpenIdConnectProviders :: ListOpenIdConnectProviders

Creates a value of ListOpenIdConnectProviders with the minimum fields required to make a request.

data ListOpenIdConnectProviders

See: listOpenIdConnectProviders smart constructor.

Destructuring the Response

listOpenIdConnectProvidersResponse

Arguments

:: Int
loicprsResponseStatus

-> ListOpenIdConnectProvidersResponse

Creates a value of ListOpenIdConnectProvidersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loicprsOpenIdConnectProviderList - The list of IAM OIDC provider resource objects defined in the AWS account.

loicprsResponseStatus - -- | The response status code.

data ListOpenIdConnectProvidersResponse

Contains the response to a successful ListOpenIDConnectProviders request.

See: listOpenIdConnectProvidersResponse smart constructor.

Response Lenses

loicprsOpenIdConnectProviderList :: Lens' ListOpenIdConnectProvidersResponse [OpenIdConnectProviderListEntry]

The list of IAM OIDC provider resource objects defined in the AWS account.

loicprsResponseStatus :: Lens' ListOpenIdConnectProvidersResponse Int

- | The response status code.