Network.AWS.IAM.ListPolicies

Description

Lists all the managed policies that are available to your account, including your own customer managed policies and all AWS managed policies.

You can filter the list of policies that is returned using the optional OnlyAttached, Scope, and PathPrefix parameters. For example, to list only the customer managed policies in your AWS account, set Scope to Local. To list only AWS managed policies, set Scope to AWS.

You can paginate the results using the MaxItems and Marker parameters.

For more information about managed policies, refer to Managed Policies andInline Policies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListPolicies.html

Request

data ListPolicies

Request constructor

listPolicies :: ListPolicies

ListPolicies constructor.

The fields accessible through corresponding lenses are:

lpMarker :: Maybe Text

lpMaxItems :: Maybe Natural

lpOnlyAttached :: Maybe Bool

lpPathPrefix :: Maybe Text

lpScope :: Maybe PolicyScopeType

Request lenses

lpMarker :: Lens' ListPolicies (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lpMaxItems :: Lens' ListPolicies (Maybe Natural)

Use this parameter only when paginating results to indicate the maximum number of policies you want in the response. If there are additional policies beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lpOnlyAttached :: Lens' ListPolicies (Maybe Bool)

A flag to filter the results to only the attached policies.

When OnlyAttached is true, the returned list contains only the policies that are attached to a user, group, or role. When OnlyAttached is false, or when the parameter is not included, all policies are returned.

lpPathPrefix :: Lens' ListPolicies (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all policies.

lpScope :: Lens' ListPolicies (Maybe PolicyScopeType)

The scope to use for filtering the results.

To list only AWS managed policies, set Scope to AWS. To list only the customer managed policies in your AWS account, set Scope to Local.

This parameter is optional. If it is not included, or if it is set to All, all policies are returned.

Response

data ListPoliciesResponse

Response constructor

listPoliciesResponse :: ListPoliciesResponse

ListPoliciesResponse constructor.

The fields accessible through corresponding lenses are:

lprIsTruncated :: Maybe Bool

lprMarker :: Maybe Text

lprPolicies :: [Policy]

Response lenses

lprIsTruncated :: Lens' ListPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more policies to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policies in the list.

lprMarker :: Lens' ListPoliciesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lprPolicies :: Lens' ListPoliciesResponse [Policy]

A list of policies.