Network.AWS.IAM.ListPolicyVersions

Description

Lists information about the versions of the specified managed policy, including the version that is set as the policy's default version.

For more information about managed policies, refer to Managed Policies andInline Policies in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListPolicyVersions.html

Request

data ListPolicyVersions

Request constructor

listPolicyVersions

Arguments

:: Text
lpvPolicyArn

-> ListPolicyVersions

ListPolicyVersions constructor.

The fields accessible through corresponding lenses are:

lpvMarker :: Maybe Text

lpvMaxItems :: Maybe Natural

lpvPolicyArn :: Text

Request lenses

lpvMarker :: Lens' ListPolicyVersions (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lpvMaxItems :: Lens' ListPolicyVersions (Maybe Natural)

Use this parameter only when paginating results to indicate the maximum number of policy versions you want in the response. If there are additional policy versions beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lpvPolicyArn :: Lens' ListPolicyVersions Text

Response

data ListPolicyVersionsResponse

Response constructor

listPolicyVersionsResponse :: ListPolicyVersionsResponse

ListPolicyVersionsResponse constructor.

The fields accessible through corresponding lenses are:

lpvrIsTruncated :: Maybe Bool

lpvrMarker :: Maybe Text

lpvrVersions :: [PolicyVersion]

Response lenses

lpvrIsTruncated :: Lens' ListPolicyVersionsResponse (Maybe Bool)

A flag that indicates whether there are more policy versions to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policy versions in the list.

lpvrMarker :: Lens' ListPolicyVersionsResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lpvrVersions :: Lens' ListPolicyVersionsResponse [PolicyVersion]

A list of policy versions.

For more information about managed policy versions, see Versioning forManaged Policies in the Using IAM guide.