Network.AWS.IAM.ListRolePolicies

Description

Lists the names of the inline policies that are embedded in the specified IAM role.

An IAM role can also have managed policies attached to it. To list the managed policies that are attached to a role, use ListAttachedRolePolicies . For more information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

You can paginate the results using the MaxItems and Marker parameters. If there are no inline policies embedded with the specified role, the action returns an empty list.

This operation returns paginated results.

Creating a Request

listRolePolicies

Arguments

:: Text
lrpRoleName

-> ListRolePolicies

Creates a value of ListRolePolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrpMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lrpMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lrpRoleName - The name of the role to list policies for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

data ListRolePolicies

See: listRolePolicies smart constructor.

Request Lenses

lrpMarker :: Lens' ListRolePolicies (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lrpMaxItems :: Lens' ListRolePolicies (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lrpRoleName :: Lens' ListRolePolicies Text

The name of the role to list policies for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

listRolePoliciesResponse

Arguments

:: Int
lrprsResponseStatus

-> ListRolePoliciesResponse

Creates a value of ListRolePoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrprsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lrprsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lrprsResponseStatus - -- | The response status code.

lrprsPolicyNames - A list of policy names.

data ListRolePoliciesResponse

Contains the response to a successful ListRolePolicies request.

See: listRolePoliciesResponse smart constructor.

Response Lenses

lrprsMarker :: Lens' ListRolePoliciesResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lrprsIsTruncated :: Lens' ListRolePoliciesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lrprsResponseStatus :: Lens' ListRolePoliciesResponse Int

- | The response status code.

lrprsPolicyNames :: Lens' ListRolePoliciesResponse [Text]

A list of policy names.