Network.AWS.IAM.ListRoles

Description

Lists the roles that have the specified path prefix. If there are none, the action returns an empty list. For more information about roles, go to Working with Roles. You can paginate the results using the MaxItems and Marker parameters. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcsrfc3986.html. https:iam.amazonaws.com ?Action=ListRoles &MaxItems=100 &PathPrefix=application_abc &Version=2010-05-08 &AUTHPARAMS false application_abccomponent_xyz/ arn:aws:iam::123456789012:roleapplication_abccomponent_xyz/S3Access S3Access {Version:"2012-10-17",Statement:[{Effect:Allow,Principal:{Service:["ec2.amazonaws.com"]},Action:["sts:AssumeRole"]}]} 2012-05-09T15:45:35Z AROACVSVTSZYEXAMPLEYK application_abccomponent_xyz/ arn:aws:iam::123456789012:roleapplication_abccomponent_xyz/SDBAccess SDBAccess {Version:"2012-10-17",Statement:[{Effect:Allow,Principal:{Service:["ec2.amazonaws.com"]},Action:["sts:AssumeRole"]}]} 2012-05-09T15:45:45Z AROAC2ICXG32EXAMPLEWK 20f7279f-99ee-11e1-a4c3-27EXAMPLE804.

Request

data ListRoles

Request constructor

listRoles :: ListRoles

Smart constructor for the minimum required parameters to construct a valid ListRoles request.

The fields accessible through corresponding lenses are:

PathPrefix :: Maybe Text

Marker :: Maybe Text

MaxItems :: Maybe Integer

Request lenses

lrPathPrefix :: Lens' ListRoles (Maybe Text)

The path prefix for filtering the results. For example: application_abccomponent_xyz/, which would get all roles whose path starts with application_abccomponent_xyz/. This parameter is optional. If it is not included, it defaults to a slash (/), listing all roles.

lrMarker :: Lens' ListRoles (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lrMaxItems :: Lens' ListRoles (Maybe Integer)

Use this parameter only when paginating results to indicate the maximum number of user names you want in the response. If there are additional user names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListRolesResponse

Contains the result of a successful invocation of the ListRoles action.

Response constructor

listRolesResponse

Arguments

:: [Role]
lrrRoles

-> Bool
lrrIsTruncated

-> ListRolesResponse

Smart constructor for the minimum required parameters to construct a valid ListRolesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Roles :: [Role]

IsTruncated :: Bool

Marker :: Maybe Text

Response lenses

lrrRoles :: Lens' ListRolesResponse [Role]

A list of roles.

lrrIsTruncated :: Lens' ListRolesResponse Bool

A flag that indicates whether there are more roles to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more roles in the list.

lrrMarker :: Lens' ListRolesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.