Network.AWS.IAM.ListSAMLProviders

Description

Lists the SAML provider resource objects defined in IAM in the account.

Creating a Request

listSAMLProviders :: ListSAMLProviders

Creates a value of ListSAMLProviders with the minimum fields required to make a request.

data ListSAMLProviders

See: listSAMLProviders smart constructor.

Destructuring the Response

listSAMLProvidersResponse

Arguments

:: Int
lsamlprsResponseStatus

-> ListSAMLProvidersResponse

Creates a value of ListSAMLProvidersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsamlprsSAMLProviderList - The list of SAML provider resource objects defined in IAM for this AWS account.

lsamlprsResponseStatus - -- | The response status code.

data ListSAMLProvidersResponse

Contains the response to a successful ListSAMLProviders request.

See: listSAMLProvidersResponse smart constructor.

Response Lenses

lsamlprsSAMLProviderList :: Lens' ListSAMLProvidersResponse [SAMLProviderListEntry]

The list of SAML provider resource objects defined in IAM for this AWS account.

lsamlprsResponseStatus :: Lens' ListSAMLProvidersResponse Int

- | The response status code.