Network.AWS.IAM.ListServerCertificates

Description

Lists the server certificates that have the specified path prefix. If none exist, the action returns an empty list. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListServerCertificates &PathPrefix=companyservercerts &Version=2010-05-08 &AUTHPARAMS false ProdServerCert companyservercerts/ arn:aws:iam::123456789012:server-certificatecompanyservercerts/ProdServerCert 2010-05-08T01:02:03.004Z ASCACKCEVSQ6CEXAMPLE1 2012-05-08T01:02:03.004Z BetaServerCert companyservercerts/ arn:aws:iam::123456789012:server-certificatecompanyservercerts/BetaServerCert 2010-05-08T02:03:01.004Z ASCACKCEVSQ6CEXAMPLE2 2012-05-08T02:03:01.004Z TestServerCert companyservercerts/ arn:aws:iam::123456789012:server-certificatecompanyservercerts/TestServerCert 2010-05-08T03:01:02.004Z ASCACKCEVSQ6CEXAMPLE3 2012-05-08T03:01:02.004Z 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data ListServerCertificates

Request constructor

listServerCertificates :: ListServerCertificates

Smart constructor for the minimum required parameters to construct a valid ListServerCertificates request.

The fields accessible through corresponding lenses are:

PathPrefix :: Maybe Text

Marker :: Maybe Text

MaxItems :: Maybe Integer

Request lenses

lscPathPrefix :: Lens' ListServerCertificates (Maybe Text)

The path prefix for filtering the results. For example: companyservercerts would get all server certificates for which the path starts with companyservercerts. This parameter is optional. If it is not included, it defaults to a slash (/), listing all server certificates.

lscMarker :: Lens' ListServerCertificates (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lscMaxItems :: Lens' ListServerCertificates (Maybe Integer)

Use this only when paginating results to indicate the maximum number of server certificates you want in the response. If there are additional server certificates beyond the maximum you specify, the IsTruncated response element will be set to true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListServerCertificatesResponse

Contains the result of a successful invocation of the ListServerCertificates action.

Response constructor

listServerCertificatesResponse

Arguments

:: [ServerCertificateMetadata]
lscrServerCertificateMetadataList

-> Bool
lscrIsTruncated

-> ListServerCertificatesResponse

Smart constructor for the minimum required parameters to construct a valid ListServerCertificatesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ServerCertificateMetadataList :: [ServerCertificateMetadata]

IsTruncated :: Bool

Marker :: Maybe Text

Response lenses

lscrServerCertificateMetadataList :: Lens' ListServerCertificatesResponse [ServerCertificateMetadata]

A list of server certificates.

lscrIsTruncated :: Lens' ListServerCertificatesResponse Bool

A flag that indicates whether there are more server certificates to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more server certificates in the list.

lscrMarker :: Lens' ListServerCertificatesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.