Network.AWS.IAM.ListServiceSpecificCredentials

Description

Returns information about the service-specific credentials associated with the specified IAM user. If there are none, the action returns an empty list. The service-specific credentials returned by this action are used only for authenticating the IAM user to a specific service. For more information about using service-specific credentials to authenticate to an AWS service, see Set Up service-specific credentials in the AWS CodeCommit User Guide.

Creating a Request

listServiceSpecificCredentials :: ListServiceSpecificCredentials

Creates a value of ListServiceSpecificCredentials with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsscUserName - The name of the user whose service-specific credentials you want information about. If this value is not specified then the operation assumes the user whose credentials are used to call the operation. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

lsscServiceName - Filters the returned results to only those for the specified AWS service. If not specified, then AWS returns service-specific credentials for all services.

data ListServiceSpecificCredentials

See: listServiceSpecificCredentials smart constructor.

Request Lenses

lsscUserName :: Lens' ListServiceSpecificCredentials (Maybe Text)

The name of the user whose service-specific credentials you want information about. If this value is not specified then the operation assumes the user whose credentials are used to call the operation. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

lsscServiceName :: Lens' ListServiceSpecificCredentials (Maybe Text)

Filters the returned results to only those for the specified AWS service. If not specified, then AWS returns service-specific credentials for all services.

Destructuring the Response

listServiceSpecificCredentialsResponse

Arguments

:: Int
lsscrsResponseStatus

-> ListServiceSpecificCredentialsResponse

Creates a value of ListServiceSpecificCredentialsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsscrsServiceSpecificCredentials - A list of structures that each contain details about a service-specific credential.

lsscrsResponseStatus - -- | The response status code.

data ListServiceSpecificCredentialsResponse

See: listServiceSpecificCredentialsResponse smart constructor.

Response Lenses

lsscrsServiceSpecificCredentials :: Lens' ListServiceSpecificCredentialsResponse [ServiceSpecificCredentialMetadata]

A list of structures that each contain details about a service-specific credential.

lsscrsResponseStatus :: Lens' ListServiceSpecificCredentialsResponse Int

- | The response status code.