Network.AWS.IAM.ListSigningCertificates

Description

Returns information about the signing certificates associated with the specified user. If there are none, the action returns an empty list.

Although each user is limited to a small number of signing certificates, you can still paginate the results using the MaxItems and Marker parameters.

If the UserName field is not specified, the user name is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListSigningCertificates.html

Request

data ListSigningCertificates

Request constructor

listSigningCertificates :: ListSigningCertificates

ListSigningCertificates constructor.

The fields accessible through corresponding lenses are:

lsc1Marker :: Maybe Text

lsc1MaxItems :: Maybe Natural

lsc1UserName :: Maybe Text

Request lenses

lsc1Marker :: Lens' ListSigningCertificates (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lsc1MaxItems :: Lens' ListSigningCertificates (Maybe Natural)

Use this only when paginating results to indicate the maximum number of certificate IDs you want in the response. If there are additional certificate IDs beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lsc1UserName :: Lens' ListSigningCertificates (Maybe Text)

The name of the user.

Response

data ListSigningCertificatesResponse

Response constructor

listSigningCertificatesResponse :: ListSigningCertificatesResponse

ListSigningCertificatesResponse constructor.

The fields accessible through corresponding lenses are:

lscr1Certificates :: [SigningCertificate]

lscr1IsTruncated :: Maybe Bool

lscr1Marker :: Maybe Text

Response lenses

lscr1Certificates :: Lens' ListSigningCertificatesResponse [SigningCertificate]

A list of the user's signing certificate information.

lscr1IsTruncated :: Lens' ListSigningCertificatesResponse (Maybe Bool)

A flag that indicates whether there are more certificate IDs to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more certificates in the list.

lscr1Marker :: Lens' ListSigningCertificatesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.