Network.AWS.IAM.ListUserPolicies

Description

Lists the names of the inline policies embedded in the specified user.

A user can also have managed policies attached to it. To list the managed policies that are attached to a user, use ListAttachedUserPolicies. For more information about policies, refer to Managed Policies and Inline Policies in the Using IAM guide.

You can paginate the results using the MaxItems and Marker parameters. If there are no inline policies embedded with the specified user, the action returns an empty list.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListUserPolicies.html

Request

data ListUserPolicies

Request constructor

listUserPolicies

Arguments

:: Text
lupUserName

-> ListUserPolicies

ListUserPolicies constructor.

The fields accessible through corresponding lenses are:

lupMarker :: Maybe Text

lupMaxItems :: Maybe Natural

lupUserName :: Text

Request lenses

lupMarker :: Lens' ListUserPolicies (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lupMaxItems :: Lens' ListUserPolicies (Maybe Natural)

Use this only when paginating results to indicate the maximum number of policy names you want in the response. If there are additional policy names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lupUserName :: Lens' ListUserPolicies Text

The name of the user to list policies for.

Response

data ListUserPoliciesResponse

Response constructor

listUserPoliciesResponse :: ListUserPoliciesResponse

ListUserPoliciesResponse constructor.

The fields accessible through corresponding lenses are:

luprIsTruncated :: Maybe Bool

luprMarker :: Maybe Text

luprPolicyNames :: [Text]

Response lenses

luprIsTruncated :: Lens' ListUserPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more policy names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policy names in the list.

luprMarker :: Lens' ListUserPoliciesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

luprPolicyNames :: Lens' ListUserPoliciesResponse [Text]

A list of policy names.