Network.AWS.IAM.ListUsers

Description

Lists the IAM users that have the specified path prefix. If no path prefix is specified, the action returns all users in the AWS account. If there are none, the action returns an empty list. You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListUsers.html

Request

data ListUsers

Request constructor

listUsers :: ListUsers

ListUsers constructor.

The fields accessible through corresponding lenses are:

luMarker :: Maybe Text

luMaxItems :: Maybe Natural

luPathPrefix :: Maybe Text

Request lenses

luMarker :: Lens' ListUsers (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

luMaxItems :: Lens' ListUsers (Maybe Natural)

Use this parameter only when paginating results to indicate the maximum number of user names you want in the response. If there are additional user names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

luPathPrefix :: Lens' ListUsers (Maybe Text)

The path prefix for filtering the results. For example: division_abcsubdivision_xyz/, which would get all user names whose path starts with division_abcsubdivision_xyz/. This parameter is optional. If it is not included, it defaults to a slash (/), listing all user names.

Response

data ListUsersResponse

Response constructor

listUsersResponse :: ListUsersResponse

ListUsersResponse constructor.

The fields accessible through corresponding lenses are:

lurIsTruncated :: Maybe Bool

lurMarker :: Maybe Text

lurUsers :: [User]

Response lenses

lurIsTruncated :: Lens' ListUsersResponse (Maybe Bool)

A flag that indicates whether there are more user names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more users in the list.

lurMarker :: Lens' ListUsersResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lurUsers :: Lens' ListUsersResponse [User]

A list of users.