Network.AWS.IAM.ListVirtualMFADevices

Description

Lists the virtual MFA devices under the AWS account by assignment status. If you do not specify an assignment status, the action returns a list of all virtual MFA devices. Assignment status can be Assigned, Unassigned, or Any. You can paginate the results using the MaxItems and Marker parameters. the AssignmentStatus is Any --> https:/iam.amazonaws.com ?Action=ListVirtualMFADevices &AssignmentStatus=Any &AUTHPARAMS associated with the account: the first device is unassigned, the second is assigned to the root account, and the third is assigned to a user named ExampleUser under the account. --> false arn:aws:iam::123456789012:mfa/MFAdeviceName arn:aws:iam::123456789012:mfa/RootMFAdeviceName 2011-10-20T20:49:03Z 123456789012 arn:aws:iam::123456789012:root 2009-10-13T22:00:36Z arn:aws:iam:::mfa/ExampleUserMFAdeviceName 2011-10-31T20:45:02Z AIDEXAMPLE4EXAMPLEXYZ / ExampleUser arn:aws:iam::111122223333:user/ExampleUser 2011-07-01T17:23:07Z b61ce1b1-0401-11e1-b2f8-2dEXAMPLEbfc.

Request

data ListVirtualMFADevices

Request constructor

listVirtualMFADevices :: ListVirtualMFADevices

Smart constructor for the minimum required parameters to construct a valid ListVirtualMFADevices request.

The fields accessible through corresponding lenses are:

AssignmentStatus :: Maybe AssignmentStatusType

Marker :: Maybe Text

MaxItems :: Maybe Integer

Request lenses

lvmfadAssignmentStatus :: Lens' ListVirtualMFADevices (Maybe AssignmentStatusType)

The status (unassigned or assigned) of the devices to list. If you do not specify an AssignmentStatus, the action defaults to Any which lists both assigned and unassigned virtual MFA devices.

lvmfadMarker :: Lens' ListVirtualMFADevices (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lvmfadMaxItems :: Lens' ListVirtualMFADevices (Maybe Integer)

Use this parameter only when paginating results to indicate the maximum number of user names you want in the response. If there are additional user names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListVirtualMFADevicesResponse

Contains the result of a successful invocation of the ListVirtualMFADevices action.

Response constructor

listVirtualMFADevicesResponse

Arguments

:: [VirtualMFADevice]
lvmfadrVirtualMFADevices

-> Bool
lvmfadrIsTruncated

-> ListVirtualMFADevicesResponse

Smart constructor for the minimum required parameters to construct a valid ListVirtualMFADevicesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VirtualMFADevices :: [VirtualMFADevice]

IsTruncated :: Bool

Marker :: Maybe Text

Response lenses

lvmfadrVirtualMFADevices :: Lens' ListVirtualMFADevicesResponse [VirtualMFADevice]

lvmfadrIsTruncated :: Lens' ListVirtualMFADevicesResponse Bool

A flag that indicates whether there are more items to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items the list.

lvmfadrMarker :: Lens' ListVirtualMFADevicesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.