Network.AWS.IAM.PutGroupPolicy

Description

Adds (or updates) an inline policy document that is embedded in the specified group.

A user can also have managed policies attached to it. To attach a managed policy to a group, use AttachGroupPolicy. To create a new managed policy, use CreatePolicy. For information about policies, refer to Managed Policies andInline Policies in the Using IAM guide.

For information about limits on the number of inline policies that you can embed in a group, see Limitations on IAM Entities in the Using IAM guide.

Because policy documents can be large, you should use POST rather than GET when calling PutGroupPolicy. For general information about using the Query API with IAM, go to Making Query Requests in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_PutGroupPolicy.html

Request

data PutGroupPolicy

Request constructor

putGroupPolicy

Arguments

:: Text
pgpGroupName

-> Text
pgpPolicyName

-> Text
pgpPolicyDocument

-> PutGroupPolicy

PutGroupPolicy constructor.

The fields accessible through corresponding lenses are:

pgpGroupName :: Text

pgpPolicyDocument :: Text

pgpPolicyName :: Text

Request lenses

pgpGroupName :: Lens' PutGroupPolicy Text

The name of the group to associate the policy with.

pgpPolicyDocument :: Lens' PutGroupPolicy Text

The policy document.

pgpPolicyName :: Lens' PutGroupPolicy Text

The name of the policy document.

Response

data PutGroupPolicyResponse

Response constructor

putGroupPolicyResponse :: PutGroupPolicyResponse

PutGroupPolicyResponse constructor.