Network.AWS.IAM.PutRolePolicy

Description

Adds (or updates) an inline policy document that is embedded in the specified role.

When you embed an inline policy in a role, the inline policy is used as the role's access (permissions) policy. The role's trust policy is created at the same time as the role, using CreateRole. You can update a role's trust policy using UpdateAssumeRolePolicy. For more information about roles, go to UsingRoles to Delegate Permissions and Federate Identities.

A role can also have a managed policy attached to it. To attach a managed policy to a role, use AttachRolePolicy. To create a new managed policy, use CreatePolicy. For information about policies, refer to Managed Policies and InlinePolicies in the Using IAM guide.

For information about limits on the number of inline policies that you can embed with a role, see Limitations on IAM Entities in the Using IAM guide.

Because policy documents can be large, you should use POST rather than GET when calling PutRolePolicy. For general information about using the Query API with IAM, go to Making Query Requests in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_PutRolePolicy.html

Request

data PutRolePolicy

Request constructor

putRolePolicy

Arguments

:: Text
prpRoleName

-> Text
prpPolicyName

-> Text
prpPolicyDocument

-> PutRolePolicy

PutRolePolicy constructor.

The fields accessible through corresponding lenses are:

prpPolicyDocument :: Text

prpPolicyName :: Text

prpRoleName :: Text

Request lenses

prpPolicyDocument :: Lens' PutRolePolicy Text

The policy document.

prpPolicyName :: Lens' PutRolePolicy Text

The name of the policy document.

prpRoleName :: Lens' PutRolePolicy Text

The name of the role to associate the policy with.

Response

data PutRolePolicyResponse

Response constructor

putRolePolicyResponse :: PutRolePolicyResponse

PutRolePolicyResponse constructor.