Network.AWS.IAM.PutUserPolicy

Description

Adds (or updates) an inline policy document that is embedded in the specified user.

A user can also have a managed policy attached to it. To attach a managed policy to a user, use AttachUserPolicy. To create a new managed policy, use CreatePolicy. For information about policies, refer to Managed Policies and InlinePolicies in the Using IAM guide.

For information about limits on the number of inline policies that you can embed in a user, see Limitations on IAM Entities in the Using IAM guide.

Because policy documents can be large, you should use POST rather than GET when calling PutUserPolicy. For general information about using the Query API with IAM, go to Making Query Requests in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_PutUserPolicy.html

Request

data PutUserPolicy

Request constructor

putUserPolicy

Arguments

:: Text
pupUserName

-> Text
pupPolicyName

-> Text
pupPolicyDocument

-> PutUserPolicy

PutUserPolicy constructor.

The fields accessible through corresponding lenses are:

pupPolicyDocument :: Text

pupPolicyName :: Text

pupUserName :: Text

Request lenses

pupPolicyDocument :: Lens' PutUserPolicy Text

The policy document.

pupPolicyName :: Lens' PutUserPolicy Text

The name of the policy document.

pupUserName :: Lens' PutUserPolicy Text

The name of the user to associate the policy with.

Response

data PutUserPolicyResponse

Response constructor

putUserPolicyResponse :: PutUserPolicyResponse

PutUserPolicyResponse constructor.