Network.AWS.IAM.RemoveRoleFromInstanceProfile

Description

Removes the specified role from the specified instance profile. Make sure you do not have any Amazon EC2 instances running with the role you are about to remove from the instance profile. Removing a role from an instance profile that is associated with a running instance will break any applications running on the instance. For more information about roles, go to Working with Roles. For more information about instance profiles, go to About Instance Profiles. https:/iam.amazonaws.com ?Action=RemoveRoleFromInstanceProfile &InstanceProfileName=Webserver &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS 29f47818-99f5-11e1-a4c3-27EXAMPLE804.

Request

data RemoveRoleFromInstanceProfile

Request constructor

removeRoleFromInstanceProfile

Arguments

:: Text
rrfipInstanceProfileName

-> Text
rrfipRoleName

-> RemoveRoleFromInstanceProfile

Smart constructor for the minimum required parameters to construct a valid RemoveRoleFromInstanceProfile request.

The fields accessible through corresponding lenses are:

InstanceProfileName :: Text

RoleName :: Text

Request lenses

rrfipInstanceProfileName :: Lens' RemoveRoleFromInstanceProfile Text

Name of the instance profile to update.

rrfipRoleName :: Lens' RemoveRoleFromInstanceProfile Text

Name of the role to remove.

Response

data RemoveRoleFromInstanceProfileResponse

Response constructor

removeRoleFromInstanceProfileResponse :: RemoveRoleFromInstanceProfileResponse

Smart constructor for the minimum required parameters to construct a valid RemoveRoleFromInstanceProfileResponse response.

This constructor is provided for convenience and testing purposes.