Network.AWS.IAM.ResetServiceSpecificCredential

Description

Resets the password for a service-specific credential. The new password is AWS generated and cryptographically strong. It cannot be configured by the user. Resetting the password immediately invalidates the previous password associated with this user.

Creating a Request

resetServiceSpecificCredential

Arguments

:: Text
rsscServiceSpecificCredentialId

-> ResetServiceSpecificCredential

Creates a value of ResetServiceSpecificCredential with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsscUserName - The name of the IAM user associated with the service-specific credential. If this value is not specified, then the operation assumes the user whose credentials are used to call the operation. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

rsscServiceSpecificCredentialId - The unique identifier of the service-specific credential. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

data ResetServiceSpecificCredential

See: resetServiceSpecificCredential smart constructor.

Request Lenses

rsscUserName :: Lens' ResetServiceSpecificCredential (Maybe Text)

The name of the IAM user associated with the service-specific credential. If this value is not specified, then the operation assumes the user whose credentials are used to call the operation. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

rsscServiceSpecificCredentialId :: Lens' ResetServiceSpecificCredential Text

The unique identifier of the service-specific credential. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

Destructuring the Response

resetServiceSpecificCredentialResponse

Arguments

:: Int
rsscrsResponseStatus

-> ResetServiceSpecificCredentialResponse

Creates a value of ResetServiceSpecificCredentialResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsscrsServiceSpecificCredential - A structure with details about the updated service-specific credential, including the new password. Important: This is the only time that you can access the password. You cannot recover the password later, but you can reset it again.

rsscrsResponseStatus - -- | The response status code.

data ResetServiceSpecificCredentialResponse

See: resetServiceSpecificCredentialResponse smart constructor.

Response Lenses

rsscrsServiceSpecificCredential :: Lens' ResetServiceSpecificCredentialResponse (Maybe ServiceSpecificCredential)

A structure with details about the updated service-specific credential, including the new password. Important: This is the only time that you can access the password. You cannot recover the password later, but you can reset it again.

rsscrsResponseStatus :: Lens' ResetServiceSpecificCredentialResponse Int

- | The response status code.