Network.AWS.IAM.ResyncMFADevice

Description

Synchronizes the specified MFA device with its IAM resource object on the AWS servers.

For more information about creating and working with virtual MFA devices, go to Using a Virtual MFA Device in the IAM User Guide .

Creating a Request

resyncMFADevice

Arguments

:: Text
rmdUserName

-> Text
rmdSerialNumber

-> Text
rmdAuthenticationCode1

-> Text
rmdAuthenticationCode2

-> ResyncMFADevice

Creates a value of ResyncMFADevice with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rmdUserName - The name of the user whose MFA device you want to resynchronize. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

rmdSerialNumber - Serial number that uniquely identifies the MFA device. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

rmdAuthenticationCode1 - An authentication code emitted by the device. The format for this parameter is a sequence of six digits.

rmdAuthenticationCode2 - A subsequent authentication code emitted by the device. The format for this parameter is a sequence of six digits.

data ResyncMFADevice

See: resyncMFADevice smart constructor.

Request Lenses

rmdUserName :: Lens' ResyncMFADevice Text

The name of the user whose MFA device you want to resynchronize. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

rmdSerialNumber :: Lens' ResyncMFADevice Text

Serial number that uniquely identifies the MFA device. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

rmdAuthenticationCode1 :: Lens' ResyncMFADevice Text

An authentication code emitted by the device. The format for this parameter is a sequence of six digits.

rmdAuthenticationCode2 :: Lens' ResyncMFADevice Text

A subsequent authentication code emitted by the device. The format for this parameter is a sequence of six digits.

Destructuring the Response

resyncMFADeviceResponse :: ResyncMFADeviceResponse

Creates a value of ResyncMFADeviceResponse with the minimum fields required to make a request.

data ResyncMFADeviceResponse

See: resyncMFADeviceResponse smart constructor.