Network.AWS.IAM.SetDefaultPolicyVersion

Description

Sets the specified version of the specified policy as the policy's default (operative) version.

This action affects all users, groups, and roles that the policy is attached to. To list the users, groups, and roles that the policy is attached to, use the ListEntitiesForPolicy API.

For information about managed policies, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

setDefaultPolicyVersion

Arguments

:: Text
sdpvPolicyARN

-> Text
sdpvVersionId

-> SetDefaultPolicyVersion

Creates a value of SetDefaultPolicyVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sdpvPolicyARN - The Amazon Resource Name (ARN) of the IAM policy whose default version you want to set. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

sdpvVersionId - The version of the policy to set as the default (operative) version. For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

data SetDefaultPolicyVersion

See: setDefaultPolicyVersion smart constructor.

Request Lenses

sdpvPolicyARN :: Lens' SetDefaultPolicyVersion Text

The Amazon Resource Name (ARN) of the IAM policy whose default version you want to set. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

sdpvVersionId :: Lens' SetDefaultPolicyVersion Text

The version of the policy to set as the default (operative) version. For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

Destructuring the Response

setDefaultPolicyVersionResponse :: SetDefaultPolicyVersionResponse

Creates a value of SetDefaultPolicyVersionResponse with the minimum fields required to make a request.

data SetDefaultPolicyVersionResponse

See: setDefaultPolicyVersionResponse smart constructor.