Network.AWS.IAM.Types

Service

data IAM

Version 2010-05-08 of the Amazon Identity and Access Management service.

Error

data RESTError :: *

XML

ns :: Text

AssignmentStatusType

data AssignmentStatusType

Constructors

Any
Any

Assigned
Assigned

Unassigned
Unassigned

PasswordPolicy

data PasswordPolicy

passwordPolicy :: PasswordPolicy

PasswordPolicy constructor.

The fields accessible through corresponding lenses are:

ppAllowUsersToChangePassword :: Maybe Bool

ppExpirePasswords :: Maybe Bool

ppHardExpiry :: Maybe Bool

ppMaxPasswordAge :: Maybe Natural

ppMinimumPasswordLength :: Maybe Natural

ppPasswordReusePrevention :: Maybe Natural

ppRequireLowercaseCharacters :: Maybe Bool

ppRequireNumbers :: Maybe Bool

ppRequireSymbols :: Maybe Bool

ppRequireUppercaseCharacters :: Maybe Bool

ppAllowUsersToChangePassword :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM users are allowed to change their own password.

ppExpirePasswords :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM users are required to change their password after a specified number of days.

ppHardExpiry :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM users are prevented from setting a new password after their password has expired.

ppMaxPasswordAge :: Lens' PasswordPolicy (Maybe Natural)

The number of days that an IAM user password is valid.

ppMinimumPasswordLength :: Lens' PasswordPolicy (Maybe Natural)

Minimum length to require for IAM user passwords.

ppPasswordReusePrevention :: Lens' PasswordPolicy (Maybe Natural)

Specifies the number of previous passwords that IAM users are prevented from reusing.

ppRequireLowercaseCharacters :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether to require lowercase characters for IAM user passwords.

ppRequireNumbers :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether to require numbers for IAM user passwords.

ppRequireSymbols :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether to require symbols for IAM user passwords.

ppRequireUppercaseCharacters :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether to require uppercase characters for IAM user passwords.

Group

data Group

group

Arguments

:: Text
gPath

-> Text
gGroupName

-> Text
gGroupId

-> Text
gArn

-> UTCTime
gCreateDate

-> Group

Group constructor.

The fields accessible through corresponding lenses are:

gArn :: Text

gCreateDate :: UTCTime

gGroupId :: Text

gGroupName :: Text

gPath :: Text

gArn :: Lens' Group Text

The Amazon Resource Name (ARN) specifying the group. For more information about ARNs and how to use them in policies, see IAM Identifiers in the Using IAM guide.

gCreateDate :: Lens' Group UTCTime

The date when the group was created.

gGroupId :: Lens' Group Text

The stable and unique string identifying the group. For more information about IDs, see IAM Identifiers in the Using IAM guide.

gGroupName :: Lens' Group Text

The name that identifies the group.

gPath :: Lens' Group Text

The path to the group. For more information about paths, see IAM Identifiers in the Using IAM guide.

MFADevice

data MFADevice

mfadevice

Arguments

:: Text
mfadUserName

-> Text
mfadSerialNumber

-> UTCTime
mfadEnableDate

-> MFADevice

MFADevice constructor.

The fields accessible through corresponding lenses are:

mfadEnableDate :: UTCTime

mfadSerialNumber :: Text

mfadUserName :: Text

mfadEnableDate :: Lens' MFADevice UTCTime

The date when the MFA device was enabled for the user.

mfadSerialNumber :: Lens' MFADevice Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

mfadUserName :: Lens' MFADevice Text

The user with whom the MFA device is associated.

InstanceProfile

data InstanceProfile

instanceProfile

Arguments

:: Text
ipPath

-> Text
ipInstanceProfileName

-> Text
ipInstanceProfileId

-> Text
ipArn

-> UTCTime
ipCreateDate

-> InstanceProfile

InstanceProfile constructor.

The fields accessible through corresponding lenses are:

ipArn :: Text

ipCreateDate :: UTCTime

ipInstanceProfileId :: Text

ipInstanceProfileName :: Text

ipPath :: Text

ipRoles :: [Role]

ipArn :: Lens' InstanceProfile Text

The Amazon Resource Name (ARN) specifying the instance profile. For more information about ARNs and how to use them in policies, see IAM Identifiers in the Using IAM guide.

ipCreateDate :: Lens' InstanceProfile UTCTime

The date when the instance profile was created.

ipInstanceProfileId :: Lens' InstanceProfile Text

The stable and unique string identifying the instance profile. For more information about IDs, see IAM Identifiers in the Using IAM guide.

ipInstanceProfileName :: Lens' InstanceProfile Text

The name identifying the instance profile.

ipPath :: Lens' InstanceProfile Text

The path to the instance profile. For more information about paths, see IAM Identifiers in the Using IAM guide.

ipRoles :: Lens' InstanceProfile [Role]

The role associated with the instance profile.

ReportFormatType

data ReportFormatType

Constructors

TextCsv
text/csv

ServerCertificateMetadata

data ServerCertificateMetadata

serverCertificateMetadata

Arguments

:: Text
scmPath

-> Text
scmServerCertificateName

-> Text
scmServerCertificateId

-> Text
scmArn

-> ServerCertificateMetadata

ServerCertificateMetadata constructor.

The fields accessible through corresponding lenses are:

scmArn :: Text

scmExpiration :: Maybe UTCTime

scmPath :: Text

scmServerCertificateId :: Text

scmServerCertificateName :: Text

scmUploadDate :: Maybe UTCTime

scmArn :: Lens' ServerCertificateMetadata Text

The Amazon Resource Name (ARN) specifying the server certificate. For more information about ARNs and how to use them in policies, see IAM Identifiers in the Using IAM guide.

scmExpiration :: Lens' ServerCertificateMetadata (Maybe UTCTime)

The date on which the certificate is set to expire.

scmPath :: Lens' ServerCertificateMetadata Text

The path to the server certificate. For more information about paths, see IAM Identifiers in the Using IAM guide.

scmServerCertificateId :: Lens' ServerCertificateMetadata Text

The stable and unique string identifying the server certificate. For more information about IDs, see IAM Identifiers in the Using IAM guide.

scmServerCertificateName :: Lens' ServerCertificateMetadata Text

The name that identifies the server certificate.

scmUploadDate :: Lens' ServerCertificateMetadata (Maybe UTCTime)

The date when the server certificate was uploaded.

OpenIDConnectProviderListEntry

data OpenIDConnectProviderListEntry

openIDConnectProviderListEntry :: OpenIDConnectProviderListEntry

OpenIDConnectProviderListEntry constructor.

The fields accessible through corresponding lenses are:

oidcpleArn :: Maybe Text

oidcpleArn :: Lens' OpenIDConnectProviderListEntry (Maybe Text)

LoginProfile

data LoginProfile

loginProfile

Arguments

:: Text
lpUserName

-> UTCTime
lpCreateDate

-> LoginProfile

LoginProfile constructor.

The fields accessible through corresponding lenses are:

lpCreateDate :: UTCTime

lpPasswordResetRequired :: Maybe Bool

lpUserName :: Text

lpCreateDate :: Lens' LoginProfile UTCTime

The date when the password for the user was created.

lpPasswordResetRequired :: Lens' LoginProfile (Maybe Bool)

Specifies whether the user is required to set a new password on next sign-in.

lpUserName :: Lens' LoginProfile Text

The name of the user, which can be used for signing in to the AWS Management Console.

SummaryKeyType

data SummaryKeyType

Constructors

AccessKeysPerUserQuota
AccessKeysPerUserQuota

AccountMFAEnabled
AccountMFAEnabled

GroupPolicySizeQuota
GroupPolicySizeQuota

Groups
Groups

GroupsPerUserQuota
GroupsPerUserQuota

GroupsQuota
GroupsQuota

MFADevices
MFADevices

MFADevicesInUse
MFADevicesInUse

ServerCertificates
ServerCertificates

ServerCertificatesQuota
ServerCertificatesQuota

SigningCertificatesPerUserQuota
SigningCertificatesPerUserQuota

UserPolicySizeQuota
UserPolicySizeQuota

Users
Users

UsersQuota
UsersQuota

ReportStateType

data ReportStateType

Constructors

Complete
COMPLETE

Inprogress
INPROGRESS

Started
STARTED

User

data User

user

Arguments

:: Text
uPath

-> Text
uUserName

-> Text
uUserId

-> Text
uArn

-> UTCTime
uCreateDate

-> User

User constructor.

The fields accessible through corresponding lenses are:

uArn :: Text

uCreateDate :: UTCTime

uPasswordLastUsed :: Maybe UTCTime

uPath :: Text

uUserId :: Text

uUserName :: Text

uArn :: Lens' User Text

The Amazon Resource Name (ARN) that identifies the user. For more information about ARNs and how to use ARNs in policies, see IAM Identifiers in the Using IAM guide.

uCreateDate :: Lens' User UTCTime

The date and time, in ISO 8601 date-time format, when the user was created.

uPasswordLastUsed :: Lens' User (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the user's password was last used to sign in to an AWS website. For a list of AWS websites that capture a user's last sign-in time, see the Credential Reports topic in the Using IAM guide. If a password is used more than once in a five-minute span, only the first use is returned in this field. When the user does not have a password, this field is null (not present). When a user's password exists but has never been used, or when there is no sign-in data associated with the user, this field is null (not present). This value is returned only in the GetUser and ListUsers actions.

uPath :: Lens' User Text

The path to the user. For more information about paths, see IAM Identifiers in the Using IAM guide.

uUserId :: Lens' User Text

The stable and unique string identifying the user. For more information about IDs, see IAM Identifiers in the Using IAM guide.

uUserName :: Lens' User Text

The friendly name identifying the user.

StatusType

data StatusType

Constructors

Active
Active

Inactive
Inactive

SAMLProviderListEntry

data SAMLProviderListEntry

samlproviderListEntry :: SAMLProviderListEntry

SAMLProviderListEntry constructor.

The fields accessible through corresponding lenses are:

samlpleArn :: Maybe Text

samlpleCreateDate :: Maybe UTCTime

samlpleValidUntil :: Maybe UTCTime

samlpleArn :: Lens' SAMLProviderListEntry (Maybe Text)

The Amazon Resource Name (ARN) of the SAML provider.

samlpleCreateDate :: Lens' SAMLProviderListEntry (Maybe UTCTime)

The date and time when the SAML provider was created.

samlpleValidUntil :: Lens' SAMLProviderListEntry (Maybe UTCTime)

The expiration date and time for the SAML provider.

Role

data Role

role

Arguments

:: Text
rPath

-> Text
rRoleName

-> Text
rRoleId

-> Text
rArn

-> UTCTime
rCreateDate

-> Role

Role constructor.

The fields accessible through corresponding lenses are:

rArn :: Text

rAssumeRolePolicyDocument :: Maybe Text

rCreateDate :: UTCTime

rPath :: Text

rRoleId :: Text

rRoleName :: Text

rArn :: Lens' Role Text

The Amazon Resource Name (ARN) specifying the role. For more information about ARNs and how to use them in policies, see IAM Identifiers in the Using IAM guide.

rAssumeRolePolicyDocument :: Lens' Role (Maybe Text)

The policy that grants an entity permission to assume the role. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcs/rfc3986.html.

rCreateDate :: Lens' Role UTCTime

The date when the role was created.

rPath :: Lens' Role Text

The path to the role. For more information about paths, see IAM Identifiers in the Using IAM guide.

rRoleId :: Lens' Role Text

The stable and unique string identifying the role. For more information about IDs, see IAM Identifiers in the Using IAM guide.

rRoleName :: Lens' Role Text

The name that identifies the role.

ServerCertificate

data ServerCertificate

serverCertificate

Arguments

:: ServerCertificateMetadata
scServerCertificateMetadata

-> Text
scCertificateBody

-> ServerCertificate

ServerCertificate constructor.

The fields accessible through corresponding lenses are:

scCertificateBody :: Text

scCertificateChain :: Maybe Text

scServerCertificateMetadata :: ServerCertificateMetadata

scCertificateBody :: Lens' ServerCertificate Text

The contents of the public key certificate.

scCertificateChain :: Lens' ServerCertificate (Maybe Text)

The contents of the public key certificate chain.

scServerCertificateMetadata :: Lens' ServerCertificate ServerCertificateMetadata

The meta information of the server certificate, such as its name, path, ID, and ARN.

AccessKey

data AccessKey

accessKey

Arguments

:: Text
akUserName

-> Text
akAccessKeyId

-> StatusType
akStatus

-> Text
akSecretAccessKey

-> AccessKey

AccessKey constructor.

The fields accessible through corresponding lenses are:

akAccessKeyId :: Text

akCreateDate :: Maybe UTCTime

akSecretAccessKey :: Text

akStatus :: StatusType

akUserName :: Text

akAccessKeyId :: Lens' AccessKey Text

The ID for this access key.

akCreateDate :: Lens' AccessKey (Maybe UTCTime)

The date when the access key was created.

akSecretAccessKey :: Lens' AccessKey Text

The secret key used to sign requests.

akStatus :: Lens' AccessKey StatusType

The status of the access key. Active means the key is valid for API calls, while Inactive means it is not.

akUserName :: Lens' AccessKey Text

The name of the IAM user that the access key is associated with.

VirtualMFADevice

data VirtualMFADevice

virtualMFADevice

Arguments

:: Text
vmfadSerialNumber

-> VirtualMFADevice

VirtualMFADevice constructor.

The fields accessible through corresponding lenses are:

vmfadBase32StringSeed :: Maybe Base64

vmfadEnableDate :: Maybe UTCTime

vmfadQRCodePNG :: Maybe Base64

vmfadSerialNumber :: Text

vmfadUser :: Maybe User

vmfadBase32StringSeed :: Lens' VirtualMFADevice (Maybe Base64)

The Base32 seed defined as specified in RFC3548. The Base32StringSeed is Base64-encoded.

vmfadEnableDate :: Lens' VirtualMFADevice (Maybe UTCTime)

The date and time on which the virtual MFA device was enabled.

vmfadQRCodePNG :: Lens' VirtualMFADevice (Maybe Base64)

A QR code PNG image that encodes otpauth:/totp$virtualMFADeviceName@$AccountName?secret=$Base32String where $virtualMFADeviceName is one of the create call arguments, AccountName is the user name if set (otherwise, the account ID otherwise), and Base32String is the seed in Base32 format. The Base32String value is Base64-encoded.

vmfadSerialNumber :: Lens' VirtualMFADevice Text

The serial number associated with VirtualMFADevice.

vmfadUser :: Lens' VirtualMFADevice (Maybe User)

SigningCertificate

data SigningCertificate

signingCertificate

Arguments

:: Text
sc1UserName

-> Text
sc1CertificateId

-> Text
sc1CertificateBody

-> StatusType
sc1Status

-> SigningCertificate

SigningCertificate constructor.

The fields accessible through corresponding lenses are:

sc1CertificateBody :: Text

sc1CertificateId :: Text

sc1Status :: StatusType

sc1UploadDate :: Maybe UTCTime

sc1UserName :: Text

sc1CertificateBody :: Lens' SigningCertificate Text

The contents of the signing certificate.

sc1CertificateId :: Lens' SigningCertificate Text

The ID for the signing certificate.

sc1Status :: Lens' SigningCertificate StatusType

The status of the signing certificate. Active means the key is valid for API calls, while Inactive means it is not.

sc1UploadDate :: Lens' SigningCertificate (Maybe UTCTime)

The date when the signing certificate was uploaded.

sc1UserName :: Lens' SigningCertificate Text

The name of the user the signing certificate is associated with.

AccessKeyMetadata

data AccessKeyMetadata

accessKeyMetadata :: AccessKeyMetadata

AccessKeyMetadata constructor.

The fields accessible through corresponding lenses are:

akmAccessKeyId :: Maybe Text

akmCreateDate :: Maybe UTCTime

akmStatus :: Maybe StatusType

akmUserName :: Maybe Text

akmAccessKeyId :: Lens' AccessKeyMetadata (Maybe Text)

The ID for this access key.

akmCreateDate :: Lens' AccessKeyMetadata (Maybe UTCTime)

The date when the access key was created.

akmStatus :: Lens' AccessKeyMetadata (Maybe StatusType)

The status of the access key. Active means the key is valid for API calls; Inactive means it is not.

akmUserName :: Lens' AccessKeyMetadata (Maybe Text)

The name of the IAM user that the key is associated with.