Network.AWS.IAM.UpdateAccessKey

Description

Changes the status of the specified access key from Active to Inactive, or vice versa. This action can be used to disable a user's key as part of a key rotation work flow.

If the UserName field is not specified, the UserName is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

For information about rotating keys, see Managing Keys and Certificates in the IAM User Guide .

Creating a Request

updateAccessKey

Arguments

:: AccessKey
uakAccessKeyId

-> StatusType
uakStatus

-> UpdateAccessKey

Creates a value of UpdateAccessKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uakUserName - The name of the user whose key you want to update. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uakAccessKeyId - The access key ID of the secret access key you want to update. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

uakStatus - The status you want to assign to the secret access key. Active means the key can be used for API calls to AWS, while Inactive means the key cannot be used.

data UpdateAccessKey

See: updateAccessKey smart constructor.

Request Lenses

uakUserName :: Lens' UpdateAccessKey (Maybe Text)

The name of the user whose key you want to update. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uakAccessKeyId :: Lens' UpdateAccessKey AccessKey

The access key ID of the secret access key you want to update. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

uakStatus :: Lens' UpdateAccessKey StatusType

The status you want to assign to the secret access key. Active means the key can be used for API calls to AWS, while Inactive means the key cannot be used.

Destructuring the Response

updateAccessKeyResponse :: UpdateAccessKeyResponse

Creates a value of UpdateAccessKeyResponse with the minimum fields required to make a request.

data UpdateAccessKeyResponse

See: updateAccessKeyResponse smart constructor.