Network.AWS.IAM.UpdateAccountPasswordPolicy

Description

Updates the password policy settings for the account. For more information about using a password policy, see Managing an IAM Password Policy in the Using IAM guide. https:/iam.amazonaws.com ?Action=UpdateAccountPasswordPolicy &MinimumPasswordLength=12 &RequireSymbols=false &RequireNumbers=true &RequireUppercaseCharacters=true &RequireLowercaseCharacters=true &AllowUsersToChangePassword=true &MaxPasswordAge=90 &PasswordReusePrevention=6 &HardExpiry=false &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data UpdateAccountPasswordPolicy

Request constructor

updateAccountPasswordPolicy :: UpdateAccountPasswordPolicy

Smart constructor for the minimum required parameters to construct a valid UpdateAccountPasswordPolicy request.

The fields accessible through corresponding lenses are:

MinimumPasswordLength :: Maybe Integer

RequireSymbols :: Maybe Bool

RequireNumbers :: Maybe Bool

RequireUppercaseCharacters :: Maybe Bool

RequireLowercaseCharacters :: Maybe Bool

AllowUsersToChangePassword :: Maybe Bool

MaxPasswordAge :: Maybe Integer

PasswordReusePrevention :: Maybe Integer

HardExpiry :: Maybe Bool

Request lenses

uappMinimumPasswordLength :: Lens' UpdateAccountPasswordPolicy (Maybe Integer)

The minimum number of characters allowed in an IAM user password.

uappRequireSymbols :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one of the following non-alphanumeric characters: ! @ # $ % ^ &amp; * ( ) _ + - = [ ] { } | '.

uappRequireNumbers :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).

uappRequireUppercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one uppercase character from the ISO basic Latin alphabet (A to Z).

uappRequireLowercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one lowercase character from the ISO basic Latin alphabet (a to z).

uappAllowUsersToChangePassword :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Allows all IAM users in your account to use the AWS Management Console to change their own passwords. For more information, see Letting IAM Users Change Their Own Passwords in the Using IAM guide.

uappMaxPasswordAge :: Lens' UpdateAccountPasswordPolicy (Maybe Integer)

The number of days that an IAM user password is valid.

uappPasswordReusePrevention :: Lens' UpdateAccountPasswordPolicy (Maybe Integer)

Specifies the number of previous passwords that IAM users are prevented from reusing.

uappHardExpiry :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Prevents IAM users from setting a new password after their password has expired.

Response

data UpdateAccountPasswordPolicyResponse

Response constructor

updateAccountPasswordPolicyResponse :: UpdateAccountPasswordPolicyResponse

Smart constructor for the minimum required parameters to construct a valid UpdateAccountPasswordPolicyResponse response.

This constructor is provided for convenience and testing purposes.