Network.AWS.IAM.UpdateAssumeRolePolicy

Description

Updates the policy that grants an IAM entity permission to assume a role. This is typically referred to as the "role trust policy". For more information about roles, go to Using Roles to Delegate Permissions and Federate Identities .

Creating a Request

updateAssumeRolePolicy

Arguments

:: Text
uarpRoleName

-> Text
uarpPolicyDocument

-> UpdateAssumeRolePolicy

Creates a value of UpdateAssumeRolePolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uarpRoleName - The name of the role to update with the new policy. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

uarpPolicyDocument - The policy that grants an entity permission to assume the role. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

data UpdateAssumeRolePolicy

See: updateAssumeRolePolicy smart constructor.

Request Lenses

uarpRoleName :: Lens' UpdateAssumeRolePolicy Text

The name of the role to update with the new policy. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

uarpPolicyDocument :: Lens' UpdateAssumeRolePolicy Text

The policy that grants an entity permission to assume the role. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

Destructuring the Response

updateAssumeRolePolicyResponse :: UpdateAssumeRolePolicyResponse

Creates a value of UpdateAssumeRolePolicyResponse with the minimum fields required to make a request.

data UpdateAssumeRolePolicyResponse

See: updateAssumeRolePolicyResponse smart constructor.