Network.AWS.IAM.UpdateGroup

Description

Updates the name and/or the path of the specified group.

You should understand the implications of changing a group's path or name. For more information, see Renaming Users and Groups in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_UpdateGroup.html

Request

data UpdateGroup

Request constructor

updateGroup

Arguments

:: Text
ugGroupName

-> UpdateGroup

UpdateGroup constructor.

The fields accessible through corresponding lenses are:

ugGroupName :: Text

ugNewGroupName :: Maybe Text

ugNewPath :: Maybe Text

Request lenses

ugGroupName :: Lens' UpdateGroup Text

Name of the group to update. If you're changing the name of the group, this is the original name.

ugNewGroupName :: Lens' UpdateGroup (Maybe Text)

New name for the group. Only include this if changing the group's name.

ugNewPath :: Lens' UpdateGroup (Maybe Text)

New path for the group. Only include this if changing the group's path.

Response

data UpdateGroupResponse

Response constructor

updateGroupResponse :: UpdateGroupResponse

UpdateGroupResponse constructor.