Network.AWS.IAM.UpdateLoginProfile

Description

Changes the password for the specified IAM user.

IAM users can change their own passwords by calling ChangePassword . For more information about modifying passwords, see Managing Passwords in the IAM User Guide .

Creating a Request

updateLoginProfile

Arguments

:: Text
ulpUserName

-> UpdateLoginProfile

Creates a value of UpdateLoginProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ulpPassword - The new password for the specified IAM user. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D). However, the format can be further restricted by the account administrator by setting a password policy on the AWS account. For more information, see UpdateAccountPasswordPolicy .

ulpPasswordResetRequired - Allows this new password to be used only once by requiring the specified IAM user to set a new password on next sign-in.

ulpUserName - The name of the user whose password you want to update. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data UpdateLoginProfile

See: updateLoginProfile smart constructor.

Request Lenses

ulpPassword :: Lens' UpdateLoginProfile (Maybe Text)

The new password for the specified IAM user. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D). However, the format can be further restricted by the account administrator by setting a password policy on the AWS account. For more information, see UpdateAccountPasswordPolicy .

ulpPasswordResetRequired :: Lens' UpdateLoginProfile (Maybe Bool)

Allows this new password to be used only once by requiring the specified IAM user to set a new password on next sign-in.

ulpUserName :: Lens' UpdateLoginProfile Text

The name of the user whose password you want to update. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

updateLoginProfileResponse :: UpdateLoginProfileResponse

Creates a value of UpdateLoginProfileResponse with the minimum fields required to make a request.

data UpdateLoginProfileResponse

See: updateLoginProfileResponse smart constructor.