Network.AWS.IAM.UpdateOpenIDConnectProviderThumbprint

Description

Replaces the existing list of server certificate thumbprints with a new list.

The list that you pass with this action completely replaces the existing list of thumbprints. (The lists are not merged.)

Typically, you need to update a thumbprint only when the identity provider's certificate changes, which occurs rarely. However, if the provider's certificate does change, any attempt to assume an IAM role that specifies the OIDC provider as a principal will fail until the certificate thumbprint is updated.

Because trust for the OpenID Connect provider is ultimately derived from the provider's certificate and is validated by the thumbprint, it is a best practice to limit access to the UpdateOpenIDConnectProviderThumbprint action to highly-privileged users.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_UpdateOpenIDConnectProviderThumbprint.html

Request

data UpdateOpenIDConnectProviderThumbprint

Request constructor

updateOpenIDConnectProviderThumbprint

Arguments

:: Text
uoidcptOpenIDConnectProviderArn

-> UpdateOpenIDConnectProviderThumbprint

UpdateOpenIDConnectProviderThumbprint constructor.

The fields accessible through corresponding lenses are:

uoidcptOpenIDConnectProviderArn :: Text

uoidcptThumbprintList :: [Text]

Request lenses

uoidcptOpenIDConnectProviderArn :: Lens' UpdateOpenIDConnectProviderThumbprint Text

The Amazon Resource Name (ARN) of the IAM OpenID Connect (OIDC) provider to update the thumbprint for. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders action.

uoidcptThumbprintList :: Lens' UpdateOpenIDConnectProviderThumbprint [Text]

A list of certificate thumbprints that are associated with the specified IAM OpenID Connect provider. For more information, see CreateOpenIDConnectProvider .

Response

data UpdateOpenIDConnectProviderThumbprintResponse

Response constructor

updateOpenIDConnectProviderThumbprintResponse :: UpdateOpenIDConnectProviderThumbprintResponse

UpdateOpenIDConnectProviderThumbprintResponse constructor.