Network.AWS.IAM.UpdateOpenIdConnectProviderThumbprint

Description

Replaces the existing list of server certificate thumbprints associated with an OpenID Connect (OIDC) provider resource object with a new list of thumbprints.

The list that you pass with this action completely replaces the existing list of thumbprints. (The lists are not merged.)

Typically, you need to update a thumbprint only when the identity provider's certificate changes, which occurs rarely. However, if the provider's certificate does change, any attempt to assume an IAM role that specifies the OIDC provider as a principal fails until the certificate thumbprint is updated.

Creating a Request

updateOpenIdConnectProviderThumbprint

Arguments

:: Text
uoicptOpenIdConnectProviderARN

-> UpdateOpenIdConnectProviderThumbprint

Creates a value of UpdateOpenIdConnectProviderThumbprint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uoicptOpenIdConnectProviderARN - The Amazon Resource Name (ARN) of the IAM OIDC provider resource object for which you want to update the thumbprint. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders action. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

uoicptThumbprintList - A list of certificate thumbprints that are associated with the specified IAM OpenID Connect provider. For more information, see CreateOpenIDConnectProvider .

data UpdateOpenIdConnectProviderThumbprint

See: updateOpenIdConnectProviderThumbprint smart constructor.

Request Lenses

uoicptOpenIdConnectProviderARN :: Lens' UpdateOpenIdConnectProviderThumbprint Text

The Amazon Resource Name (ARN) of the IAM OIDC provider resource object for which you want to update the thumbprint. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders action. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

uoicptThumbprintList :: Lens' UpdateOpenIdConnectProviderThumbprint [Text]

A list of certificate thumbprints that are associated with the specified IAM OpenID Connect provider. For more information, see CreateOpenIDConnectProvider .

Destructuring the Response

updateOpenIdConnectProviderThumbprintResponse :: UpdateOpenIdConnectProviderThumbprintResponse

Creates a value of UpdateOpenIdConnectProviderThumbprintResponse with the minimum fields required to make a request.

data UpdateOpenIdConnectProviderThumbprintResponse

See: updateOpenIdConnectProviderThumbprintResponse smart constructor.