Network.AWS.IAM.UpdateRoleDescription

Description

Modifies the description of a role.

Creating a Request

updateRoleDescription

Arguments

:: Text
urdRoleName

-> Text
urdDescription

-> UpdateRoleDescription

Creates a value of UpdateRoleDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urdRoleName - The name of the role that you want to modify.

urdDescription - The new description that you want to apply to the specified role.

data UpdateRoleDescription

See: updateRoleDescription smart constructor.

Request Lenses

urdRoleName :: Lens' UpdateRoleDescription Text

The name of the role that you want to modify.

urdDescription :: Lens' UpdateRoleDescription Text

The new description that you want to apply to the specified role.

Destructuring the Response

updateRoleDescriptionResponse

Arguments

:: Int
urdrsResponseStatus

-> UpdateRoleDescriptionResponse

Creates a value of UpdateRoleDescriptionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urdrsRole - A structure that contains details about the modified role.

urdrsResponseStatus - -- | The response status code.

data UpdateRoleDescriptionResponse

See: updateRoleDescriptionResponse smart constructor.

Response Lenses

urdrsRole :: Lens' UpdateRoleDescriptionResponse (Maybe Role)

A structure that contains details about the modified role.

urdrsResponseStatus :: Lens' UpdateRoleDescriptionResponse Int

- | The response status code.