Network.AWS.IAM.UpdateSAMLProvider

Description

Updates the metadata document for an existing SAML provider.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_UpdateSAMLProvider.html

Request

data UpdateSAMLProvider

Request constructor

updateSAMLProvider

Arguments

:: Text
usamlpSAMLMetadataDocument

-> Text
usamlpSAMLProviderArn

-> UpdateSAMLProvider

UpdateSAMLProvider constructor.

The fields accessible through corresponding lenses are:

usamlpSAMLMetadataDocument :: Text

usamlpSAMLProviderArn :: Text

Request lenses

usamlpSAMLMetadataDocument :: Lens' UpdateSAMLProvider Text

An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

usamlpSAMLProviderArn :: Lens' UpdateSAMLProvider Text

The Amazon Resource Name (ARN) of the SAML provider to update.

Response

data UpdateSAMLProviderResponse

Response constructor

updateSAMLProviderResponse :: UpdateSAMLProviderResponse

UpdateSAMLProviderResponse constructor.

The fields accessible through corresponding lenses are:

usamlprSAMLProviderArn :: Maybe Text

Response lenses

usamlprSAMLProviderArn :: Lens' UpdateSAMLProviderResponse (Maybe Text)

The Amazon Resource Name (ARN) of the SAML provider that was updated.