Network.AWS.IAM.UpdateServerCertificate

Description

Updates the name and/or the path of the specified server certificate. You should understand the implications of changing a server certificate's path or name. For more information, see Managing Server Certificates in the Using IAM guide. To change a server certificate name the requester must have appropriate permissions on both the source object and the target object. For example, to change the name from ProductionCert to ProdCert, the entity making the request must have permission on ProductionCert and ProdCert, or must have permission on all (*). For more information about permissions, see Permissions and Policies. https:/iam.amazonaws.com ?Action=UpdateServerCertificate &ServerCertificateName=ProdServerCert &NewServerCertificateName=ProdServerCertName &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data UpdateServerCertificate

Request constructor

mkUpdateServerCertificate

Arguments

:: Text
uscServerCertificateName

-> UpdateServerCertificate

Smart constructor for the minimum required parameters to construct a valid UpdateServerCertificate request.

The fields accessible through corresponding lenses are:

ServerCertificateName :: Text

NewPath :: Maybe Text

NewServerCertificateName :: Maybe Text

Request lenses

uscServerCertificateName :: Lens' UpdateServerCertificate Text

The name of the server certificate that you want to update.

uscNewPath :: Lens' UpdateServerCertificate (Maybe Text)

The new path for the server certificate. Include this only if you are updating the server certificate's path.

uscNewServerCertificateName :: Lens' UpdateServerCertificate (Maybe Text)

The new name for the server certificate. Include this only if you are updating the server certificate's name.

Response

data UpdateServerCertificateResponse

Response constructor

mkUpdateServerCertificateResponse :: UpdateServerCertificateResponse

Smart constructor for the minimum required parameters to construct a valid UpdateServerCertificateResponse response.

This constructor is provided for convenience and testing purposes.