Network.AWS.IAM.UpdateServiceSpecificCredential

Description

Sets the status of a service-specific credential to Active or Inactive . Service-specific credentials that are inactive cannot be used for authentication to the service. This action can be used to disable a user’s service-specific credential as part of a credential rotation work flow.

Creating a Request

updateServiceSpecificCredential

Arguments

:: Text
usscServiceSpecificCredentialId

-> StatusType
usscStatus

-> UpdateServiceSpecificCredential

Creates a value of UpdateServiceSpecificCredential with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

usscUserName - The name of the IAM user associated with the service-specific credential. If you do not specify this value, then the operation assumes the user whose credentials are used to call the operation. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

usscServiceSpecificCredentialId - The unique identifier of the service-specific credential. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

usscStatus - The status to be assigned to the service-specific credential.

data UpdateServiceSpecificCredential

See: updateServiceSpecificCredential smart constructor.

Request Lenses

usscUserName :: Lens' UpdateServiceSpecificCredential (Maybe Text)

The name of the IAM user associated with the service-specific credential. If you do not specify this value, then the operation assumes the user whose credentials are used to call the operation. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

usscServiceSpecificCredentialId :: Lens' UpdateServiceSpecificCredential Text

The unique identifier of the service-specific credential. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

usscStatus :: Lens' UpdateServiceSpecificCredential StatusType

The status to be assigned to the service-specific credential.

Destructuring the Response

updateServiceSpecificCredentialResponse :: UpdateServiceSpecificCredentialResponse

Creates a value of UpdateServiceSpecificCredentialResponse with the minimum fields required to make a request.

data UpdateServiceSpecificCredentialResponse

See: updateServiceSpecificCredentialResponse smart constructor.