Network.AWS.IAM.UpdateSigningCertificate

Description

Changes the status of the specified signing certificate from active to disabled, or vice versa. This action can be used to disable a user's signing certificate as part of a certificate rotation work flow.

If the UserName field is not specified, the UserName is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

For information about rotating certificates, see Managing Keys andCertificates in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_UpdateSigningCertificate.html

Request

data UpdateSigningCertificate

Request constructor

updateSigningCertificate

Arguments

:: Text
uscCertificateId

-> StatusType
uscStatus

-> UpdateSigningCertificate

UpdateSigningCertificate constructor.

The fields accessible through corresponding lenses are:

uscCertificateId :: Text

uscStatus :: StatusType

uscUserName :: Maybe Text

Request lenses

uscCertificateId :: Lens' UpdateSigningCertificate Text

The ID of the signing certificate you want to update.

uscStatus :: Lens' UpdateSigningCertificate StatusType

The status you want to assign to the certificate. Active means the certificate can be used for API calls to AWS, while Inactive means the certificate cannot be used.

uscUserName :: Lens' UpdateSigningCertificate (Maybe Text)

The name of the user the signing certificate belongs to.

Response

data UpdateSigningCertificateResponse

Response constructor

updateSigningCertificateResponse :: UpdateSigningCertificateResponse

UpdateSigningCertificateResponse constructor.