Network.AWS.IAM.UploadSSHPublicKey

Description

Uploads an SSH public key and associates it with the specified IAM user.

The SSH public key uploaded by this action can be used only for authenticating the associated IAM user to an AWS CodeCommit repository. For more information about using SSH keys to authenticate to an AWS CodeCommit repository, see Set up AWS CodeCommit for SSH Connections in the AWS CodeCommit User Guide .

Creating a Request

uploadSSHPublicKey

Arguments

:: Text
usshpkUserName

-> Text
usshpkSSHPublicKeyBody

-> UploadSSHPublicKey

Creates a value of UploadSSHPublicKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

usshpkUserName - The name of the IAM user to associate the SSH public key with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

usshpkSSHPublicKeyBody - The SSH public key. The public key must be encoded in ssh-rsa format or PEM format. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

data UploadSSHPublicKey

See: uploadSSHPublicKey smart constructor.

Request Lenses

usshpkUserName :: Lens' UploadSSHPublicKey Text

The name of the IAM user to associate the SSH public key with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

usshpkSSHPublicKeyBody :: Lens' UploadSSHPublicKey Text

The SSH public key. The public key must be encoded in ssh-rsa format or PEM format. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

Destructuring the Response

uploadSSHPublicKeyResponse

Arguments

:: Int
uspkrsResponseStatus

-> UploadSSHPublicKeyResponse

Creates a value of UploadSSHPublicKeyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uspkrsSSHPublicKey - Contains information about the SSH public key.

uspkrsResponseStatus - -- | The response status code.

data UploadSSHPublicKeyResponse

Contains the response to a successful UploadSSHPublicKey request.

See: uploadSSHPublicKeyResponse smart constructor.

Response Lenses

uspkrsSSHPublicKey :: Lens' UploadSSHPublicKeyResponse (Maybe SSHPublicKey)

Contains information about the SSH public key.

uspkrsResponseStatus :: Lens' UploadSSHPublicKeyResponse Int

- | The response status code.