Network.AWS.IAM.V2010_05_08.CreateAccessKey

Description

Creates a new AWS secret access key and corresponding AWS access key ID for the specified user. The default status for new keys is Active. If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request. Because this action works for access keys under the AWS account, you can use this API to manage root credentials even if the AWS account has no associated users. For information about limits on the number of keys you can create, see Limitations on IAM Entities in the Using IAM guide. To ensure the security of your AWS account, the secret access key is accessible only during key and user creation. You must save the key (for example, in a text file) if you want to be able to access it again. If a secret key is lost, you can delete the access keys for the associated user and then create new keys. https:/iam.amazonaws.com ?Action=CreateAccessKey &UserName=Bob &Version=2010-05-08 &AUTHPARAMS Bob AKIAIOSFODNN7EXAMPLE Active wJalrXUtnFEMIK7MDENGbPxRfiCYzEXAMPLEKEY 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

createAccessKey :: CreateAccessKey

Minimum specification for a CreateAccessKey request.

data CreateAccessKey

Constructors

CreateAccessKey

Fields

_cakrUserName :: Maybe Text

The user name that the new key will belong to.

cakrUserName :: Iso' CreateAccessKey (Maybe Text)

data CreateAccessKeyResponse

Constructors

CreateAccessKeyResponse

Fields

_caksAccessKey :: AccessKey

Information about the access key.

caksAccessKey :: Iso' CreateAccessKeyResponse AccessKey