Network.AWS.IAM.V2010_05_08.CreateLoginProfile

Description

Creates a password for the specified user, giving the user the ability to access AWS services through the AWS Management Console. For more information about managing passwords, see Managing Passwords in the Using IAM guide. https:/iam.amazonaws.com ?Action=CreateLoginProfile &UserName=Bob &Password=Password1 &AUTHPARAMS Bob 2011-09-19T23:00:56Z 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

createLoginProfile

Arguments

:: Text
_clprPassword

-> Text
_clprUserName

-> CreateLoginProfile

Minimum specification for a CreateLoginProfile request.

data CreateLoginProfile

Constructors

CreateLoginProfile

Fields

_clprPassword :: Text

The new password for the user.

_clprUserName :: Text

Name of the user to create a password for.

_clprPasswordResetRequired :: Maybe Bool

Specifies whether the user is required to set a new password on next sign-in.

clprUserName :: Lens' CreateLoginProfile Text

clprPasswordResetRequired :: Lens' CreateLoginProfile (Maybe Bool)

clprPassword :: Lens' CreateLoginProfile Text

data CreateLoginProfileResponse

Constructors

CreateLoginProfileResponse

Fields

_clpsLoginProfile :: LoginProfile

The user name and password create date.

clpsLoginProfile :: Iso' CreateLoginProfileResponse LoginProfile