Network.AWS.IAM.V2010_05_08.CreateVirtualMFADevice

Description

Creates a new virtual MFA device for the AWS account. After creating the virtual MFA, use EnableMFADevice to attach the MFA device to an IAM user. For more information about creating and working with virtual MFA devices, go to Using a Virtual MFA Device in the Using IAM guide. For information about limits on the number of MFA devices you can create, see Limitations on Entities in the Using IAM guide. The seed information contained in the QR code and the Base32 string should be treated like any other secret access information, such as your AWS access keys or your passwords. After you provision your virtual device, you should ensure that the information is destroyed following secure procedures. https:/iam.amazonaws.com ?Action=CreateVirtualMFADevice &VirtualMFADeviceName=ExampleName &Path=/ &Version=2010-05-08 &AUTHPARAMS arn:aws:iam::123456789012:mfa/ExampleName 2K5K5XTLA7GGE75TQLYEXAMPLEEXAMPLEEXAMPLECHDFW4KJYZ6 UFQ75LL7COCYKM 89504E470D0A1A0AASDFAHSDFKJKLJFKALSDFJASDF 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data CreateVirtualMFADevice

Request constructor

mkCreateVirtualMFADevice

Arguments

:: Text
cvmfadVirtualMFADeviceName

-> CreateVirtualMFADevice

Smart constructor for the minimum required parameters to construct a valid CreateVirtualMFADevice request.

Request lenses

cvmfadPath :: Lens' CreateVirtualMFADevice (Maybe Text)

The path for the virtual MFA device. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide. This parameter is optional. If it is not included, it defaults to a slash (/).

cvmfadVirtualMFADeviceName :: Lens' CreateVirtualMFADevice Text

The name of the virtual MFA device. Use with path to uniquely identify a virtual MFA device.

Response

data CreateVirtualMFADeviceResponse

Contains the result of a successful invocation of the CreateVirtualMFADevice action.

Response lenses

cvmfadrsVirtualMFADevice :: Lens' CreateVirtualMFADeviceResponse VirtualMFADevice

A newly created virtual MFA device.