Network.AWS.IAM.V2010_05_08.DeleteGroupPolicy

Description

Deletes the specified policy that is associated with the specified group. https:/iam.amazonaws.com ?Action=DeleteGroupPolicy &GroupName=Admins &PolicyName=AdminRoot &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

data DeleteGroupPolicy

Constructors

DeleteGroupPolicy

Fields

_dgprGroupName :: Text

Name of the group the policy is associated with.

_dgprPolicyName :: Text

Name of the policy document to delete.

dgprPolicyName :: Lens' DeleteGroupPolicy Text

dgprGroupName :: Lens' DeleteGroupPolicy Text

data DeleteGroupPolicyResponse

Constructors

DeleteGroupPolicyResponse