Network.AWS.IAM.V2010_05_08.DeleteLoginProfile

Description

Deletes the password for the specified user, which terminates the user's ability to access AWS services through the AWS Management Console. Deleting a user's password does not prevent a user from accessing IAM through the command line interface or the API. To prevent all user access you must also either make the access key inactive or delete it. For more information about making keys inactive or deleting them, see UpdateAccessKey and DeleteAccessKey. https:/iam.amazonaws.com ?Action=DeleteLoginProfile &UserName=Bob &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

data DeleteLoginProfile

Constructors

DeleteLoginProfile

Fields

_dlprUserName :: Text

Name of the user whose password you want to delete.

dlprUserName :: Iso' DeleteLoginProfile Text

data DeleteLoginProfileResponse

Constructors

DeleteLoginProfileResponse