Network.AWS.IAM.V2010_05_08.DeleteRole

Description

Deletes the specified role. The role must not have any policies attached. For more information about roles, go to Working with Roles. Make sure you do not have any Amazon EC2 instances running with the role you are about to delete. Deleting a role or instance profile that is associated with a running instance will break any applications running on the instance. https:/iam.amazonaws.com ?Action=DeleteRole &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS 913e3f37-99ed-11e1-a4c3-270EXAMPLE04.

Request

data DeleteRole

Request constructor

mkDeleteRole

Arguments

:: Text
drRoleName

-> DeleteRole

Smart constructor for the minimum required parameters to construct a valid DeleteRole request.

Request lenses

drRoleName :: Lens' DeleteRole Text

Name of the role to delete.

Response

data DeleteRoleResponse

Response constructor

mkDeleteRoleResponse :: DeleteRoleResponse

Smart constructor for the minimum required parameters to construct a valid DeleteRoleResponse response.

This constructor is provided for convenience and testing purposes.