Network.AWS.IAM.V2010_05_08.DeleteRolePolicy

Description

Deletes the specified policy associated with the specified role. https:/iam.amazonaws.com ?Action=DeleteRolePolicy &PolicyName=S3AccessPolicy &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS c749ee7f-99ef-11e1-a4c3-27EXAMPLE804.

Request

data DeleteRolePolicy

Request constructor

mkDeleteRolePolicy

Arguments

:: Text
drpRoleName

-> Text
drpPolicyName

-> DeleteRolePolicy

Smart constructor for the minimum required parameters to construct a valid DeleteRolePolicy request.

Request lenses

drpRoleName :: Lens' DeleteRolePolicy Text

Name of the role the associated with the policy.

drpPolicyName :: Lens' DeleteRolePolicy Text

Name of the policy document to delete.

Response

data DeleteRolePolicyResponse

Response constructor

mkDeleteRolePolicyResponse :: DeleteRolePolicyResponse

Smart constructor for the minimum required parameters to construct a valid DeleteRolePolicyResponse response.

This constructor is provided for convenience and testing purposes.