Network.AWS.IAM.V2010_05_08.DeleteSigningCertificate

Description

Deletes the specified signing certificate associated with the specified user. If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request. Because this action works for access keys under the AWS account, you can use this API to manage root credentials even if the AWS account has no associated users. https:/iam.amazonaws.com ?Action=DeleteSigningCertificate &UserName=Bob &CertificateId=TA7SMP42TDN5Z26OBPJE7EXAMPLE &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

deleteSigningCertificate

Arguments

:: Text
_dscrCertificateId

-> DeleteSigningCertificate

Minimum specification for a DeleteSigningCertificate request.

data DeleteSigningCertificate

Constructors

DeleteSigningCertificate

Fields

_dscrCertificateId :: Text

ID of the signing certificate to delete.

_dscrUserName :: Maybe Text

Name of the user the signing certificate belongs to.

dscrUserName :: Lens' DeleteSigningCertificate (Maybe Text)

dscrCertificateId :: Lens' DeleteSigningCertificate Text

data DeleteSigningCertificateResponse

Constructors

DeleteSigningCertificateResponse