Network.AWS.IAM.V2010_05_08.DeleteUser

Description

Deletes the specified user. The user must not belong to any groups, have any keys or signing certificates, or have any attached policies. https:/iam.amazonaws.com ?Action=DeleteUser &UserName=Bob &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

data DeleteUser

Constructors

DeleteUser

Fields

_durUserName :: Text

Name of the user to delete.

durUserName :: Iso' DeleteUser Text

data DeleteUserResponse

Constructors

DeleteUserResponse