Network.AWS.IAM.V2010_05_08.EnableMFADevice

Description

Enables the specified MFA device and associates it with the specified user name. When enabled, the MFA device is required for every subsequent login by the user name associated with the device. https:/iam.amazonaws.com ?Action=EnableMFADevice &UserName=Bob &SerialNumber=R1234 &AuthenticationCode1=234567 &AuthenticationCode2=987654 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

data EnableMFADevice

Constructors

EnableMFADevice

Fields

_emfadrAuthenticationCode1 :: Text

An authentication code emitted by the device.

_emfadrAuthenticationCode2 :: Text

A subsequent authentication code emitted by the device.

_emfadrUserName :: Text

Name of the user for whom you want to enable the MFA device.

_emfadrSerialNumber :: Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

emfadrUserName :: Lens' EnableMFADevice Text

emfadrSerialNumber :: Lens' EnableMFADevice Text

emfadrAuthenticationCode2 :: Lens' EnableMFADevice Text

emfadrAuthenticationCode1 :: Lens' EnableMFADevice Text

data EnableMFADeviceResponse

Constructors

EnableMFADeviceResponse