Network.AWS.IAM.V2010_05_08.GetGroupPolicy

Description

Retrieves the specified policy document for the specified group. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcs/rfc3986.html. https:/iam.amazonaws.com ?Action=GetGroupPolicy &GroupName=Admins &PolicyName=AdminRoot &AUTHPARAMS Admins AdminRoot {Version:"2012-10-17",Statement:[{Effect:Allow,Action:"*",Resource:"*"}]} 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

data GetGroupPolicy

Constructors

GetGroupPolicy

Fields

_ggprGroupName :: Text

Name of the group the policy is associated with.

_ggprPolicyName :: Text

Name of the policy document to get.

ggprPolicyName :: Lens' GetGroupPolicy Text

ggprGroupName :: Lens' GetGroupPolicy Text

data GetGroupPolicyResponse

Constructors

GetGroupPolicyResponse

Fields

_ggpsGroupName :: Text

The group the policy is associated with.

_ggpsPolicyDocument :: Text

The policy document.

_ggpsPolicyName :: Text

The name of the policy.

ggpsPolicyName :: Lens' GetGroupPolicyResponse Text

ggpsPolicyDocument :: Lens' GetGroupPolicyResponse Text

ggpsGroupName :: Lens' GetGroupPolicyResponse Text