Network.AWS.IAM.V2010_05_08.GetLoginProfile

Description

Retrieves the user name and password-creation date for the specified user. If the user has not been assigned a password, the action returns a 404 (NoSuchEntity) error. https:/iam.amazonaws.com ?Action=GetLoginProfile &UserName=Bob &AUTHPARAMS Bob 2011-09-19T23:00:56Z 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data GetLoginProfile

Request constructor

mkGetLoginProfile

Arguments

:: Text
glpUserName

-> GetLoginProfile

Smart constructor for the minimum required parameters to construct a valid GetLoginProfile request.

Request lenses

glpUserName :: Lens' GetLoginProfile Text

Name of the user whose login profile you want to retrieve.

Response

data GetLoginProfileResponse

Contains the result of a successful invocation of the GetLoginProfile action.

Response constructor

mkGetLoginProfileResponse

Arguments

:: LoginProfile
glprLoginProfile

-> GetLoginProfileResponse

Smart constructor for the minimum required parameters to construct a valid GetLoginProfileResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

glprLoginProfile :: Lens' GetLoginProfileResponse LoginProfile

User name and password create date for the user.