Network.AWS.IAM.V2010_05_08.GetServerCertificate

Description

Retrieves information about the specified server certificate. https:/iam.amazonaws.com ?Action=GetServerCertificate &ServerCertificateName=ProdServerCert &Version=2010-05-08 &AUTHPARAMS ProdServerCert companyservercerts/ arn:aws:iam::123456789012:server-certificatecompanyservercerts/ProdServerCert 2010-05-08T01:02:03.004Z ASCACKCEVSQ6C2EXAMPLE 2012-05-08T01:02:03.004Z -----BEGIN CERTIFICATE----- MIICdzCCAeCgAwIBAgIGANc+Ha2wMA0GCSqGSIb3DQEBBQUAMFMxCzAJBgNVBAYT AlVTMRMwEQYDVQQKEwpBbWF6b24uY29tMQwwCgYDVQQLEwNBV1MxITAfBgNVBAMT GEFXUyBMaW1pdGVkLUFzc3VyYW5jZSBDQTAeFw0wOTAyMDQxNzE5MjdaFw0xMDAy MDQxNzE5MjdaMFIxCzAJBgNVBAYTAlVTMRMwEQYDVQQKEwpBbWF6b24uY29tMRcw FQYDVQQLEw5BV1MtRGV2ZWxvcGVyczEVMBMGA1UEAxMMNTdxNDl0c3ZwYjRtMIGf MA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQCpBvsOwmTO0td1RqzKjttSBaPjbr dqwNe9BrOyB08fw2+Ch5oonZYXfGUrT6mkYXH5fQot9HvASrzAKHO596FdJA6DmL ywdWe1Oggk7zFSXO1Xv+3vPrJtaYxYo3eRIp7w80PMkiOv6M0XK8ubcTouODeJbf suDqcLnLDxwsvwIDAQABo1cwVTAOBgNVHQ8BAf8EBAMCBaAwFgYDVR0lAQH/BAww CgYIKwYBBQUHAwIwDAYDVR0TAQH/BAIwADAdBgNVHQ4EFgQULGNaBphBumaKbDRK CAi0mH8B3mowDQYJKoZIhvcNAQEFBQADgYEAuKxhkXaCLGcqDuweKtO/AEw9ZePH wr0XqsaIK2HZboqruebXEGsojK4Ks0WzwgrEynuHJwTn760xe39rSqXWIOGrOBaX wFpWHVjTFMKk+tSDG1lssLHyYWWdFFU4AnejRGORJYNaRHgVTKjHphc5jEhHm0BX AEaHzTpmEXAMPLE= -----END CERTIFICATE----- 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data GetServerCertificate

Request constructor

mkGetServerCertificate

Arguments

:: Text
gscServerCertificateName

-> GetServerCertificate

Smart constructor for the minimum required parameters to construct a valid GetServerCertificate request.

The fields accessible through corresponding lenses are:

ServerCertificateName :: Text

Request lenses

gscServerCertificateName :: Lens' GetServerCertificate Text

The name of the server certificate you want to retrieve information about.

Response

data GetServerCertificateResponse

Contains the result of a successful invocation of the GetServerCertificate action.

Response constructor

mkGetServerCertificateResponse

Arguments

:: ServerCertificate
gscrServerCertificate

-> GetServerCertificateResponse

Smart constructor for the minimum required parameters to construct a valid GetServerCertificateResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ServerCertificate :: ServerCertificate

Response lenses

gscrServerCertificate :: Lens' GetServerCertificateResponse ServerCertificate

Information about the server certificate.