Network.AWS.IAM.V2010_05_08.GetUserPolicy

Description

Retrieves the specified policy document for the specified user. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcs/rfc3986.html. https:/iam.amazonaws.com ?Action=GetUserPolicy &UserName=Bob &PolicyName=AllAccessPolicy &AUTHPARAMS Bob AllAccessPolicy {Version:"2012-10-17",Statement:[{Effect:Allow,Action:"*",Resource:"*"}]} 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

data GetUserPolicy

Constructors

GetUserPolicy

Fields

_guprUserName :: Text

Name of the user who the policy is associated with.

_guprPolicyName :: Text

Name of the policy document to get.

guprUserName :: Lens' GetUserPolicy Text

guprPolicyName :: Lens' GetUserPolicy Text

data GetUserPolicyResponse

Constructors

GetUserPolicyResponse

Fields

_gupsUserName :: Text

The user the policy is associated with.

_gupsPolicyDocument :: Text

The policy document.

_gupsPolicyName :: Text

The name of the policy.

gupsUserName :: Lens' GetUserPolicyResponse Text

gupsPolicyName :: Lens' GetUserPolicyResponse Text

gupsPolicyDocument :: Lens' GetUserPolicyResponse Text