Network.AWS.IAM.V2010_05_08.ListGroupPolicies

Description

Lists the names of the policies associated with the specified group. If there are none, the action returns an empty list. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListGroupPolicies &GroupName=Admins &AUTHPARAMS AdminRoot KeyPolicy false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

listGroupPolicies

Arguments

:: Text
_lgprGroupName

-> ListGroupPolicies

Minimum specification for a ListGroupPolicies request.

data ListGroupPolicies

Constructors

ListGroupPolicies

Fields

_lgprGroupName :: Text

The name of the group to list policies for.

_lgprMarker :: Maybe Text

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

_lgprMaxItems :: Maybe Integer

Use this only when paginating results to indicate the maximum number of policy names you want in the response. If there are additional policy names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lgprMaxItems :: Lens' ListGroupPolicies (Maybe Integer)

lgprMarker :: Lens' ListGroupPolicies (Maybe Text)

lgprGroupName :: Lens' ListGroupPolicies Text

data ListGroupPoliciesResponse

Constructors

ListGroupPoliciesResponse

Fields

_lgpsIsTruncated :: Bool

A flag that indicates whether there are more policy names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policy names in the list.

_lgpsPolicyNames :: [Text]

A list of policy names.

_lgpsMarker :: Maybe Text

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lgpsPolicyNames :: Lens' ListGroupPoliciesResponse [Text]

lgpsMarker :: Lens' ListGroupPoliciesResponse (Maybe Text)

lgpsIsTruncated :: Lens' ListGroupPoliciesResponse Bool