Network.AWS.IAM.V2010_05_08.ListRolePolicies

Description

Lists the names of the policies associated with the specified role. If there are none, the action returns an empty list. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListRolePolicies &MaxItems=100 &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS CloudwatchPutMetricPolicy S3AccessPolicy false 8c7e1816-99f0-11e1-a4c3-27EXAMPLE804.

Documentation

listRolePolicies

Arguments

:: Text
_lrprRoleName

-> ListRolePolicies

Minimum specification for a ListRolePolicies request.

data ListRolePolicies

Constructors

ListRolePolicies

Fields

_lrprRoleName :: Text

The name of the role to list policies for.

_lrprMarker :: Maybe Text

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

_lrprMaxItems :: Maybe Integer

Use this parameter only when paginating results to indicate the maximum number of user names you want in the response. If there are additional user names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lrprRoleName :: Lens' ListRolePolicies Text

lrprMaxItems :: Lens' ListRolePolicies (Maybe Integer)

lrprMarker :: Lens' ListRolePolicies (Maybe Text)

data ListRolePoliciesResponse

Constructors

ListRolePoliciesResponse

Fields

_lrpsIsTruncated :: Bool

A flag that indicates whether there are more policy names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policy names in the list.

_lrpsPolicyNames :: [Text]

A list of policy names.

_lrpsMarker :: Maybe Text

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lrpsPolicyNames :: Lens' ListRolePoliciesResponse [Text]

lrpsMarker :: Lens' ListRolePoliciesResponse (Maybe Text)

lrpsIsTruncated :: Lens' ListRolePoliciesResponse Bool