Network.AWS.IAM.V2010_05_08.ListSAMLProviders

Description

Lists the SAML providers in the account. This operation requires Signature Version 4. https:/iam.amazonaws.com ?Action=ListSAMLProviders &MaxItems=100 &PathPrefix=application_abc &Version=2010-05-08 &AUTHPARAMS arn:aws:iam::123456789012:instance-profileapplication_abccomponent_xyz/Database 2032-05-09T16:27:11Z 2012-05-09T16:27:03Z arn:aws:iam::123456789012:instance-profileapplication_abccomponent_xyz/Webserver 2015-03-11T13:11:02Z 2012-05-09T16:27:11Z fd74fa8d-99f3-11e1-a4c3-27EXAMPLE804.

Request

data ListSAMLProviders

Request constructor

mkListSAMLProviders :: ListSAMLProviders

Smart constructor for the minimum required parameters to construct a valid ListSAMLProviders request.

Response

data ListSAMLProvidersResponse

Contains the result of a successful invocation of the ListSAMLProviders action.

Response constructor

mkListSAMLProvidersResponse :: ListSAMLProvidersResponse

Smart constructor for the minimum required parameters to construct a valid ListSAMLProvidersResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

lsamlprSAMLProviderList :: Lens' ListSAMLProvidersResponse [SAMLProviderListEntry]

The list of SAML providers for this account.