Network.AWS.IAM.V2010_05_08.ListServerCertificates

Description

Lists the server certificates that have the specified path prefix. If none exist, the action returns an empty list. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListServerCertificates &PathPrefix=companyservercerts &Version=2010-05-08 &AUTHPARAMS false ProdServerCert companyservercerts/ arn:aws:iam::123456789012:server-certificatecompanyservercerts/ProdServerCert 2010-05-08T01:02:03.004Z ASCACKCEVSQ6CEXAMPLE1 2012-05-08T01:02:03.004Z BetaServerCert companyservercerts/ arn:aws:iam::123456789012:server-certificatecompanyservercerts/BetaServerCert 2010-05-08T02:03:01.004Z ASCACKCEVSQ6CEXAMPLE2 2012-05-08T02:03:01.004Z TestServerCert companyservercerts/ arn:aws:iam::123456789012:server-certificatecompanyservercerts/TestServerCert 2010-05-08T03:01:02.004Z ASCACKCEVSQ6CEXAMPLE3 2012-05-08T03:01:02.004Z 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

listServerCertificates :: ListServerCertificates

Minimum specification for a ListServerCertificates request.

data ListServerCertificates

Constructors

ListServerCertificates

Fields

_lscrMarker :: Maybe Text

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

_lscrMaxItems :: Maybe Integer

Use this only when paginating results to indicate the maximum number of server certificates you want in the response. If there are additional server certificates beyond the maximum you specify, the IsTruncated response element will be set to true. This parameter is optional. If you do not include it, it defaults to 100.

_lscrPathPrefix :: Maybe Text

The path prefix for filtering the results. For example: companyservercerts would get all server certificates for which the path starts with companyservercerts. This parameter is optional. If it is not included, it defaults to a slash (/), listing all server certificates.

lscrPathPrefix :: Lens' ListServerCertificates (Maybe Text)

lscrMaxItems :: Lens' ListServerCertificates (Maybe Integer)

lscrMarker :: Lens' ListServerCertificates (Maybe Text)

data ListServerCertificatesResponse

Constructors

ListServerCertificatesResponse

Fields

_lscsIsTruncated :: Bool

A flag that indicates whether there are more server certificates to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more server certificates in the list.

_lscsServerCertificateMetadataList :: [ServerCertificateMetadata]

A list of server certificates.

_lscsMarker :: Maybe Text

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lscsServerCertificateMetadataList :: Lens' ListServerCertificatesResponse [ServerCertificateMetadata]

lscsMarker :: Lens' ListServerCertificatesResponse (Maybe Text)

lscsIsTruncated :: Lens' ListServerCertificatesResponse Bool