Network.AWS.IAM.V2010_05_08.ListSigningCertificates

Description

Returns information about the signing certificates associated with the specified user. If there are none, the action returns an empty list. Although each user is limited to a small number of signing certificates, you can still paginate the results using the MaxItems and Marker parameters. If the UserName field is not specified, the user name is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, this API can be used to manage root credentials even if the AWS account has no associated users. https:/iam.amazonaws.com ?Action=ListSigningCertificates &UserName=Bob &Version=2010-05-08 &AUTHPARAMS Bob Bob TA7SMP42TDN5Z26OBPJE7EXAMPLE -----BEGIN CERTIFICATE----- MIICdzCCAeCgAwIBAgIGANc+Ha2wMA0GCSqGSIb3DQEBBQUAMFMxCzAJBgNVBAYT AlVTMRMwEQYDVQQKEwpBbWF6b24uY29tMQwwCgYDVQQLEwNBV1MxITAfBgNVBAMT GEFXUyBMaW1pdGVkLUFzc3VyYW5jZSBDQTAeFw0wOTAyMDQxNzE5MjdaFw0xMDAy MDQxNzE5MjdaMFIxCzAJBgNVBAYTAlVTMRMwEQYDVQQKEwpBbWF6b24uY29tMRcw FQYDVQQLEw5BV1MtRGV2ZWxvcGVyczEVMBMGA1UEAxMMNTdxNDl0c3ZwYjRtMIGf MA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQCpBvsOwmTO0td1RqzKjttSBaPjbr dqwNe9BrOyB08fw2+Ch5oonZYXfGUrT6mkYXH5fQot9HvASrzAKHO596FdJA6DmL ywdWe1Oggk7zFSXO1Xv+3vPrJtaYxYo3eRIp7w80PMkiOv6M0XK8ubcTouODeJbf suDqcLnLDxwsvwIDAQABo1cwVTAOBgNVHQ8BAf8EBAMCBaAwFgYDVR0lAQH/BAww CgYIKwYBBQUHAwIwDAYDVR0TAQH/BAIwADAdBgNVHQ4EFgQULGNaBphBumaKbDRK CAi0mH8B3mowDQYJKoZIhvcNAQEFBQADgYEAuKxhkXaCLGcqDuweKtO/AEw9ZePH wr0XqsaIK2HZboqruebXEGsojK4Ks0WzwgrEynuHJwTn760xe39rSqXWIOGrOBaX wFpWHVjTFMKk+tSDG1lssLHyYWWdFFU4AnejRGORJYNaRHgVTKjHphc5jEhHm0BX AEaHzTpmEXAMPLE= -----END CERTIFICATE----- Active false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data ListSigningCertificates

Request constructor

mkListSigningCertificates :: ListSigningCertificates

Smart constructor for the minimum required parameters to construct a valid ListSigningCertificates request.

Request lenses

lsc1UserName :: Lens' ListSigningCertificates (Maybe Text)

The name of the user.

lsc1Marker :: Lens' ListSigningCertificates (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lsc1MaxItems :: Lens' ListSigningCertificates (Maybe Integer)

Use this only when paginating results to indicate the maximum number of certificate IDs you want in the response. If there are additional certificate IDs beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListSigningCertificatesResponse

Contains the result of a successful invocation of the ListSigningCertificates action.

Response lenses

lscrrCertificates :: Lens' ListSigningCertificatesResponse [SigningCertificate]

A list of the user's signing certificate information.

lscrrIsTruncated :: Lens' ListSigningCertificatesResponse Bool

A flag that indicates whether there are more certificate IDs to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more certificates in the list.

lscrrMarker :: Lens' ListSigningCertificatesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.