Network.AWS.IAM.V2010_05_08.ListUserPolicies

Description

Lists the names of the policies associated with the specified user. If there are none, the action returns an empty list. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListUserPolicies &UserName=Bob &AUTHPARAMS AllAccessPolicy KeyPolicy false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

listUserPolicies

Arguments

:: Text
_luprUserName

-> ListUserPolicies

Minimum specification for a ListUserPolicies request.

data ListUserPolicies

Constructors

ListUserPolicies

Fields

_luprUserName :: Text

The name of the user to list policies for.

_luprMarker :: Maybe Text

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

_luprMaxItems :: Maybe Integer

Use this only when paginating results to indicate the maximum number of policy names you want in the response. If there are additional policy names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

luprUserName :: Lens' ListUserPolicies Text

luprMaxItems :: Lens' ListUserPolicies (Maybe Integer)

luprMarker :: Lens' ListUserPolicies (Maybe Text)

data ListUserPoliciesResponse

Constructors

ListUserPoliciesResponse

Fields

_lupsIsTruncated :: Bool

A flag that indicates whether there are more policy names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policy names in the list.

_lupsPolicyNames :: [Text]

A list of policy names.

_lupsMarker :: Maybe Text

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lupsPolicyNames :: Lens' ListUserPoliciesResponse [Text]

lupsMarker :: Lens' ListUserPoliciesResponse (Maybe Text)

lupsIsTruncated :: Lens' ListUserPoliciesResponse Bool