Network.AWS.IAM.V2010_05_08.PutGroupPolicy

Description

Adds (or updates) a policy document associated with the specified group. For information about policies, refer to Overview of Policies in the Using IAM guide. For information about limits on the number of policies you can associate with a group, see Limitations on IAM Entities in the Using IAM guide. Because policy documents can be large, you should use POST rather than GET when calling PutGroupPolicy. For information about setting up signatures and authorization through the API, go to Signing AWS API Requests in the AWS General Reference. For general information about using the Query API with IAM, go to Making Query Requests in the Using IAM guide. https:/iam.amazonaws.com ?Action=PutGroupPolicy &GroupName=Admins &PolicyName=AdminRoot &PolicyDocument={Version:"2012-10-17",Statement:[{Effect:Allow,Action:"*",Resource:"*"}]} &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data PutGroupPolicy

Request constructor

mkPutGroupPolicy

Arguments

:: Text
pgpGroupName

-> Text
pgpPolicyName

-> Text
pgpPolicyDocument

-> PutGroupPolicy

Smart constructor for the minimum required parameters to construct a valid PutGroupPolicy request.

Request lenses

pgpGroupName :: Lens' PutGroupPolicy Text

Name of the group to associate the policy with.

pgpPolicyName :: Lens' PutGroupPolicy Text

Name of the policy document.

pgpPolicyDocument :: Lens' PutGroupPolicy Text

The policy document.

Response

data PutGroupPolicyResponse

Response constructor

mkPutGroupPolicyResponse :: PutGroupPolicyResponse

Smart constructor for the minimum required parameters to construct a valid PutGroupPolicyResponse response.

This constructor is provided for convenience and testing purposes.