Network.AWS.IAM.V2010_05_08.PutRolePolicy

Description

Adds (or updates) a policy document associated with the specified role. For information about policies, go to Overview of Policies in the Using IAM guide. For information about limits on the policies you can associate with a role, see Limitations on IAM Entities in the Using IAM guide. Because policy documents can be large, you should use POST rather than GET when calling PutRolePolicy. For information about setting up signatures and authorization through the API, go to Signing AWS API Requests in the AWS General Reference. For general information about using the Query API with IAM, go to Making Query Requests in the Using IAM guide. https:/iam.amazonaws.com ?Action=PutRolePolicy &RoleName=S3Access &PolicyName=S3AccessPolicy &PolicyDocument={Version:"2012-10-17",Statement:[{Effect:Allow,Action:"s3:*",Resource:"*"}]} &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

data PutRolePolicy

Constructors

PutRolePolicy

Fields

_prprPolicyDocument :: Text

The policy document.

_prprPolicyName :: Text

Name of the policy document.

_prprRoleName :: Text

Name of the role to associate the policy with.

prprRoleName :: Lens' PutRolePolicy Text

prprPolicyName :: Lens' PutRolePolicy Text

prprPolicyDocument :: Lens' PutRolePolicy Text

data PutRolePolicyResponse

Constructors

PutRolePolicyResponse