Network.AWS.IAM.V2010_05_08.PutUserPolicy

Description

Adds (or updates) a policy document associated with the specified user. For information about policies, refer to Overview of Policies in the Using IAM guide. For information about limits on the number of policies you can associate with a user, see Limitations on IAM Entities in the Using IAM guide. Because policy documents can be large, you should use POST rather than GET when calling PutUserPolicy. For information about setting up signatures and authorization through the API, go to Signing AWS API Requests in the AWS General Reference. For general information about using the Query API with IAM, go to Making Query Requests in the Using IAM guide. https:/iam.amazonaws.com ?Action=PutUserPolicy &UserName=Bob &PolicyName=AllAccessPolicy &PolicyDocument={Version:"2012-10-17",Statement:[{Effect:Allow,Action:"*",Resource:"*"}]} &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

data PutUserPolicy

Constructors

PutUserPolicy

Fields

_puprUserName :: Text

Name of the user to associate the policy with.

_puprPolicyDocument :: Text

The policy document.

_puprPolicyName :: Text

Name of the policy document.

puprUserName :: Lens' PutUserPolicy Text

puprPolicyName :: Lens' PutUserPolicy Text

puprPolicyDocument :: Lens' PutUserPolicy Text

data PutUserPolicyResponse

Constructors

PutUserPolicyResponse