Network.AWS.IAM.V2010_05_08.ResyncMFADevice

Description

Synchronizes the specified MFA device with AWS servers. https:/iam.amazonaws.com ?Action=ResyncMFADevice &UserName=Bob &SerialNumber=R1234 &AuthenticationCode1=234567 &AuthenticationCode2=987654 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data ResyncMFADevice

Request constructor

mkResyncMFADevice

Arguments

:: Text
rmfadUserName

-> Text
rmfadSerialNumber

-> Text
rmfadAuthenticationCode1

-> Text
rmfadAuthenticationCode2

-> ResyncMFADevice

Smart constructor for the minimum required parameters to construct a valid ResyncMFADevice request.

The fields accessible through corresponding lenses are:

UserName :: Text

SerialNumber :: Text

AuthenticationCode1 :: Text

AuthenticationCode2 :: Text

Request lenses

rmfadUserName :: Lens' ResyncMFADevice Text

Name of the user whose MFA device you want to resynchronize.

rmfadSerialNumber :: Lens' ResyncMFADevice Text

Serial number that uniquely identifies the MFA device.

rmfadAuthenticationCode1 :: Lens' ResyncMFADevice Text

An authentication code emitted by the device.

rmfadAuthenticationCode2 :: Lens' ResyncMFADevice Text

A subsequent authentication code emitted by the device.

Response

data ResyncMFADeviceResponse

Response constructor

mkResyncMFADeviceResponse :: ResyncMFADeviceResponse

Smart constructor for the minimum required parameters to construct a valid ResyncMFADeviceResponse response.

This constructor is provided for convenience and testing purposes.