Network.AWS.IAM.V2010_05_08.Types

Description

AWS Identity and Access Management (IAM) enables you to securely control access to AWS services and resources for your users. Using IAM, you can create and manage AWS users and groups and use permissions to allow and deny their access to AWS resources.

Documentation

data IAM

Supported version (2010-05-08) of the AWS Identity and Access Management service.

xmlOptions :: Tagged a XMLOptions

data AssignmentStatusType

The status (unassigned or assigned) of the devices to list. If you do not specify an AssignmentStatus, the action defaults to Any which lists both assigned and unassigned virtual MFA devices.

Constructors

AssignmentStatusTypeAny
Any

AssignmentStatusTypeAssigned
Assigned

AssignmentStatusTypeUnassigned
Unassigned

data ReportFormatType

The format (MIME type) of the credential report.

Constructors

ReportFormatTypeTextCsv
text/csv

data ReportStateType

Information about the state of a credential report.

Constructors

ReportStateTypeComplete
COMPLETE

ReportStateTypeInprogress
INPROGRESS

ReportStateTypeStarted
STARTED

data StatusType

The status of the access key. Active means the key is valid for API calls, while Inactive means it is not.

Constructors

StatusTypeActive
Active

StatusTypeInactive
Inactive

data SummaryKeyType

Constructors

SummaryKeyTypeAccessKeysPerUserQuota
AccessKeysPerUserQuota

SummaryKeyTypeAccountMFAEnabled
AccountMFAEnabled

SummaryKeyTypeGroupPolicySizeQuota
GroupPolicySizeQuota

SummaryKeyTypeGroups
Groups

SummaryKeyTypeGroupsPerUserQuota
GroupsPerUserQuota

SummaryKeyTypeGroupsQuota
GroupsQuota

SummaryKeyTypeMFADevices
MFADevices

SummaryKeyTypeMFADevicesInUse
MFADevicesInUse

SummaryKeyTypeServerCertificates
ServerCertificates

SummaryKeyTypeServerCertificatesQuota
ServerCertificatesQuota

SummaryKeyTypeSigningCertificatesPerUserQuota
SigningCertificatesPerUserQuota

SummaryKeyTypeUserPolicySizeQuota
UserPolicySizeQuota

SummaryKeyTypeUsers
Users

SummaryKeyTypeUsersQuota
UsersQuota

data AccessKey

Information about the access key.

Constructors

AccessKey

Fields

_akStatus :: StatusType

The status of the access key. Active means the key is valid for API calls, while Inactive means it is not.

_akSecretAccessKey :: Text

The secret key used to sign requests.

_akCreateDate :: Maybe ISO8601

The date when the access key was created.

_akUserName :: Text

Name of the user the key is associated with.

_akAccessKeyId :: Text

The ID for this access key.

data AccessKeyMetadata

The AccessKey data type contains information about an AWS access key, without its secret key. This data type is used as a response element in the action ListAccessKeys.

Constructors

AccessKeyMetadata

Fields

_akmStatus :: Maybe StatusType

The status of the access key. Active means the key is valid for API calls, while Inactive means it is not.

_akmCreateDate :: Maybe ISO8601

The date when the access key was created.

_akmUserName :: Maybe Text

Name of the user the key is associated with.

_akmAccessKeyId :: Maybe Text

The ID for this access key.

data Group

The Group data type contains information about a group. This data type is used as a response element in the following actions: CreateGroup GetGroup ListGroups.

Constructors

Group

Fields

_gpArn :: Text

The Amazon Resource Name (ARN) specifying the group. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

_gpPath :: Text

Path to the group. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

_gpCreateDate :: ISO8601

The date when the group was created.

_gpGroupId :: Text

The stable and unique string identifying the group. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

_gpGroupName :: Text

The name that identifies the group.

data InstanceProfile

The InstanceProfile data type contains information about an instance profile. This data type is used as a response element in the following actions: CreateInstanceProfile GetInstanceProfile ListInstanceProfiles ListInstanceProfilesForRole.

Constructors

InstanceProfile

Fields

_ipRoles :: [Role]

The role associated with the instance profile.

_ipArn :: Text

The Amazon Resource Name (ARN) specifying the instance profile. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

_ipPath :: Text

Path to the instance profile. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

_ipCreateDate :: ISO8601

The date when the instance profile was created.

_ipInstanceProfileId :: Text

The stable and unique string identifying the instance profile. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

_ipInstanceProfileName :: Text

The name identifying the instance profile.

data LoginProfile

User name and password create date for the user.

Constructors

LoginProfile

Fields

_lpCreateDate :: ISO8601

The date when the password for the user was created.

_lpUserName :: Text

The name of the user, which can be used for signing in to the AWS Management Console.

_lpPasswordResetRequired :: Maybe Bool

Specifies whether the user is required to set a new password on next sign-in.

data MFADevice

The MFADevice data type contains information about an MFA device. This data type is used as a response element in the action ListMFADevices.

Constructors

MFADevice

Fields

_mfadUserName :: Text

The user with whom the MFA device is associated.

_mfadEnableDate :: ISO8601

The date when the MFA device was enabled for the user.

_mfadSerialNumber :: Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

data PasswordPolicy

The PasswordPolicy data type contains information about the account password policy. This data type is used as a response element in the action GetAccountPasswordPolicy.

Constructors

PasswordPolicy

Fields

_ppExpirePasswords :: Maybe Bool

Specifies whether IAM users are required to change their password after a specified number of days.

_ppMinimumPasswordLength :: Maybe Integer

Minimum length to require for IAM user passwords.

_ppRequireNumbers :: Maybe Bool

Specifies whether to require numbers for IAM user passwords.

_ppPasswordReusePrevention :: Maybe Integer

Specifies the number of previous passwords that IAM users are prevented from reusing.

_ppRequireLowercaseCharacters :: Maybe Bool

Specifies whether to require lowercase characters for IAM user passwords.

_ppMaxPasswordAge :: Maybe Integer

The number of days that an IAM user password is valid.

_ppHardExpiry :: Maybe Bool

Specifies whether IAM users are prevented from setting a new password after their password has expired.

_ppRequireSymbols :: Maybe Bool

Specifies whether to require symbols for IAM user passwords.

_ppRequireUppercaseCharacters :: Maybe Bool

Specifies whether to require uppercase characters for IAM user passwords.

_ppAllowUsersToChangePassword :: Maybe Bool

Specifies whether IAM users are allowed to change their own password.

data Role

The Role data type contains information about a role. This data type is used as a response element in the following actions: CreateRole GetRole ListRoles.

Constructors

Role

Fields

_rAssumeRolePolicyDocument :: Maybe Text

The policy that grants an entity permission to assume the role. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcs/rfc3986.html.

_rArn :: Text

The Amazon Resource Name (ARN) specifying the role. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

_rPath :: Text

Path to the role. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

_rCreateDate :: ISO8601

The date when the role was created.

_rRoleName :: Text

The name identifying the role.

_rRoleId :: Text

The stable and unique string identifying the role. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

data SAMLProviderListEntry

The list of SAML providers for this account.

Constructors

SAMLProviderListEntry

Fields

_samlpleArn :: Maybe Text

The Amazon Resource Name (ARN) of the SAML provider.

_samlpleCreateDate :: Maybe ISO8601

The date and time when the SAML provider was created.

_samlpleValidUntil :: Maybe ISO8601

The expiration date and time for the SAML provider.

data ServerCertificate

Information about the server certificate.

Constructors

ServerCertificate

Fields

_seServerCertificateMetadata :: ServerCertificateMetadata

The meta information of the server certificate, such as its name, path, ID, and ARN.

_seCertificateBody :: Text

The contents of the public key certificate.

_seCertificateChain :: Maybe Text

The contents of the public key certificate chain.

data ServerCertificateMetadata

ServerCertificateMetadata contains information about a server certificate without its certificate body, certificate chain, and private key. This data type is used as a response element in the action UploadServerCertificate and ListServerCertificates.

Constructors

ServerCertificateMetadata

Fields

_scmServerCertificateName :: Text

The name that identifies the server certificate.

_scmUploadDate :: Maybe ISO8601

The date when the server certificate was uploaded.

_scmServerCertificateId :: Text

The stable and unique string identifying the server certificate. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

_scmArn :: Text

The Amazon Resource Name (ARN) specifying the server certificate. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

_scmPath :: Text

Path to the server certificate. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

_scmExpiration :: Maybe ISO8601

The date on which the certificate is set to expire.

data SigningCertificate

The SigningCertificate data type contains information about an X.509 signing certificate. This data type is used as a response element in the actions UploadSigningCertificate and ListSigningCertificates.

Constructors

SigningCertificate

Fields

_scStatus :: StatusType

The status of the signing certificate. Active means the key is valid for API calls, while Inactive means it is not.

_scUploadDate :: Maybe ISO8601

The date when the signing certificate was uploaded.

_scCertificateId :: Text

The ID for the signing certificate.

_scUserName :: Text

Name of the user the signing certificate is associated with.

_scCertificateBody :: Text

The contents of the signing certificate.

data User

The User data type contains information about a user. This data type is used as a response element in the following actions: CreateUser GetUser ListUsers.

Constructors

User

Fields

_uArn :: Text

The Amazon Resource Name (ARN) specifying the user. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

_uPath :: Text

Path to the user. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

_uCreateDate :: ISO8601

The date when the user was created.

_uUserName :: Text

The name identifying the user.

_uUserId :: Text

The stable and unique string identifying the user. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

data VirtualMFADevice

A newly created virtual MFA device.

Constructors

VirtualMFADevice

Fields

_vmfadQRCodePNG :: Maybe ByteString

A QR code PNG image that encodes otpauth:/totp$virtualMFADeviceName@$AccountName? secret=$Base32String where $virtualMFADeviceName is one of the create call arguments, AccountName is the user name if set (accountId otherwise), and Base32String is the seed in Base32 format. The Base32String is Base64-encoded.

_vmfadBase32StringSeed :: Maybe ByteString

The Base32 seed defined as specified in RFC3548. The Base32StringSeed is Base64-encoded.

_vmfadUser :: Maybe User

The User data type contains information about a user. This data type is used as a response element in the following actions: CreateUser GetUser ListUsers.

_vmfadEnableDate :: Maybe ISO8601

_vmfadSerialNumber :: Text

The serial number associated with VirtualMFADevice.

akUserName :: Lens' AccessKey Text

akStatus :: Lens' AccessKey StatusType

akSecretAccessKey :: Lens' AccessKey Text

akCreateDate :: Lens' AccessKey (Maybe ISO8601)

akAccessKeyId :: Lens' AccessKey Text

akmUserName :: Lens' AccessKeyMetadata (Maybe Text)

akmStatus :: Lens' AccessKeyMetadata (Maybe StatusType)

akmCreateDate :: Lens' AccessKeyMetadata (Maybe ISO8601)

akmAccessKeyId :: Lens' AccessKeyMetadata (Maybe Text)

gpPath :: Lens' Group Text

gpGroupName :: Lens' Group Text

gpGroupId :: Lens' Group Text

gpCreateDate :: Lens' Group ISO8601

gpArn :: Lens' Group Text

ipRoles :: Lens' InstanceProfile [Role]

ipPath :: Lens' InstanceProfile Text

ipInstanceProfileName :: Lens' InstanceProfile Text

ipInstanceProfileId :: Lens' InstanceProfile Text

ipCreateDate :: Lens' InstanceProfile ISO8601

ipArn :: Lens' InstanceProfile Text

lpUserName :: Lens' LoginProfile Text

lpPasswordResetRequired :: Lens' LoginProfile (Maybe Bool)

lpCreateDate :: Lens' LoginProfile ISO8601

mfadUserName :: Lens' MFADevice Text

mfadSerialNumber :: Lens' MFADevice Text

mfadEnableDate :: Lens' MFADevice ISO8601

ppRequireUppercaseCharacters :: Lens' PasswordPolicy (Maybe Bool)

ppRequireSymbols :: Lens' PasswordPolicy (Maybe Bool)

ppRequireNumbers :: Lens' PasswordPolicy (Maybe Bool)

ppRequireLowercaseCharacters :: Lens' PasswordPolicy (Maybe Bool)

ppPasswordReusePrevention :: Lens' PasswordPolicy (Maybe Integer)

ppMinimumPasswordLength :: Lens' PasswordPolicy (Maybe Integer)

ppMaxPasswordAge :: Lens' PasswordPolicy (Maybe Integer)

ppHardExpiry :: Lens' PasswordPolicy (Maybe Bool)

ppExpirePasswords :: Lens' PasswordPolicy (Maybe Bool)

ppAllowUsersToChangePassword :: Lens' PasswordPolicy (Maybe Bool)

rRoleName :: Lens' Role Text

rRoleId :: Lens' Role Text

rPath :: Lens' Role Text

rCreateDate :: Lens' Role ISO8601

rAssumeRolePolicyDocument :: Lens' Role (Maybe Text)

rArn :: Lens' Role Text

samlpleValidUntil :: Lens' SAMLProviderListEntry (Maybe ISO8601)

samlpleCreateDate :: Lens' SAMLProviderListEntry (Maybe ISO8601)

samlpleArn :: Lens' SAMLProviderListEntry (Maybe Text)

seServerCertificateMetadata :: Lens' ServerCertificate ServerCertificateMetadata

seCertificateChain :: Lens' ServerCertificate (Maybe Text)

seCertificateBody :: Lens' ServerCertificate Text

scmUploadDate :: Lens' ServerCertificateMetadata (Maybe ISO8601)

scmServerCertificateName :: Lens' ServerCertificateMetadata Text

scmServerCertificateId :: Lens' ServerCertificateMetadata Text

scmPath :: Lens' ServerCertificateMetadata Text

scmExpiration :: Lens' ServerCertificateMetadata (Maybe ISO8601)

scmArn :: Lens' ServerCertificateMetadata Text

scUserName :: Lens' SigningCertificate Text

scUploadDate :: Lens' SigningCertificate (Maybe ISO8601)

scStatus :: Lens' SigningCertificate StatusType

scCertificateId :: Lens' SigningCertificate Text

scCertificateBody :: Lens' SigningCertificate Text

uUserName :: Lens' User Text

uUserId :: Lens' User Text

uPath :: Lens' User Text

uCreateDate :: Lens' User ISO8601

uArn :: Lens' User Text

vmfadUser :: Lens' VirtualMFADevice (Maybe User)

vmfadSerialNumber :: Lens' VirtualMFADevice Text

vmfadQRCodePNG :: Lens' VirtualMFADevice (Maybe ByteString)

vmfadEnableDate :: Lens' VirtualMFADevice (Maybe ISO8601)

vmfadBase32StringSeed :: Lens' VirtualMFADevice (Maybe ByteString)