Network.AWS.IAM.V2010_05_08.UpdateAccessKey

Description

Changes the status of the specified access key from Active to Inactive, or vice versa. This action can be used to disable a user's key as part of a key rotation work flow. If the UserName field is not specified, the UserName is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, this API can be used to manage root credentials even if the AWS account has no associated users. For information about rotating keys, see Managing Keys and Certificates in the Using IAM guide. https:/iam.amazonaws.com ?Action=UpdateAccessKey &UserName=Bob &AccessKeyId=AKIAIOSFODNN7EXAMPLE &Status=Inactive &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data UpdateAccessKey

Request constructor

mkUpdateAccessKey

Arguments

:: Text
uakAccessKeyId

-> StatusType
uakStatus

-> UpdateAccessKey

Smart constructor for the minimum required parameters to construct a valid UpdateAccessKey request.

Request lenses

uakUserName :: Lens' UpdateAccessKey (Maybe Text)

Name of the user whose key you want to update.

uakAccessKeyId :: Lens' UpdateAccessKey Text

The access key ID of the secret access key you want to update.

uakStatus :: Lens' UpdateAccessKey StatusType

The status you want to assign to the secret access key. Active means the key can be used for API calls to AWS, while Inactive means the key cannot be used.

Response

data UpdateAccessKeyResponse