Network.AWS.IAM.V2010_05_08.UpdateAccountPasswordPolicy

Description

Updates the password policy settings for the account. For more information about using a password policy, see Managing an IAM Password Policy in the Using IAM guide. https:/iam.amazonaws.com ?Action=UpdateAccountPasswordPolicy &MinimumPasswordLength=12 &RequireSymbols=false &RequireNumbers=true &RequireUppercaseCharacters=true &RequireLowercaseCharacters=true &AllowUsersToChangePassword=true &MaxPasswordAge=90 &PasswordReusePrevention=6 &HardExpiry=false &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

updateAccountPasswordPolicy :: UpdateAccountPasswordPolicy

Minimum specification for a UpdateAccountPasswordPolicy request.

data UpdateAccountPasswordPolicy

Constructors

UpdateAccountPasswordPolicy

Fields

_uapprHardExpiry :: Maybe Bool

Prevents IAM users from setting a new password after their password has expired.

_uapprRequireNumbers :: Maybe Bool

Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).

_uapprRequireLowercaseCharacters :: Maybe Bool

Specifies whether IAM user passwords must contain at least one lowercase character from the ISO basic Latin alphabet (a to z).

_uapprRequireSymbols :: Maybe Bool

Specifies whether IAM user passwords must contain at least one of the following non-alphanumeric characters: ! @ # $ % ^ &amp; * ( ) _ + - = [ ] { } | '.

_uapprRequireUppercaseCharacters :: Maybe Bool

Specifies whether IAM user passwords must contain at least one uppercase character from the ISO basic Latin alphabet (A to Z).

_uapprAllowUsersToChangePassword :: Maybe Bool

Allows all IAM users in your account to use the AWS Management Console to change their own passwords. For more information, see Letting IAM Users Change Their Own Passwords in the Using IAM guide.

_uapprMaxPasswordAge :: Maybe Integer

The number of days that an IAM user password is valid.

_uapprMinimumPasswordLength :: Maybe Integer

The minimum number of characters allowed in an IAM user password.

_uapprPasswordReusePrevention :: Maybe Integer

Specifies the number of previous passwords that IAM users are prevented from reusing.

uapprRequireUppercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

uapprRequireSymbols :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

uapprRequireNumbers :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

uapprRequireLowercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

uapprPasswordReusePrevention :: Lens' UpdateAccountPasswordPolicy (Maybe Integer)

uapprMinimumPasswordLength :: Lens' UpdateAccountPasswordPolicy (Maybe Integer)

uapprMaxPasswordAge :: Lens' UpdateAccountPasswordPolicy (Maybe Integer)

uapprHardExpiry :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

uapprAllowUsersToChangePassword :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

data UpdateAccountPasswordPolicyResponse

Constructors

UpdateAccountPasswordPolicyResponse