Network.AWS.IAM.V2010_05_08.UpdateLoginProfile

Description

Changes the password for the specified user. https:/iam.amazonaws.com ?Action=UpdateLoginProfile &UserName=Bob &Password=NewPassword &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data UpdateLoginProfile

Request constructor

mkUpdateLoginProfile

Arguments

:: Text
ulpUserName

-> UpdateLoginProfile

Smart constructor for the minimum required parameters to construct a valid UpdateLoginProfile request.

Request lenses

ulpUserName :: Lens' UpdateLoginProfile Text

Name of the user whose password you want to update.

ulpPassword :: Lens' UpdateLoginProfile (Maybe Text)

The new password for the specified user.

ulpPasswordResetRequired :: Lens' UpdateLoginProfile (Maybe Bool)

Require the specified user to set a new password on next sign-in.

Response

data UpdateLoginProfileResponse