Network.AWS.IAM.V2010_05_08.UpdateSigningCertificate

Description

Changes the status of the specified signing certificate from active to disabled, or vice versa. This action can be used to disable a user's signing certificate as part of a certificate rotation work flow. If the UserName field is not specified, the UserName is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, this API can be used to manage root credentials even if the AWS account has no associated users. For information about rotating certificates, see Managing Keys and Certificates in the Using IAM guide. https:/iam.amazonaws.com ?Action=UpdateSigningCertificate &UserName=Bob &CertificateId=TA7SMP42TDN5Z26OBPJE7EXAMPLE &Status=Inactive &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data UpdateSigningCertificate

Request constructor

mkUpdateSigningCertificate

Arguments

:: Text
usc1CertificateId

-> StatusType
usc1Status

-> UpdateSigningCertificate

Smart constructor for the minimum required parameters to construct a valid UpdateSigningCertificate request.

The fields accessible through corresponding lenses are:

UserName :: Maybe Text

CertificateId :: Text

Status :: StatusType

Request lenses

usc1UserName :: Lens' UpdateSigningCertificate (Maybe Text)

Name of the user the signing certificate belongs to.

usc1CertificateId :: Lens' UpdateSigningCertificate Text

The ID of the signing certificate you want to update.

usc1Status :: Lens' UpdateSigningCertificate StatusType

The status you want to assign to the certificate. Active means the certificate can be used for API calls to AWS, while Inactive means the certificate cannot be used.

Response

data UpdateSigningCertificateResponse

Response constructor

mkUpdateSigningCertificateResponse :: UpdateSigningCertificateResponse

Smart constructor for the minimum required parameters to construct a valid UpdateSigningCertificateResponse response.

This constructor is provided for convenience and testing purposes.