Network.AWS.IAM.V2010_05_08.UpdateUser

Description

Updates the name and/or the path of the specified user. You should understand the implications of changing a user's path or name. For more information, see Renaming Users and Groups in the Using IAM guide. To change a user name the requester must have appropriate permissions on both the source object and the target object. For example, to change Bob to Robert, the entity making the request must have permission on Bob and Robert, or must have permission on all (*). For more information about permissions, see Permissions and Policies. https:/iam.amazonaws.com ?Action=UpdateUser &UserName=Bob &NewUserName=Robert &Version=2010-05-08 &AUTHPARAMS division_abcsubdivision_xyz/ Robert AIDACKCEVSQ6C2EXAMPLE arn:aws::123456789012:userdivision_abcsubdivision_xyz/Robert 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

updateUser

Arguments

:: Text
_uurUserName

-> UpdateUser

Minimum specification for a UpdateUser request.

data UpdateUser

Constructors

UpdateUser

Fields

_uurUserName :: Text

Name of the user to update. If you're changing the name of the user, this is the original user name.

_uurNewPath :: Maybe Text

New path for the user. Include this parameter only if you're changing the user's path.

_uurNewUserName :: Maybe Text

New name for the user. Include this parameter only if you're changing the user's name.

uurUserName :: Lens' UpdateUser Text

uurNewUserName :: Lens' UpdateUser (Maybe Text)

uurNewPath :: Lens' UpdateUser (Maybe Text)

data UpdateUserResponse

Constructors

UpdateUserResponse