Network.AWS.IAM.AttachGroupPolicy

Description

Attaches the specified managed policy to the specified IAM group.

You use this API to attach a managed policy to a group. To embed an inline policy in a group, use PutGroupPolicy .

For more information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

attachGroupPolicy

Arguments

:: Text
agpGroupName

-> Text
agpPolicyARN

-> AttachGroupPolicy

Creates a value of AttachGroupPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

agpGroupName - The name (friendly name, not ARN) of the group to attach the policy to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

agpPolicyARN - The Amazon Resource Name (ARN) of the IAM policy you want to attach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data AttachGroupPolicy

See: attachGroupPolicy smart constructor.

Request Lenses

agpGroupName :: Lens' AttachGroupPolicy Text

The name (friendly name, not ARN) of the group to attach the policy to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

agpPolicyARN :: Lens' AttachGroupPolicy Text

The Amazon Resource Name (ARN) of the IAM policy you want to attach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

attachGroupPolicyResponse :: AttachGroupPolicyResponse

Creates a value of AttachGroupPolicyResponse with the minimum fields required to make a request.

data AttachGroupPolicyResponse

See: attachGroupPolicyResponse smart constructor.