Network.AWS.IAM.AttachRolePolicy

Description

Attaches the specified managed policy to the specified IAM role. When you attach a managed policy to a role, the managed policy becomes part of the role's permission (access) policy.

Use this API to attach a managed policy to a role. To embed an inline policy in a role, use PutRolePolicy . For more information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

attachRolePolicy

Arguments

:: Text
arpRoleName

-> Text
arpPolicyARN

-> AttachRolePolicy

Creates a value of AttachRolePolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

arpRoleName - The name (friendly name, not ARN) of the role to attach the policy to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

arpPolicyARN - The Amazon Resource Name (ARN) of the IAM policy you want to attach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data AttachRolePolicy

See: attachRolePolicy smart constructor.

Request Lenses

arpRoleName :: Lens' AttachRolePolicy Text

The name (friendly name, not ARN) of the role to attach the policy to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

arpPolicyARN :: Lens' AttachRolePolicy Text

The Amazon Resource Name (ARN) of the IAM policy you want to attach. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

attachRolePolicyResponse :: AttachRolePolicyResponse

Creates a value of AttachRolePolicyResponse with the minimum fields required to make a request.

data AttachRolePolicyResponse

See: attachRolePolicyResponse smart constructor.