Network.AWS.IAM.ChangePassword

Description

Changes the password of the IAM user who is calling this action. The root account password is not affected by this action.

To change the password for a different user, see UpdateLoginProfile . For more information about modifying passwords, see Managing Passwords in the IAM User Guide .

Creating a Request

changePassword

Arguments

:: Text
cpOldPassword

-> Text
cpNewPassword

-> ChangePassword

Creates a value of ChangePassword with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpOldPassword - The IAM user's current password.

cpNewPassword - The new password. The new password must conform to the AWS account's password policy, if one exists. The regex pattern used to validate this parameter is a string of characters consisting of almost any printable ASCII character from the space (u0020) through the end of the ASCII character range (u00FF). You can also include the tab (u0009), line feed (u000A), and carriage return (u000D) characters. Although any of these characters are valid in a password, note that many tools, such as the AWS Management Console, might restrict the ability to enter certain characters because they have special meaning within that tool.

data ChangePassword

See: changePassword smart constructor.

Request Lenses

cpOldPassword :: Lens' ChangePassword Text

The IAM user's current password.

cpNewPassword :: Lens' ChangePassword Text

The new password. The new password must conform to the AWS account's password policy, if one exists. The regex pattern used to validate this parameter is a string of characters consisting of almost any printable ASCII character from the space (u0020) through the end of the ASCII character range (u00FF). You can also include the tab (u0009), line feed (u000A), and carriage return (u000D) characters. Although any of these characters are valid in a password, note that many tools, such as the AWS Management Console, might restrict the ability to enter certain characters because they have special meaning within that tool.

Destructuring the Response

changePasswordResponse :: ChangePasswordResponse

Creates a value of ChangePasswordResponse with the minimum fields required to make a request.

data ChangePasswordResponse

See: changePasswordResponse smart constructor.