Network.AWS.IAM.ChangePassword

Description

Changes the password of the IAM user who is calling this action. The root account password is not affected by this action.

To change the password for a different user, see UpdateLoginProfile. For more information about modifying passwords, see Managing Passwords in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ChangePassword.html

Request

data ChangePassword

Request constructor

changePassword

Arguments

:: Text
cpOldPassword

-> Text
cpNewPassword

-> ChangePassword

ChangePassword constructor.

The fields accessible through corresponding lenses are:

cpNewPassword :: Text

cpOldPassword :: Text

Request lenses

cpNewPassword :: Lens' ChangePassword Text

The new password. The new password must conform to the AWS account's password policy, if one exists.

cpOldPassword :: Lens' ChangePassword Text

The IAM user's current password.

Response

data ChangePasswordResponse

Response constructor

changePasswordResponse :: ChangePasswordResponse

ChangePasswordResponse constructor.