Network.AWS.IAM.CreateAccessKey

Description

Creates a new AWS secret access key and corresponding AWS access key ID for the specified user. The default status for new keys is Active .

If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

For information about limits on the number of keys you can create, see Limitations on IAM Entities in the IAM User Guide .

Important: To ensure the security of your AWS account, the secret access key is accessible only during key and user creation. You must save the key (for example, in a text file) if you want to be able to access it again. If a secret key is lost, you can delete the access keys for the associated user and then create new keys.

Creating a Request

createAccessKey :: CreateAccessKey

Creates a value of CreateAccessKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cakUserName - The name of the IAM user that the new key will belong to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data CreateAccessKey

See: createAccessKey smart constructor.

Request Lenses

cakUserName :: Lens' CreateAccessKey (Maybe Text)

The name of the IAM user that the new key will belong to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

createAccessKeyResponse

Arguments

:: Int
cakrsResponseStatus

-> AccessKeyInfo
cakrsAccessKey

-> CreateAccessKeyResponse

Creates a value of CreateAccessKeyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cakrsResponseStatus - -- | The response status code.

cakrsAccessKey - A structure with details about the access key.

data CreateAccessKeyResponse

Contains the response to a successful CreateAccessKey request.

See: createAccessKeyResponse smart constructor.

Response Lenses

cakrsResponseStatus :: Lens' CreateAccessKeyResponse Int

- | The response status code.

cakrsAccessKey :: Lens' CreateAccessKeyResponse AccessKeyInfo

A structure with details about the access key.