Network.AWS.IAM.CreateGroup

Description

Creates a new group.

For information about the number of groups you can create, see Limitations on IAM Entities in the IAM User Guide .

Creating a Request

createGroup

Arguments

:: Text
cgGroupName

-> CreateGroup

Creates a value of CreateGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgPath - The path to the group. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cgGroupName - The name of the group to create. Do not include the path in this value. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-. The group name must be unique within the account. Group names are not distinguished by case. For example, you cannot create groups named both ADMINS and "admins".

data CreateGroup

See: createGroup smart constructor.

Request Lenses

cgPath :: Lens' CreateGroup (Maybe Text)

The path to the group. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cgGroupName :: Lens' CreateGroup Text

The name of the group to create. Do not include the path in this value. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-. The group name must be unique within the account. Group names are not distinguished by case. For example, you cannot create groups named both ADMINS and "admins".

Destructuring the Response

createGroupResponse

Arguments

:: Int
cgrsResponseStatus

-> Group
cgrsGroup

-> CreateGroupResponse

Creates a value of CreateGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgrsResponseStatus - -- | The response status code.

cgrsGroup - A structure containing details about the new group.

data CreateGroupResponse

Contains the response to a successful CreateGroup request.

See: createGroupResponse smart constructor.

Response Lenses

cgrsResponseStatus :: Lens' CreateGroupResponse Int

- | The response status code.

cgrsGroup :: Lens' CreateGroupResponse Group

A structure containing details about the new group.