Network.AWS.IAM.CreateGroup

Description

Creates a new group.

For information about the number of groups you can create, see Limitationson IAM Entities in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateGroup.html

Request

data CreateGroup

Request constructor

createGroup

Arguments

:: Text
cgGroupName

-> CreateGroup

CreateGroup constructor.

The fields accessible through corresponding lenses are:

cgGroupName :: Text

cgPath :: Maybe Text

Request lenses

cgGroupName :: Lens' CreateGroup Text

The name of the group to create. Do not include the path in this value.

cgPath :: Lens' CreateGroup (Maybe Text)

The path to the group. For more information about paths, see IAM Identifiers in the Using IAM guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

Response

data CreateGroupResponse

Response constructor

createGroupResponse

Arguments

:: Group
cgrGroup

-> CreateGroupResponse

CreateGroupResponse constructor.

The fields accessible through corresponding lenses are:

cgrGroup :: Group

Response lenses

cgrGroup :: Lens' CreateGroupResponse Group

Information about the group.