Network.AWS.IAM.CreateInstanceProfile

Description

Creates a new instance profile. For information about instance profiles, go to About Instance Profiles .

For information about the number of instance profiles you can create, see Limitations on IAM Entities in the IAM User Guide .

Creating a Request

createInstanceProfile

Arguments

:: Text
cipInstanceProfileName

-> CreateInstanceProfile

Creates a value of CreateInstanceProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cipPath - The path to the instance profile. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cipInstanceProfileName - The name of the instance profile to create. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data CreateInstanceProfile

See: createInstanceProfile smart constructor.

Request Lenses

cipPath :: Lens' CreateInstanceProfile (Maybe Text)

The path to the instance profile. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cipInstanceProfileName :: Lens' CreateInstanceProfile Text

The name of the instance profile to create. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

createInstanceProfileResponse

Arguments

:: Int
ciprsResponseStatus

-> InstanceProfile
ciprsInstanceProfile

-> CreateInstanceProfileResponse

Creates a value of CreateInstanceProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciprsResponseStatus - -- | The response status code.

ciprsInstanceProfile - A structure containing details about the new instance profile.

data CreateInstanceProfileResponse

Contains the response to a successful CreateInstanceProfile request.

See: createInstanceProfileResponse smart constructor.

Response Lenses

ciprsResponseStatus :: Lens' CreateInstanceProfileResponse Int

- | The response status code.

ciprsInstanceProfile :: Lens' CreateInstanceProfileResponse InstanceProfile

A structure containing details about the new instance profile.