Network.AWS.IAM.CreateLoginProfile

Description

Creates a password for the specified user, giving the user the ability to access AWS services through the AWS Management Console. For more information about managing passwords, see Managing Passwords in the IAM User Guide .

Creating a Request

createLoginProfile

Arguments

:: Text
clpUserName

-> Text
clpPassword

-> CreateLoginProfile

Creates a value of CreateLoginProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clpPasswordResetRequired - Specifies whether the user is required to set a new password on next sign-in.

clpUserName - The name of the IAM user to create a password for. The user must already exist. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

clpPassword - The new password for the user. The regex pattern used to validate this parameter is a string of characters consisting of almost any printable ASCII character from the space (u0020) through the end of the ASCII character range (u00FF). You can also include the tab (u0009), line feed (u000A), and carriage return (u000D) characters. Although any of these characters are valid in a password, note that many tools, such as the AWS Management Console, might restrict the ability to enter certain characters because they have special meaning within that tool.

data CreateLoginProfile

See: createLoginProfile smart constructor.

Request Lenses

clpPasswordResetRequired :: Lens' CreateLoginProfile (Maybe Bool)

Specifies whether the user is required to set a new password on next sign-in.

clpUserName :: Lens' CreateLoginProfile Text

The name of the IAM user to create a password for. The user must already exist. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

clpPassword :: Lens' CreateLoginProfile Text

The new password for the user. The regex pattern used to validate this parameter is a string of characters consisting of almost any printable ASCII character from the space (u0020) through the end of the ASCII character range (u00FF). You can also include the tab (u0009), line feed (u000A), and carriage return (u000D) characters. Although any of these characters are valid in a password, note that many tools, such as the AWS Management Console, might restrict the ability to enter certain characters because they have special meaning within that tool.

Destructuring the Response

createLoginProfileResponse

Arguments

:: Int
clprsResponseStatus

-> LoginProfile
clprsLoginProfile

-> CreateLoginProfileResponse

Creates a value of CreateLoginProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clprsResponseStatus - -- | The response status code.

clprsLoginProfile - A structure containing the user name and password create date.

data CreateLoginProfileResponse

Contains the response to a successful CreateLoginProfile request.

See: createLoginProfileResponse smart constructor.

Response Lenses

clprsResponseStatus :: Lens' CreateLoginProfileResponse Int

- | The response status code.

clprsLoginProfile :: Lens' CreateLoginProfileResponse LoginProfile

A structure containing the user name and password create date.