Network.AWS.IAM.CreatePolicy

Description

Creates a new managed policy for your AWS account.

This operation creates a policy version with a version identifier of v1 and sets v1 as the policy's default version. For more information about policy versions, see Versioning for Managed Policies in the IAM User Guide .

For more information about managed policies in general, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

createPolicy

Arguments

:: Text
cpPolicyName

-> Text
cpPolicyDocument

-> CreatePolicy

Creates a value of CreatePolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpPath - The path for the policy. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cpDescription - A friendly description of the policy. Typically used to store information about the permissions defined in the policy. For example, "Grants access to production DynamoDB tables." The policy description is immutable. After a value is assigned, it cannot be changed.

cpPolicyName - The friendly name of the policy. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

cpPolicyDocument - The JSON policy document that you want to use as the content for the new policy. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

data CreatePolicy

See: createPolicy smart constructor.

Request Lenses

cpPath :: Lens' CreatePolicy (Maybe Text)

The path for the policy. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

cpDescription :: Lens' CreatePolicy (Maybe Text)

A friendly description of the policy. Typically used to store information about the permissions defined in the policy. For example, "Grants access to production DynamoDB tables." The policy description is immutable. After a value is assigned, it cannot be changed.

cpPolicyName :: Lens' CreatePolicy Text

The friendly name of the policy. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

cpPolicyDocument :: Lens' CreatePolicy Text

The JSON policy document that you want to use as the content for the new policy. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

Destructuring the Response

createPolicyResponse

Arguments

:: Int
cprsResponseStatus

-> CreatePolicyResponse

Creates a value of CreatePolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cprsPolicy - A structure containing details about the new policy.

cprsResponseStatus - -- | The response status code.

data CreatePolicyResponse

Contains the response to a successful CreatePolicy request.

See: createPolicyResponse smart constructor.

Response Lenses

cprsPolicy :: Lens' CreatePolicyResponse (Maybe Policy)

A structure containing details about the new policy.

cprsResponseStatus :: Lens' CreatePolicyResponse Int

- | The response status code.