Network.AWS.IAM.CreatePolicyVersion

Description

Creates a new version of the specified managed policy. To update a managed policy, you create a new policy version. A managed policy can have up to five versions. If the policy has five versions, you must delete an existing version using DeletePolicyVersion before you create a new version.

Optionally, you can set the new version as the policy's default version. The default version is the version that is in effect for the IAM users, groups, and roles to which the policy is attached.

For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

Creating a Request

createPolicyVersion

Arguments

:: Text
cpvPolicyARN

-> Text
cpvPolicyDocument

-> CreatePolicyVersion

Creates a value of CreatePolicyVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpvSetAsDefault - Specifies whether to set this version as the policy's default version. When this parameter is true , the new policy version becomes the operative version; that is, the version that is in effect for the IAM users, groups, and roles that the policy is attached to. For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

cpvPolicyARN - The Amazon Resource Name (ARN) of the IAM policy to which you want to add a new version. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

cpvPolicyDocument - The JSON policy document that you want to use as the content for this new version of the policy. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

data CreatePolicyVersion

See: createPolicyVersion smart constructor.

Request Lenses

cpvSetAsDefault :: Lens' CreatePolicyVersion (Maybe Bool)

Specifies whether to set this version as the policy's default version. When this parameter is true , the new policy version becomes the operative version; that is, the version that is in effect for the IAM users, groups, and roles that the policy is attached to. For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

cpvPolicyARN :: Lens' CreatePolicyVersion Text

The Amazon Resource Name (ARN) of the IAM policy to which you want to add a new version. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

cpvPolicyDocument :: Lens' CreatePolicyVersion Text

The JSON policy document that you want to use as the content for this new version of the policy. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

Destructuring the Response

createPolicyVersionResponse

Arguments

:: Int
cpvrsResponseStatus

-> CreatePolicyVersionResponse

Creates a value of CreatePolicyVersionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpvrsPolicyVersion - A structure containing details about the new policy version.

cpvrsResponseStatus - -- | The response status code.

data CreatePolicyVersionResponse

Contains the response to a successful CreatePolicyVersion request.

See: createPolicyVersionResponse smart constructor.

Response Lenses

cpvrsPolicyVersion :: Lens' CreatePolicyVersionResponse (Maybe PolicyVersion)

A structure containing details about the new policy version.

cpvrsResponseStatus :: Lens' CreatePolicyVersionResponse Int

- | The response status code.