Network.AWS.IAM.CreateRole

Description

Creates a new role for your AWS account. For more information about roles, go to Working with Roles . For information about limitations on role names and the number of roles you can create, go to Limitations on IAM Entities in the IAM User Guide .

Creating a Request

createRole

Arguments

:: Text
crRoleName

-> Text
crAssumeRolePolicyDocument

-> CreateRole

Creates a value of CreateRole with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crPath - The path to the role. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

crDescription - A customer-provided description of the role.

crRoleName - The name of the role to create. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@- Role names are not distinguished by case. For example, you cannot create roles named both PRODROLE and "prodrole".

crAssumeRolePolicyDocument - The trust relationship policy document that grants an entity permission to assume the role. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

data CreateRole

See: createRole smart constructor.

Request Lenses

crPath :: Lens' CreateRole (Maybe Text)

The path to the role. For more information about paths, see IAM Identifiers in the IAM User Guide . This parameter is optional. If it is not included, it defaults to a slash (). This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

crDescription :: Lens' CreateRole (Maybe Text)

A customer-provided description of the role.

crRoleName :: Lens' CreateRole Text

The name of the role to create. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@- Role names are not distinguished by case. For example, you cannot create roles named both PRODROLE and "prodrole".

crAssumeRolePolicyDocument :: Lens' CreateRole Text

The trust relationship policy document that grants an entity permission to assume the role. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

Destructuring the Response

createRoleResponse

Arguments

:: Int
crrsResponseStatus

-> Role
crrsRole

-> CreateRoleResponse

Creates a value of CreateRoleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crrsResponseStatus - -- | The response status code.

crrsRole - A structure containing details about the new role.

data CreateRoleResponse

Contains the response to a successful CreateRole request.

See: createRoleResponse smart constructor.

Response Lenses

crrsResponseStatus :: Lens' CreateRoleResponse Int

- | The response status code.

crrsRole :: Lens' CreateRoleResponse Role

A structure containing details about the new role.