Network.AWS.IAM.CreateRole

Description

Creates a new role for your AWS account. For more information about roles, go to Working with Roles. For information about limitations on role names and the number of roles you can create, go to Limitations on IAM Entities in the Using IAM guide.

The example policy grants permission to an EC2 instance to assume the role. The policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http://www.faqs.org/rfcs/rfc3986.html.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateRole.html

Request

data CreateRole

Request constructor

createRole

Arguments

:: Text
crRoleName

-> Text
crAssumeRolePolicyDocument

-> CreateRole

CreateRole constructor.

The fields accessible through corresponding lenses are:

crAssumeRolePolicyDocument :: Text

crPath :: Maybe Text

crRoleName :: Text

Request lenses

crAssumeRolePolicyDocument :: Lens' CreateRole Text

The policy that grants an entity permission to assume the role.

crPath :: Lens' CreateRole (Maybe Text)

The path to the role. For more information about paths, see IAM Identifiers in the Using IAM guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

crRoleName :: Lens' CreateRole Text

The name of the role to create.

Response

data CreateRoleResponse

Response constructor

createRoleResponse

Arguments

:: Role
crrRole

-> CreateRoleResponse

CreateRoleResponse constructor.

The fields accessible through corresponding lenses are:

crrRole :: Role

Response lenses

crrRole :: Lens' CreateRoleResponse Role

Information about the role.