Network.AWS.IAM.CreateSAMLProvider

Description

Creates an IAM resource that describes an identity provider (IdP) that supports SAML 2.0.

The SAML provider resource that you create with this operation can be used as a principal in an IAM role's trust policy to enable federated users who sign-in using the SAML IdP to assume the role. You can create an IAM role that supports Web-based single sign-on (SSO) to the AWS Management Console or one that supports API access to AWS.

When you create the SAML provider resource, you upload an a SAML metadata document that you get from your IdP and that includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that the IdP sends. You must generate the metadata document using the identity management software that is used as your organization's IdP.

For more information, see Enabling SAML 2.0 Federated Users to Access the AWS Management Console and About SAML 2.0-based Federation in the IAM User Guide .

Creating a Request

createSAMLProvider

Arguments

:: Text
csamlpSAMLMetadataDocument

-> Text
csamlpName

-> CreateSAMLProvider

Creates a value of CreateSAMLProvider with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csamlpSAMLMetadataDocument - An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP. For more information, see About SAML 2.0-based Federation in the IAM User Guide

csamlpName - The name of the provider to create. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data CreateSAMLProvider

See: createSAMLProvider smart constructor.

Request Lenses

csamlpSAMLMetadataDocument :: Lens' CreateSAMLProvider Text

An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP. For more information, see About SAML 2.0-based Federation in the IAM User Guide

csamlpName :: Lens' CreateSAMLProvider Text

The name of the provider to create. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

createSAMLProviderResponse

Arguments

:: Int
csamlprsResponseStatus

-> CreateSAMLProviderResponse

Creates a value of CreateSAMLProviderResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csamlprsSAMLProviderARN - The Amazon Resource Name (ARN) of the new SAML provider resource in IAM.

csamlprsResponseStatus - -- | The response status code.

data CreateSAMLProviderResponse

Contains the response to a successful CreateSAMLProvider request.

See: createSAMLProviderResponse smart constructor.

Response Lenses

csamlprsSAMLProviderARN :: Lens' CreateSAMLProviderResponse (Maybe Text)

The Amazon Resource Name (ARN) of the new SAML provider resource in IAM.

csamlprsResponseStatus :: Lens' CreateSAMLProviderResponse Int

- | The response status code.