Network.AWS.IAM.CreateSAMLProvider

Description

Creates an IAM entity to describe an identity provider (IdP) that supports SAML 2.0.

The SAML provider that you create with this operation can be used as a principal in a role's trust policy to establish a trust relationship between AWS and a SAML identity provider. You can create an IAM role that supports Web-based single sign-on (SSO) to the AWS Management Console or one that supports API access to AWS.

When you create the SAML provider, you upload an a SAML metadata document that you get from your IdP and that includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

This operation requires Signature Version 4. For more information, see Giving Console Access Using SAML and Creating Temporary Security Credentials for SAML Federation in the UsingTemporary Credentials guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateSAMLProvider.html

Request

data CreateSAMLProvider

Request constructor

createSAMLProvider

Arguments

:: Text
csamlpSAMLMetadataDocument

-> Text
csamlpName

-> CreateSAMLProvider

CreateSAMLProvider constructor.

The fields accessible through corresponding lenses are:

csamlpName :: Text

csamlpSAMLMetadataDocument :: Text

Request lenses

csamlpName :: Lens' CreateSAMLProvider Text

The name of the provider to create.

csamlpSAMLMetadataDocument :: Lens' CreateSAMLProvider Text

An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

For more information, see Creating Temporary Security Credentials for SAMLFederation in the Using Temporary Security Credentials guide.

Response

data CreateSAMLProviderResponse

Response constructor

createSAMLProviderResponse :: CreateSAMLProviderResponse

CreateSAMLProviderResponse constructor.

The fields accessible through corresponding lenses are:

csamlprSAMLProviderArn :: Maybe Text

Response lenses

csamlprSAMLProviderArn :: Lens' CreateSAMLProviderResponse (Maybe Text)

The Amazon Resource Name (ARN) of the SAML provider.