Network.AWS.IAM.CreateServiceSpecificCredential

Description

Generates a set of credentials consisting of a user name and password that can be used to access the service specified in the request. These credentials are generated by IAM, and can be used only for the specified service.

You can have a maximum of two sets of service-specific credentials for each supported service per user.

The only supported service at this time is AWS CodeCommit.

You can reset the password to a new service-generated value by calling ResetServiceSpecificCredential .

For more information about service-specific credentials, see Using IAM with AWS CodeCommit: Git Credentials, SSH Keys, and AWS Access Keys in the IAM User Guide .

Creating a Request

createServiceSpecificCredential

Arguments

:: Text
csscUserName

-> Text
csscServiceName

-> CreateServiceSpecificCredential

Creates a value of CreateServiceSpecificCredential with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csscUserName - The name of the IAM user that is to be associated with the credentials. The new service-specific credentials have the same permissions as the associated user except that they can be used only to access the specified service. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

csscServiceName - The name of the AWS service that is to be associated with the credentials. The service you specify here is the only service that can be accessed using these credentials.

data CreateServiceSpecificCredential

See: createServiceSpecificCredential smart constructor.

Request Lenses

csscUserName :: Lens' CreateServiceSpecificCredential Text

The name of the IAM user that is to be associated with the credentials. The new service-specific credentials have the same permissions as the associated user except that they can be used only to access the specified service. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

csscServiceName :: Lens' CreateServiceSpecificCredential Text

The name of the AWS service that is to be associated with the credentials. The service you specify here is the only service that can be accessed using these credentials.

Destructuring the Response

createServiceSpecificCredentialResponse

Arguments

:: Int
csscrsResponseStatus

-> CreateServiceSpecificCredentialResponse

Creates a value of CreateServiceSpecificCredentialResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csscrsServiceSpecificCredential - A structure that contains information about the newly created service-specific credential. Important: This is the only time that the password for this credential set is available. It cannot be recovered later. Instead, you will have to reset the password with ResetServiceSpecificCredential .

csscrsResponseStatus - -- | The response status code.

data CreateServiceSpecificCredentialResponse

See: createServiceSpecificCredentialResponse smart constructor.

Response Lenses

csscrsServiceSpecificCredential :: Lens' CreateServiceSpecificCredentialResponse (Maybe ServiceSpecificCredential)

A structure that contains information about the newly created service-specific credential. Important: This is the only time that the password for this credential set is available. It cannot be recovered later. Instead, you will have to reset the password with ResetServiceSpecificCredential .

csscrsResponseStatus :: Lens' CreateServiceSpecificCredentialResponse Int

- | The response status code.