Network.AWS.IAM.DeleteGroupPolicy

Description

Deletes the specified inline policy that is embedded in the specified IAM group.

A group can also have managed policies attached to it. To detach a managed policy from a group, use DetachGroupPolicy . For more information about policies, refer to Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

deleteGroupPolicy

Arguments

:: Text
dGroupName

-> Text
dPolicyName

-> DeleteGroupPolicy

Creates a value of DeleteGroupPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dGroupName - The name (friendly name, not ARN) identifying the group that the policy is embedded in. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

dPolicyName - The name identifying the policy document to delete. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data DeleteGroupPolicy

See: deleteGroupPolicy smart constructor.

Request Lenses

dGroupName :: Lens' DeleteGroupPolicy Text

The name (friendly name, not ARN) identifying the group that the policy is embedded in. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

dPolicyName :: Lens' DeleteGroupPolicy Text

The name identifying the policy document to delete. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

deleteGroupPolicyResponse :: DeleteGroupPolicyResponse

Creates a value of DeleteGroupPolicyResponse with the minimum fields required to make a request.

data DeleteGroupPolicyResponse

See: deleteGroupPolicyResponse smart constructor.