Network.AWS.IAM.DeletePolicyVersion

Description

Deletes the specified version from the specified managed policy.

You cannot delete the default version from a policy using this API. To delete the default version from a policy, use DeletePolicy . To find out which version of a policy is marked as the default version, use ListPolicyVersions .

For information about versions for managed policies, see Versioning for Managed Policies in the IAM User Guide .

Creating a Request

deletePolicyVersion

Arguments

:: Text
dpvPolicyARN

-> Text
dpvVersionId

-> DeletePolicyVersion

Creates a value of DeletePolicyVersion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpvPolicyARN - The Amazon Resource Name (ARN) of the IAM policy from which you want to delete a version. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

dpvVersionId - The policy version to delete. This parameter allows (per its regex pattern ) a string of characters that consists of the lowercase letter v followed by one or two digits, and optionally followed by a period . and a string of letters and digits. For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

data DeletePolicyVersion

See: deletePolicyVersion smart constructor.

Request Lenses

dpvPolicyARN :: Lens' DeletePolicyVersion Text

The Amazon Resource Name (ARN) of the IAM policy from which you want to delete a version. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

dpvVersionId :: Lens' DeletePolicyVersion Text

The policy version to delete. This parameter allows (per its regex pattern ) a string of characters that consists of the lowercase letter v followed by one or two digits, and optionally followed by a period . and a string of letters and digits. For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

Destructuring the Response

deletePolicyVersionResponse :: DeletePolicyVersionResponse

Creates a value of DeletePolicyVersionResponse with the minimum fields required to make a request.

data DeletePolicyVersionResponse

See: deletePolicyVersionResponse smart constructor.