Network.AWS.IAM.DeleteSAMLProvider

Description

Deletes a SAML provider resource in IAM.

Deleting the provider resource from IAM does not update any roles that reference the SAML provider resource's ARN as a principal in their trust policies. Any attempt to assume a role that references a non-existent provider resource ARN fails.

Creating a Request

deleteSAMLProvider

Arguments

:: Text
dsamlpSAMLProviderARN

-> DeleteSAMLProvider

Creates a value of DeleteSAMLProvider with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsamlpSAMLProviderARN - The Amazon Resource Name (ARN) of the SAML provider to delete.

data DeleteSAMLProvider

See: deleteSAMLProvider smart constructor.

Request Lenses

dsamlpSAMLProviderARN :: Lens' DeleteSAMLProvider Text

The Amazon Resource Name (ARN) of the SAML provider to delete.

Destructuring the Response

deleteSAMLProviderResponse :: DeleteSAMLProviderResponse

Creates a value of DeleteSAMLProviderResponse with the minimum fields required to make a request.

data DeleteSAMLProviderResponse

See: deleteSAMLProviderResponse smart constructor.