Network.AWS.IAM.EnableMFADevice

Description

Enables the specified MFA device and associates it with the specified IAM user. When enabled, the MFA device is required for every subsequent login by the IAM user associated with the device.

Creating a Request

enableMFADevice

Arguments

:: Text
emdUserName

-> Text
emdSerialNumber

-> Text
emdAuthenticationCode1

-> Text
emdAuthenticationCode2

-> EnableMFADevice

Creates a value of EnableMFADevice with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

emdUserName - The name of the IAM user for whom you want to enable the MFA device. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

emdSerialNumber - The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@:/-

emdAuthenticationCode1 - An authentication code emitted by the device. The format for this parameter is a string of 6 digits. Important: Submit your request immediately after generating the authentication codes. If you generate the codes and then wait too long to submit the request, the MFA device successfully associates with the user but the MFA device becomes out of sync. This happens because time-based one-time passwords (TOTP) expire after a short period of time. If this happens, you can resync the device .

emdAuthenticationCode2 - A subsequent authentication code emitted by the device. The format for this parameter is a string of 6 digits. Important: Submit your request immediately after generating the authentication codes. If you generate the codes and then wait too long to submit the request, the MFA device successfully associates with the user but the MFA device becomes out of sync. This happens because time-based one-time passwords (TOTP) expire after a short period of time. If this happens, you can resync the device .

data EnableMFADevice

See: enableMFADevice smart constructor.

Request Lenses

emdUserName :: Lens' EnableMFADevice Text

The name of the IAM user for whom you want to enable the MFA device. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

emdSerialNumber :: Lens' EnableMFADevice Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@:/-

emdAuthenticationCode1 :: Lens' EnableMFADevice Text

An authentication code emitted by the device. The format for this parameter is a string of 6 digits. Important: Submit your request immediately after generating the authentication codes. If you generate the codes and then wait too long to submit the request, the MFA device successfully associates with the user but the MFA device becomes out of sync. This happens because time-based one-time passwords (TOTP) expire after a short period of time. If this happens, you can resync the device .

emdAuthenticationCode2 :: Lens' EnableMFADevice Text

A subsequent authentication code emitted by the device. The format for this parameter is a string of 6 digits. Important: Submit your request immediately after generating the authentication codes. If you generate the codes and then wait too long to submit the request, the MFA device successfully associates with the user but the MFA device becomes out of sync. This happens because time-based one-time passwords (TOTP) expire after a short period of time. If this happens, you can resync the device .

Destructuring the Response

enableMFADeviceResponse :: EnableMFADeviceResponse

Creates a value of EnableMFADeviceResponse with the minimum fields required to make a request.

data EnableMFADeviceResponse

See: enableMFADeviceResponse smart constructor.