Network.AWS.IAM.GetAccountAuthorizationDetails

Description

Retrieves information about all IAM users, groups, roles, and policies in your AWS account, including their relationships to one another. Use this API to obtain a snapshot of the configuration of IAM permissions (users, groups, roles, and policies) in your account.

You can optionally filter the results using the Filter parameter. You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

getAccountAuthorizationDetails :: GetAccountAuthorizationDetails

Creates a value of GetAccountAuthorizationDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaadMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

gaadMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

gaadFilter - A list of entity types used to filter the results. Only the entities that match the types you specify are included in the output. Use the value LocalManagedPolicy to include customer managed policies. The format for this parameter is a comma-separated (if more than one) list of strings. Each string value in the list must be one of the valid values listed below.

data GetAccountAuthorizationDetails

See: getAccountAuthorizationDetails smart constructor.

Request Lenses

gaadMarker :: Lens' GetAccountAuthorizationDetails (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

gaadMaxItems :: Lens' GetAccountAuthorizationDetails (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

gaadFilter :: Lens' GetAccountAuthorizationDetails [EntityType]

A list of entity types used to filter the results. Only the entities that match the types you specify are included in the output. Use the value LocalManagedPolicy to include customer managed policies. The format for this parameter is a comma-separated (if more than one) list of strings. Each string value in the list must be one of the valid values listed below.

Destructuring the Response

getAccountAuthorizationDetailsResponse

Arguments

:: Int
gaadrsResponseStatus

-> GetAccountAuthorizationDetailsResponse

Creates a value of GetAccountAuthorizationDetailsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gaadrsRoleDetailList - A list containing information about IAM roles.

gaadrsGroupDetailList - A list containing information about IAM groups.

gaadrsUserDetailList - A list containing information about IAM users.

gaadrsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

gaadrsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

gaadrsPolicies - A list containing information about managed policies.

gaadrsResponseStatus - -- | The response status code.

data GetAccountAuthorizationDetailsResponse

Contains the response to a successful GetAccountAuthorizationDetails request.

See: getAccountAuthorizationDetailsResponse smart constructor.

Response Lenses

gaadrsRoleDetailList :: Lens' GetAccountAuthorizationDetailsResponse [RoleDetail]

A list containing information about IAM roles.

gaadrsGroupDetailList :: Lens' GetAccountAuthorizationDetailsResponse [GroupDetail]

A list containing information about IAM groups.

gaadrsUserDetailList :: Lens' GetAccountAuthorizationDetailsResponse [UserDetail]

A list containing information about IAM users.

gaadrsMarker :: Lens' GetAccountAuthorizationDetailsResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

gaadrsIsTruncated :: Lens' GetAccountAuthorizationDetailsResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

gaadrsPolicies :: Lens' GetAccountAuthorizationDetailsResponse [ManagedPolicyDetail]

A list containing information about managed policies.

gaadrsResponseStatus :: Lens' GetAccountAuthorizationDetailsResponse Int

- | The response status code.