Network.AWS.IAM.GetAccountPasswordPolicy

Description

Retrieves the password policy for the AWS account. For more information about using a password policy, go to Managing an IAM Password Policy .

Creating a Request

getAccountPasswordPolicy :: GetAccountPasswordPolicy

Creates a value of GetAccountPasswordPolicy with the minimum fields required to make a request.

data GetAccountPasswordPolicy

See: getAccountPasswordPolicy smart constructor.

Destructuring the Response

getAccountPasswordPolicyResponse

Arguments

:: Int
gapprsResponseStatus

-> PasswordPolicy
gapprsPasswordPolicy

-> GetAccountPasswordPolicyResponse

Creates a value of GetAccountPasswordPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gapprsResponseStatus - -- | The response status code.

gapprsPasswordPolicy - A structure that contains details about the account's password policy.

data GetAccountPasswordPolicyResponse

Contains the response to a successful GetAccountPasswordPolicy request.

See: getAccountPasswordPolicyResponse smart constructor.

Response Lenses

gapprsResponseStatus :: Lens' GetAccountPasswordPolicyResponse Int

- | The response status code.

gapprsPasswordPolicy :: Lens' GetAccountPasswordPolicyResponse PasswordPolicy

A structure that contains details about the account's password policy.