Network.AWS.IAM.GetAccountSummary

Description

Retrieves information about IAM entity usage and IAM quotas in the AWS account.

For information about limitations on IAM entities, see Limitations on IAM Entities in the IAM User Guide .

Creating a Request

getAccountSummary :: GetAccountSummary

Creates a value of GetAccountSummary with the minimum fields required to make a request.

data GetAccountSummary

See: getAccountSummary smart constructor.

Destructuring the Response

getAccountSummaryResponse

Arguments

:: Int
gasrsResponseStatus

-> GetAccountSummaryResponse

Creates a value of GetAccountSummaryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gasrsSummaryMap - A set of key value pairs containing information about IAM entity usage and IAM quotas.

gasrsResponseStatus - -- | The response status code.

data GetAccountSummaryResponse

Contains the response to a successful GetAccountSummary request.

See: getAccountSummaryResponse smart constructor.

Response Lenses

gasrsSummaryMap :: Lens' GetAccountSummaryResponse (HashMap SummaryKeyType Int)

A set of key value pairs containing information about IAM entity usage and IAM quotas.

gasrsResponseStatus :: Lens' GetAccountSummaryResponse Int

- | The response status code.