Network.AWS.IAM.GetCredentialReport

Description

Retrieves a credential report for the AWS account. For more information about the credential report, see Getting Credential Reports in the IAM User Guide .

Creating a Request

getCredentialReport :: GetCredentialReport

Creates a value of GetCredentialReport with the minimum fields required to make a request.

data GetCredentialReport

See: getCredentialReport smart constructor.

Destructuring the Response

getCredentialReportResponse

Arguments

:: Int
grsResponseStatus

-> GetCredentialReportResponse

Creates a value of GetCredentialReportResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grsContent - Contains the credential report. The report is Base64-encoded.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

grsGeneratedTime - The date and time when the credential report was created, in ISO 8601 date-time format .

grsReportFormat - The format (MIME type) of the credential report.

grsResponseStatus - -- | The response status code.

data GetCredentialReportResponse

Contains the response to a successful GetCredentialReport request.

See: getCredentialReportResponse smart constructor.

Response Lenses

grsContent :: Lens' GetCredentialReportResponse (Maybe ByteString)

Contains the credential report. The report is Base64-encoded.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

grsGeneratedTime :: Lens' GetCredentialReportResponse (Maybe UTCTime)

The date and time when the credential report was created, in ISO 8601 date-time format .

grsReportFormat :: Lens' GetCredentialReportResponse (Maybe ReportFormatType)

The format (MIME type) of the credential report.

grsResponseStatus :: Lens' GetCredentialReportResponse Int

- | The response status code.