Network.AWS.IAM.GetCredentialReport

Description

Retrieves a credential report for the AWS account. For more information about the credential report, see Getting Credential Reports in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetCredentialReport.html

Request

data GetCredentialReport

Request constructor

getCredentialReport :: GetCredentialReport

GetCredentialReport constructor.

Response

data GetCredentialReportResponse

Response constructor

getCredentialReportResponse :: GetCredentialReportResponse

GetCredentialReportResponse constructor.

The fields accessible through corresponding lenses are:

gcrrContent :: Maybe Base64

gcrrGeneratedTime :: Maybe UTCTime

gcrrReportFormat :: Maybe ReportFormatType

Response lenses

gcrrContent :: Lens' GetCredentialReportResponse (Maybe Base64)

Contains the credential report. The report is Base64-encoded.

gcrrGeneratedTime :: Lens' GetCredentialReportResponse (Maybe UTCTime)

The date and time when the credential report was created, in ISO 8601date-time format.

gcrrReportFormat :: Lens' GetCredentialReportResponse (Maybe ReportFormatType)

The format (MIME type) of the credential report.