Network.AWS.IAM.GetGroupPolicy

Description

Retrieves the specified inline policy document that is embedded in the specified IAM group.

An IAM group can also have managed policies attached to it. To retrieve a managed policy document that is attached to a group, use GetPolicy to determine the policy's default version, then use GetPolicyVersion to retrieve the policy document.

For more information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

getGroupPolicy

Arguments

:: Text
ggpGroupName

-> Text
ggpPolicyName

-> GetGroupPolicy

Creates a value of GetGroupPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ggpGroupName - The name of the group the policy is associated with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

ggpPolicyName - The name of the policy document to get. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data GetGroupPolicy

See: getGroupPolicy smart constructor.

Request Lenses

ggpGroupName :: Lens' GetGroupPolicy Text

The name of the group the policy is associated with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

ggpPolicyName :: Lens' GetGroupPolicy Text

The name of the policy document to get. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

getGroupPolicyResponse

Arguments

:: Int
ggprsResponseStatus

-> Text
ggprsGroupName

-> Text
ggprsPolicyName

-> Text
ggprsPolicyDocument

-> GetGroupPolicyResponse

Creates a value of GetGroupPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ggprsResponseStatus - -- | The response status code.

ggprsGroupName - The group the policy is associated with.

ggprsPolicyName - The name of the policy.

ggprsPolicyDocument - The policy document.

data GetGroupPolicyResponse

Contains the response to a successful GetGroupPolicy request.

See: getGroupPolicyResponse smart constructor.

Response Lenses

ggprsResponseStatus :: Lens' GetGroupPolicyResponse Int

- | The response status code.

ggprsGroupName :: Lens' GetGroupPolicyResponse Text

The group the policy is associated with.

ggprsPolicyName :: Lens' GetGroupPolicyResponse Text

The name of the policy.

ggprsPolicyDocument :: Lens' GetGroupPolicyResponse Text

The policy document.