Network.AWS.IAM.GetLoginProfile

Description

Retrieves the user name and password-creation date for the specified IAM user. If the user has not been assigned a password, the action returns a 404 (NoSuchEntity ) error.

Creating a Request

getLoginProfile

Arguments

:: Text
glpUserName

-> GetLoginProfile

Creates a value of GetLoginProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

glpUserName - The name of the user whose login profile you want to retrieve. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data GetLoginProfile

See: getLoginProfile smart constructor.

Request Lenses

glpUserName :: Lens' GetLoginProfile Text

The name of the user whose login profile you want to retrieve. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

getLoginProfileResponse

Arguments

:: Int
glprsResponseStatus

-> LoginProfile
glprsLoginProfile

-> GetLoginProfileResponse

Creates a value of GetLoginProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

glprsResponseStatus - -- | The response status code.

glprsLoginProfile - A structure containing the user name and password create date for the user.

data GetLoginProfileResponse

Contains the response to a successful GetLoginProfile request.

See: getLoginProfileResponse smart constructor.

Response Lenses

glprsResponseStatus :: Lens' GetLoginProfileResponse Int

- | The response status code.

glprsLoginProfile :: Lens' GetLoginProfileResponse LoginProfile

A structure containing the user name and password create date for the user.