Network.AWS.IAM.GetPolicy

Description

Retrieves information about the specified managed policy, including the policy's default version and the total number of IAM users, groups, and roles to which the policy is attached. To retrieve the list of the specific users, groups, and roles that the policy is attached to, use the ListEntitiesForPolicy API. This API returns metadata about the policy. To retrieve the actual policy document for a specific version of the policy, use GetPolicyVersion .

This API retrieves information about managed policies. To retrieve information about an inline policy that is embedded with an IAM user, group, or role, use the GetUserPolicy , GetGroupPolicy , or GetRolePolicy API.

For more information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

getPolicy

Arguments

:: Text
gpPolicyARN

-> GetPolicy

Creates a value of GetPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpPolicyARN - The Amazon Resource Name (ARN) of the managed policy that you want information about. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data GetPolicy

See: getPolicy smart constructor.

Request Lenses

gpPolicyARN :: Lens' GetPolicy Text

The Amazon Resource Name (ARN) of the managed policy that you want information about. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

getPolicyResponse

Arguments

:: Int
gprsResponseStatus

-> GetPolicyResponse

Creates a value of GetPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gprsPolicy - A structure containing details about the policy.

gprsResponseStatus - -- | The response status code.

data GetPolicyResponse

Contains the response to a successful GetPolicy request.

See: getPolicyResponse smart constructor.

Response Lenses

gprsPolicy :: Lens' GetPolicyResponse (Maybe Policy)

A structure containing details about the policy.

gprsResponseStatus :: Lens' GetPolicyResponse Int

- | The response status code.