Network.AWS.IAM.GetRole

Description

Retrieves information about the specified role, including the role's path, GUID, ARN, and the role's trust policy that grants permission to assume the role. For more information about roles, see Working with Roles .

Creating a Request

getRole

Arguments

:: Text
grRoleName

-> GetRole

Creates a value of GetRole with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grRoleName - The name of the IAM role to get information about. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

data GetRole

See: getRole smart constructor.

Request Lenses

grRoleName :: Lens' GetRole Text

The name of the IAM role to get information about. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

getRoleResponse

Arguments

:: Int
grrsResponseStatus

-> Role
grrsRole

-> GetRoleResponse

Creates a value of GetRoleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grrsResponseStatus - -- | The response status code.

grrsRole - A structure containing details about the IAM role.

data GetRoleResponse

Contains the response to a successful GetRole request.

See: getRoleResponse smart constructor.

Response Lenses

grrsResponseStatus :: Lens' GetRoleResponse Int

- | The response status code.

grrsRole :: Lens' GetRoleResponse Role

A structure containing details about the IAM role.