Network.AWS.IAM.GetSAMLProvider

Description

Returns the SAML provider metadocument that was uploaded when the IAM SAML provider resource object was created or updated.

Creating a Request

getSAMLProvider

Arguments

:: Text
gsamlpSAMLProviderARN

-> GetSAMLProvider

Creates a value of GetSAMLProvider with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsamlpSAMLProviderARN - The Amazon Resource Name (ARN) of the SAML provider resource object in IAM to get information about. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data GetSAMLProvider

See: getSAMLProvider smart constructor.

Request Lenses

gsamlpSAMLProviderARN :: Lens' GetSAMLProvider Text

The Amazon Resource Name (ARN) of the SAML provider resource object in IAM to get information about. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

getSAMLProviderResponse

Arguments

:: Int
gsamlprsResponseStatus

-> GetSAMLProviderResponse

Creates a value of GetSAMLProviderResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsamlprsCreateDate - The date and time when the SAML provider was created.

gsamlprsValidUntil - The expiration date and time for the SAML provider.

gsamlprsSAMLMetadataDocument - The XML metadata document that includes information about an identity provider.

gsamlprsResponseStatus - -- | The response status code.

data GetSAMLProviderResponse

Contains the response to a successful GetSAMLProvider request.

See: getSAMLProviderResponse smart constructor.

Response Lenses

gsamlprsCreateDate :: Lens' GetSAMLProviderResponse (Maybe UTCTime)

The date and time when the SAML provider was created.

gsamlprsValidUntil :: Lens' GetSAMLProviderResponse (Maybe UTCTime)

The expiration date and time for the SAML provider.

gsamlprsSAMLMetadataDocument :: Lens' GetSAMLProviderResponse (Maybe Text)

The XML metadata document that includes information about an identity provider.

gsamlprsResponseStatus :: Lens' GetSAMLProviderResponse Int

- | The response status code.