Network.AWS.IAM.GetSAMLProvider

Description

Returns the SAML provider metadocument that was uploaded when the provider was created or updated.

This operation requires Signature Version 4.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetSAMLProvider.html

Request

data GetSAMLProvider

Request constructor

getSAMLProvider

Arguments

:: Text
gsamlpSAMLProviderArn

-> GetSAMLProvider

GetSAMLProvider constructor.

The fields accessible through corresponding lenses are:

gsamlpSAMLProviderArn :: Text

Request lenses

gsamlpSAMLProviderArn :: Lens' GetSAMLProvider Text

The Amazon Resource Name (ARN) of the SAML provider to get information about.

Response

data GetSAMLProviderResponse

Response constructor

getSAMLProviderResponse :: GetSAMLProviderResponse

GetSAMLProviderResponse constructor.

The fields accessible through corresponding lenses are:

gsamlprCreateDate :: Maybe UTCTime

gsamlprSAMLMetadataDocument :: Maybe Text

gsamlprValidUntil :: Maybe UTCTime

Response lenses

gsamlprCreateDate :: Lens' GetSAMLProviderResponse (Maybe UTCTime)

The date and time when the SAML provider was created.

gsamlprSAMLMetadataDocument :: Lens' GetSAMLProviderResponse (Maybe Text)

The XML metadata document that includes information about an identity provider.

gsamlprValidUntil :: Lens' GetSAMLProviderResponse (Maybe UTCTime)

The expiration date and time for the SAML provider.