Network.AWS.IAM.GetSSHPublicKey

Description

Retrieves the specified SSH public key, including metadata about the key.

The SSH public key retrieved by this action is used only for authenticating the associated IAM user to an AWS CodeCommit repository. For more information about using SSH keys to authenticate to an AWS CodeCommit repository, see Set up AWS CodeCommit for SSH Connections in the AWS CodeCommit User Guide .

Creating a Request

getSSHPublicKey

Arguments

:: Text
gspkUserName

-> Text
gspkSSHPublicKeyId

-> EncodingType
gspkEncoding

-> GetSSHPublicKey

Creates a value of GetSSHPublicKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gspkUserName - The name of the IAM user associated with the SSH public key. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

gspkSSHPublicKeyId - The unique identifier for the SSH public key. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

gspkEncoding - Specifies the public key encoding format to use in the response. To retrieve the public key in ssh-rsa format, use SSH . To retrieve the public key in PEM format, use PEM .

data GetSSHPublicKey

See: getSSHPublicKey smart constructor.

Request Lenses

gspkUserName :: Lens' GetSSHPublicKey Text

The name of the IAM user associated with the SSH public key. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

gspkSSHPublicKeyId :: Lens' GetSSHPublicKey Text

The unique identifier for the SSH public key. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

gspkEncoding :: Lens' GetSSHPublicKey EncodingType

Specifies the public key encoding format to use in the response. To retrieve the public key in ssh-rsa format, use SSH . To retrieve the public key in PEM format, use PEM .

Destructuring the Response

getSSHPublicKeyResponse

Arguments

:: Int
gspkrsResponseStatus

-> GetSSHPublicKeyResponse

Creates a value of GetSSHPublicKeyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gspkrsSSHPublicKey - A structure containing details about the SSH public key.

gspkrsResponseStatus - -- | The response status code.

data GetSSHPublicKeyResponse

Contains the response to a successful GetSSHPublicKey request.

See: getSSHPublicKeyResponse smart constructor.

Response Lenses

gspkrsSSHPublicKey :: Lens' GetSSHPublicKeyResponse (Maybe SSHPublicKey)

A structure containing details about the SSH public key.

gspkrsResponseStatus :: Lens' GetSSHPublicKeyResponse Int

- | The response status code.