Network.AWS.IAM.GetUser

Description

Retrieves information about the specified IAM user, including the user's creation date, path, unique ID, and ARN.

If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID used to sign the request to this API.

Creating a Request

getUser :: GetUser

Creates a value of GetUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

guUserName - The name of the user to get information about. This parameter is optional. If it is not included, it defaults to the user making the request. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data GetUser

See: getUser smart constructor.

Request Lenses

guUserName :: Lens' GetUser (Maybe Text)

The name of the user to get information about. This parameter is optional. If it is not included, it defaults to the user making the request. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

getUserResponse

Arguments

:: Int
gursResponseStatus

-> User
gursUser

-> GetUserResponse

Creates a value of GetUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gursResponseStatus - -- | The response status code.

gursUser - A structure containing details about the IAM user.

data GetUserResponse

Contains the response to a successful GetUser request.

See: getUserResponse smart constructor.

Response Lenses

gursResponseStatus :: Lens' GetUserResponse Int

- | The response status code.

gursUser :: Lens' GetUserResponse User

A structure containing details about the IAM user.