Network.AWS.IAM.GetUser

Description

Retrieves information about the specified user, including the user's creation date, path, unique ID, and ARN.

If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID used to sign the request.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetUser.html

Request

data GetUser

Request constructor

getUser :: GetUser

GetUser constructor.

The fields accessible through corresponding lenses are:

guUserName :: Maybe Text

Request lenses

guUserName :: Lens' GetUser (Maybe Text)

The name of the user to get information about.

This parameter is optional. If it is not included, it defaults to the user making the request.

Response

data GetUserResponse

Response constructor

getUserResponse

Arguments

:: User
gurUser

-> GetUserResponse

GetUserResponse constructor.

The fields accessible through corresponding lenses are:

gurUser :: User

Response lenses

gurUser :: Lens' GetUserResponse User

Information about the user.