Network.AWS.IAM.GetUserPolicy

Description

Retrieves the specified inline policy document that is embedded in the specified IAM user.

An IAM user can also have managed policies attached to it. To retrieve a managed policy document that is attached to a user, use GetPolicy to determine the policy's default version, then use GetPolicyVersion to retrieve the policy document.

For more information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

Creating a Request

getUserPolicy

Arguments

:: Text
gupUserName

-> Text
gupPolicyName

-> GetUserPolicy

Creates a value of GetUserPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gupUserName - The name of the user who the policy is associated with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

gupPolicyName - The name of the policy document to get. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data GetUserPolicy

See: getUserPolicy smart constructor.

Request Lenses

gupUserName :: Lens' GetUserPolicy Text

The name of the user who the policy is associated with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

gupPolicyName :: Lens' GetUserPolicy Text

The name of the policy document to get. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

getUserPolicyResponse

Arguments

:: Int
guprsResponseStatus

-> Text
guprsUserName

-> Text
guprsPolicyName

-> Text
guprsPolicyDocument

-> GetUserPolicyResponse

Creates a value of GetUserPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

guprsResponseStatus - -- | The response status code.

guprsUserName - The user the policy is associated with.

guprsPolicyName - The name of the policy.

guprsPolicyDocument - The policy document.

data GetUserPolicyResponse

Contains the response to a successful GetUserPolicy request.

See: getUserPolicyResponse smart constructor.

Response Lenses

guprsResponseStatus :: Lens' GetUserPolicyResponse Int

- | The response status code.

guprsUserName :: Lens' GetUserPolicyResponse Text

The user the policy is associated with.

guprsPolicyName :: Lens' GetUserPolicyResponse Text

The name of the policy.

guprsPolicyDocument :: Lens' GetUserPolicyResponse Text

The policy document.