Network.AWS.IAM.ListAccountAliases

Description

Lists the account alias associated with the AWS account (Note: you can have only one). For information about using an AWS account alias, see Using an Alias for Your AWS Account ID in the IAM User Guide .

This operation returns paginated results.

Creating a Request

listAccountAliases :: ListAccountAliases

Creates a value of ListAccountAliases with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laaMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

laaMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

data ListAccountAliases

See: listAccountAliases smart constructor.

Request Lenses

laaMarker :: Lens' ListAccountAliases (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

laaMaxItems :: Lens' ListAccountAliases (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Destructuring the Response

listAccountAliasesResponse

Arguments

:: Int
laarsResponseStatus

-> ListAccountAliasesResponse

Creates a value of ListAccountAliasesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laarsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

laarsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

laarsResponseStatus - -- | The response status code.

laarsAccountAliases - A list of aliases associated with the account. AWS supports only one alias per account.

data ListAccountAliasesResponse

Contains the response to a successful ListAccountAliases request.

See: listAccountAliasesResponse smart constructor.

Response Lenses

laarsMarker :: Lens' ListAccountAliasesResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

laarsIsTruncated :: Lens' ListAccountAliasesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

laarsResponseStatus :: Lens' ListAccountAliasesResponse Int

- | The response status code.

laarsAccountAliases :: Lens' ListAccountAliasesResponse [Text]

A list of aliases associated with the account. AWS supports only one alias per account.