Network.AWS.IAM.ListAttachedGroupPolicies

Description

Lists all managed policies that are attached to the specified IAM group.

An IAM group can also have inline policies embedded with it. To list the inline policies for a group, use the ListGroupPolicies API. For information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

You can paginate the results using the MaxItems and Marker parameters. You can use the PathPrefix parameter to limit the list of policies to only those matching the specified path prefix. If there are no policies attached to the specified group (or none that match the specified path prefix), the action returns an empty list.

This operation returns paginated results.

Creating a Request

listAttachedGroupPolicies

Arguments

:: Text
lagpGroupName

-> ListAttachedGroupPolicies

Creates a value of ListAttachedGroupPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lagpPathPrefix - The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (), listing all policies. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lagpMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lagpMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lagpGroupName - The name (friendly name, not ARN) of the group to list attached policies for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data ListAttachedGroupPolicies

See: listAttachedGroupPolicies smart constructor.

Request Lenses

lagpPathPrefix :: Lens' ListAttachedGroupPolicies (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (), listing all policies. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lagpMarker :: Lens' ListAttachedGroupPolicies (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lagpMaxItems :: Lens' ListAttachedGroupPolicies (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lagpGroupName :: Lens' ListAttachedGroupPolicies Text

The name (friendly name, not ARN) of the group to list attached policies for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

listAttachedGroupPoliciesResponse

Arguments

:: Int
lagprsResponseStatus

-> ListAttachedGroupPoliciesResponse

Creates a value of ListAttachedGroupPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lagprsAttachedPolicies - A list of the attached policies.

lagprsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lagprsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lagprsResponseStatus - -- | The response status code.

data ListAttachedGroupPoliciesResponse

Contains the response to a successful ListAttachedGroupPolicies request.

See: listAttachedGroupPoliciesResponse smart constructor.

Response Lenses

lagprsAttachedPolicies :: Lens' ListAttachedGroupPoliciesResponse [AttachedPolicy]

A list of the attached policies.

lagprsMarker :: Lens' ListAttachedGroupPoliciesResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lagprsIsTruncated :: Lens' ListAttachedGroupPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lagprsResponseStatus :: Lens' ListAttachedGroupPoliciesResponse Int

- | The response status code.