Network.AWS.IAM.ListEntitiesForPolicy

Description

Lists all IAM users, groups, and roles that the specified managed policy is attached to.

You can use the optional EntityFilter parameter to limit the results to a particular type of entity (users, groups, or roles). For example, to list only the roles that are attached to the specified policy, set EntityFilter to Role .

You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

listEntitiesForPolicy

Arguments

:: Text
lefpPolicyARN

-> ListEntitiesForPolicy

Creates a value of ListEntitiesForPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lefpPathPrefix - The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (), listing all entities. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lefpEntityFilter - The entity type to use for filtering the results. For example, when EntityFilter is Role , only the roles that are attached to the specified policy are returned. This parameter is optional. If it is not included, all attached entities (users, groups, and roles) are returned. The argument for this parameter must be one of the valid values listed below.

lefpMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lefpMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lefpPolicyARN - The Amazon Resource Name (ARN) of the IAM policy for which you want the versions. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data ListEntitiesForPolicy

See: listEntitiesForPolicy smart constructor.

Request Lenses

lefpPathPrefix :: Lens' ListEntitiesForPolicy (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (), listing all entities. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lefpEntityFilter :: Lens' ListEntitiesForPolicy (Maybe EntityType)

The entity type to use for filtering the results. For example, when EntityFilter is Role , only the roles that are attached to the specified policy are returned. This parameter is optional. If it is not included, all attached entities (users, groups, and roles) are returned. The argument for this parameter must be one of the valid values listed below.

lefpMarker :: Lens' ListEntitiesForPolicy (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lefpMaxItems :: Lens' ListEntitiesForPolicy (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lefpPolicyARN :: Lens' ListEntitiesForPolicy Text

The Amazon Resource Name (ARN) of the IAM policy for which you want the versions. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

listEntitiesForPolicyResponse

Arguments

:: Int
lefprsResponseStatus

-> ListEntitiesForPolicyResponse

Creates a value of ListEntitiesForPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lefprsPolicyGroups - A list of IAM groups that the policy is attached to.

lefprsPolicyRoles - A list of IAM roles that the policy is attached to.

lefprsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lefprsPolicyUsers - A list of IAM users that the policy is attached to.

lefprsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lefprsResponseStatus - -- | The response status code.

data ListEntitiesForPolicyResponse

Contains the response to a successful ListEntitiesForPolicy request.

See: listEntitiesForPolicyResponse smart constructor.

Response Lenses

lefprsPolicyGroups :: Lens' ListEntitiesForPolicyResponse [PolicyGroup]

A list of IAM groups that the policy is attached to.

lefprsPolicyRoles :: Lens' ListEntitiesForPolicyResponse [PolicyRole]

A list of IAM roles that the policy is attached to.

lefprsMarker :: Lens' ListEntitiesForPolicyResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lefprsPolicyUsers :: Lens' ListEntitiesForPolicyResponse [PolicyUser]

A list of IAM users that the policy is attached to.

lefprsIsTruncated :: Lens' ListEntitiesForPolicyResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lefprsResponseStatus :: Lens' ListEntitiesForPolicyResponse Int

- | The response status code.