Network.AWS.IAM.ListGroupPolicies

Description

Lists the names of the policies associated with the specified group. If there are none, the action returns an empty list. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListGroupPolicies &GroupName=Admins &AUTHPARAMS AdminRoot KeyPolicy false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data ListGroupPolicies

Request constructor

listGroupPolicies

Arguments

:: Text
lgpGroupName

-> ListGroupPolicies

Smart constructor for the minimum required parameters to construct a valid ListGroupPolicies request.

The fields accessible through corresponding lenses are:

GroupName :: Text

Marker :: Maybe Text

MaxItems :: Maybe Integer

Request lenses

lgpGroupName :: Lens' ListGroupPolicies Text

The name of the group to list policies for.

lgpMarker :: Lens' ListGroupPolicies (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lgpMaxItems :: Lens' ListGroupPolicies (Maybe Integer)

Use this only when paginating results to indicate the maximum number of policy names you want in the response. If there are additional policy names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListGroupPoliciesResponse

Contains the result of a successful invocation of the ListGroupPolicies action.

Response constructor

listGroupPoliciesResponse

Arguments

:: [Text]
lgprPolicyNames

-> Bool
lgprIsTruncated

-> ListGroupPoliciesResponse

Smart constructor for the minimum required parameters to construct a valid ListGroupPoliciesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

PolicyNames :: [Text]

IsTruncated :: Bool

Marker :: Maybe Text

Response lenses

lgprPolicyNames :: Lens' ListGroupPoliciesResponse [Text]

A list of policy names.

lgprIsTruncated :: Lens' ListGroupPoliciesResponse Bool

A flag that indicates whether there are more policy names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policy names in the list.

lgprMarker :: Lens' ListGroupPoliciesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.