Network.AWS.IAM.ListGroups

Description

Lists the IAM groups that have the specified path prefix.

You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

listGroups :: ListGroups

Creates a value of ListGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgPathPrefix - The path prefix for filtering the results. For example, the prefix division_abcsubdivision_xyz/ gets all groups whose path starts with division_abcsubdivision_xyz/ . This parameter is optional. If it is not included, it defaults to a slash (), listing all groups. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lgMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lgMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

data ListGroups

See: listGroups smart constructor.

Request Lenses

lgPathPrefix :: Lens' ListGroups (Maybe Text)

The path prefix for filtering the results. For example, the prefix division_abcsubdivision_xyz/ gets all groups whose path starts with division_abcsubdivision_xyz/ . This parameter is optional. If it is not included, it defaults to a slash (), listing all groups. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lgMarker :: Lens' ListGroups (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lgMaxItems :: Lens' ListGroups (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Destructuring the Response

listGroupsResponse

Arguments

:: Int
lgrsResponseStatus

-> ListGroupsResponse

Creates a value of ListGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgrsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lgrsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lgrsResponseStatus - -- | The response status code.

lgrsGroups - A list of groups.

data ListGroupsResponse

Contains the response to a successful ListGroups request.

See: listGroupsResponse smart constructor.

Response Lenses

lgrsMarker :: Lens' ListGroupsResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lgrsIsTruncated :: Lens' ListGroupsResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lgrsResponseStatus :: Lens' ListGroupsResponse Int

- | The response status code.

lgrsGroups :: Lens' ListGroupsResponse [Group]

A list of groups.