Network.AWS.IAM.ListGroups

Description

Lists the groups that have the specified path prefix.

You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListGroups.html

Request

data ListGroups

Request constructor

listGroups :: ListGroups

ListGroups constructor.

The fields accessible through corresponding lenses are:

lgMarker :: Maybe Text

lgMaxItems :: Maybe Natural

lgPathPrefix :: Maybe Text

Request lenses

lgMarker :: Lens' ListGroups (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lgMaxItems :: Lens' ListGroups (Maybe Natural)

Use this only when paginating results to indicate the maximum number of groups you want in the response. If there are additional groups beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lgPathPrefix :: Lens' ListGroups (Maybe Text)

The path prefix for filtering the results. For example, the prefix '/division_abc/subdivision_xyz/' gets all groups whose path starts with '/division_abc/subdivision_xyz/'.

This parameter is optional. If it is not included, it defaults to a slash (/), listing all groups.

Response

data ListGroupsResponse

Response constructor

listGroupsResponse :: ListGroupsResponse

ListGroupsResponse constructor.

The fields accessible through corresponding lenses are:

lgrGroups :: [Group]

lgrIsTruncated :: Maybe Bool

lgrMarker :: Maybe Text

Response lenses

lgrGroups :: Lens' ListGroupsResponse [Group]

A list of groups.

lgrIsTruncated :: Lens' ListGroupsResponse (Maybe Bool)

A flag that indicates whether there are more groups to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more groups in the list.

lgrMarker :: Lens' ListGroupsResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.