Network.AWS.IAM.ListGroupsForUser

Description

Lists the IAM groups that the specified IAM user belongs to.

You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

listGroupsForUser

Arguments

:: Text
lgfuUserName

-> ListGroupsForUser

Creates a value of ListGroupsForUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgfuMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lgfuMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lgfuUserName - The name of the user to list groups for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data ListGroupsForUser

See: listGroupsForUser smart constructor.

Request Lenses

lgfuMarker :: Lens' ListGroupsForUser (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lgfuMaxItems :: Lens' ListGroupsForUser (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lgfuUserName :: Lens' ListGroupsForUser Text

The name of the user to list groups for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

listGroupsForUserResponse

Arguments

:: Int
lgfursResponseStatus

-> ListGroupsForUserResponse

Creates a value of ListGroupsForUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgfursMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lgfursIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lgfursResponseStatus - -- | The response status code.

lgfursGroups - A list of groups.

data ListGroupsForUserResponse

Contains the response to a successful ListGroupsForUser request.

See: listGroupsForUserResponse smart constructor.

Response Lenses

lgfursMarker :: Lens' ListGroupsForUserResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lgfursIsTruncated :: Lens' ListGroupsForUserResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lgfursResponseStatus :: Lens' ListGroupsForUserResponse Int

- | The response status code.

lgfursGroups :: Lens' ListGroupsForUserResponse [Group]

A list of groups.