Network.AWS.IAM.ListInstanceProfilesForRole

Description

Lists the instance profiles that have the specified associated IAM role. If there are none, the action returns an empty list. For more information about instance profiles, go to About Instance Profiles .

You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

listInstanceProfilesForRole

Arguments

:: Text
lipfrRoleName

-> ListInstanceProfilesForRole

Creates a value of ListInstanceProfilesForRole with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lipfrMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lipfrMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lipfrRoleName - The name of the role to list instance profiles for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

data ListInstanceProfilesForRole

See: listInstanceProfilesForRole smart constructor.

Request Lenses

lipfrMarker :: Lens' ListInstanceProfilesForRole (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lipfrMaxItems :: Lens' ListInstanceProfilesForRole (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lipfrRoleName :: Lens' ListInstanceProfilesForRole Text

The name of the role to list instance profiles for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

listInstanceProfilesForRoleResponse

Arguments

:: Int
lipfrrsResponseStatus

-> ListInstanceProfilesForRoleResponse

Creates a value of ListInstanceProfilesForRoleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lipfrrsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lipfrrsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lipfrrsResponseStatus - -- | The response status code.

lipfrrsInstanceProfiles - A list of instance profiles.

data ListInstanceProfilesForRoleResponse

Contains the response to a successful ListInstanceProfilesForRole request.

See: listInstanceProfilesForRoleResponse smart constructor.

Response Lenses

lipfrrsMarker :: Lens' ListInstanceProfilesForRoleResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lipfrrsIsTruncated :: Lens' ListInstanceProfilesForRoleResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lipfrrsResponseStatus :: Lens' ListInstanceProfilesForRoleResponse Int

- | The response status code.

lipfrrsInstanceProfiles :: Lens' ListInstanceProfilesForRoleResponse [InstanceProfile]

A list of instance profiles.