Network.AWS.IAM.ListMFADevices

Description

Lists the MFA devices. If the request includes the user name, then this action lists all the MFA devices associated with the specified user name. If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request.

You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListMFADevices.html

Request

data ListMFADevices

Request constructor

listMFADevices :: ListMFADevices

ListMFADevices constructor.

The fields accessible through corresponding lenses are:

lmfadMarker :: Maybe Text

lmfadMaxItems :: Maybe Natural

lmfadUserName :: Maybe Text

Request lenses

lmfadMarker :: Lens' ListMFADevices (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lmfadMaxItems :: Lens' ListMFADevices (Maybe Natural)

Use this only when paginating results to indicate the maximum number of MFA devices you want in the response. If there are additional MFA devices beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lmfadUserName :: Lens' ListMFADevices (Maybe Text)

The name of the user whose MFA devices you want to list.

Response

data ListMFADevicesResponse

Response constructor

listMFADevicesResponse :: ListMFADevicesResponse

ListMFADevicesResponse constructor.

The fields accessible through corresponding lenses are:

lmfadrIsTruncated :: Maybe Bool

lmfadrMFADevices :: [MFADevice]

lmfadrMarker :: Maybe Text

Response lenses

lmfadrIsTruncated :: Lens' ListMFADevicesResponse (Maybe Bool)

A flag that indicates whether there are more MFA devices to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more MFA devices in the list.

lmfadrMFADevices :: Lens' ListMFADevicesResponse [MFADevice]

A list of MFA devices.

lmfadrMarker :: Lens' ListMFADevicesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.