Network.AWS.IAM.ListPolicies

Description

Lists all the managed policies that are available in your AWS account, including your own customer-defined managed policies and all AWS managed policies.

You can filter the list of policies that is returned using the optional OnlyAttached , Scope , and PathPrefix parameters. For example, to list only the customer managed policies in your AWS account, set Scope to Local . To list only AWS managed policies, set Scope to AWS .

You can paginate the results using the MaxItems and Marker parameters.

For more information about managed policies, see Managed Policies and Inline Policies in the IAM User Guide .

This operation returns paginated results.

Creating a Request

listPolicies :: ListPolicies

Creates a value of ListPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpPathPrefix - The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (), listing all policies. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lpOnlyAttached - A flag to filter the results to only the attached policies. When OnlyAttached is true , the returned list contains only the policies that are attached to an IAM user, group, or role. When OnlyAttached is false , or when the parameter is not included, all policies are returned.

lpMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lpScope - The scope to use for filtering the results. To list only AWS managed policies, set Scope to AWS . To list only the customer managed policies in your AWS account, set Scope to Local . This parameter is optional. If it is not included, or if it is set to All , all policies are returned.

lpMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

data ListPolicies

See: listPolicies smart constructor.

Request Lenses

lpPathPrefix :: Lens' ListPolicies (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (), listing all policies. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lpOnlyAttached :: Lens' ListPolicies (Maybe Bool)

A flag to filter the results to only the attached policies. When OnlyAttached is true , the returned list contains only the policies that are attached to an IAM user, group, or role. When OnlyAttached is false , or when the parameter is not included, all policies are returned.

lpMarker :: Lens' ListPolicies (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lpScope :: Lens' ListPolicies (Maybe PolicyScopeType)

The scope to use for filtering the results. To list only AWS managed policies, set Scope to AWS . To list only the customer managed policies in your AWS account, set Scope to Local . This parameter is optional. If it is not included, or if it is set to All , all policies are returned.

lpMaxItems :: Lens' ListPolicies (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Destructuring the Response

listPoliciesResponse

Arguments

:: Int
lprsResponseStatus

-> ListPoliciesResponse

Creates a value of ListPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lprsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lprsPolicies - A list of policies.

lprsResponseStatus - -- | The response status code.

data ListPoliciesResponse

Contains the response to a successful ListPolicies request.

See: listPoliciesResponse smart constructor.

Response Lenses

lprsMarker :: Lens' ListPoliciesResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lprsIsTruncated :: Lens' ListPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lprsPolicies :: Lens' ListPoliciesResponse [Policy]

A list of policies.

lprsResponseStatus :: Lens' ListPoliciesResponse Int

- | The response status code.