Network.AWS.IAM.ListPolicyVersions

Description

Lists information about the versions of the specified managed policy, including the version that is currently set as the policy's default version.

For more information about managed policies, see Managed Policies and Inline Policies in the IAM User Guide .

This operation returns paginated results.

Creating a Request

listPolicyVersions

Arguments

:: Text
lpvPolicyARN

-> ListPolicyVersions

Creates a value of ListPolicyVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpvMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lpvMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lpvPolicyARN - The Amazon Resource Name (ARN) of the IAM policy for which you want the versions. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

data ListPolicyVersions

See: listPolicyVersions smart constructor.

Request Lenses

lpvMarker :: Lens' ListPolicyVersions (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lpvMaxItems :: Lens' ListPolicyVersions (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lpvPolicyARN :: Lens' ListPolicyVersions Text

The Amazon Resource Name (ARN) of the IAM policy for which you want the versions. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference .

Destructuring the Response

listPolicyVersionsResponse

Arguments

:: Int
lpvrsResponseStatus

-> ListPolicyVersionsResponse

Creates a value of ListPolicyVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpvrsVersions - A list of policy versions. For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

lpvrsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lpvrsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lpvrsResponseStatus - -- | The response status code.

data ListPolicyVersionsResponse

Contains the response to a successful ListPolicyVersions request.

See: listPolicyVersionsResponse smart constructor.

Response Lenses

lpvrsVersions :: Lens' ListPolicyVersionsResponse [PolicyVersion]

A list of policy versions. For more information about managed policy versions, see Versioning for Managed Policies in the IAM User Guide .

lpvrsMarker :: Lens' ListPolicyVersionsResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lpvrsIsTruncated :: Lens' ListPolicyVersionsResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lpvrsResponseStatus :: Lens' ListPolicyVersionsResponse Int

- | The response status code.