Network.AWS.IAM.ListRoles

Description

Lists the IAM roles that have the specified path prefix. If there are none, the action returns an empty list. For more information about roles, go to Working with Roles .

You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

listRoles :: ListRoles

Creates a value of ListRoles with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrPathPrefix - The path prefix for filtering the results. For example, the prefix application_abccomponent_xyz/ gets all roles whose path starts with application_abccomponent_xyz/ . This parameter is optional. If it is not included, it defaults to a slash (), listing all roles. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lrMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lrMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

data ListRoles

See: listRoles smart constructor.

Request Lenses

lrPathPrefix :: Lens' ListRoles (Maybe Text)

The path prefix for filtering the results. For example, the prefix application_abccomponent_xyz/ gets all roles whose path starts with application_abccomponent_xyz/ . This parameter is optional. If it is not included, it defaults to a slash (), listing all roles. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lrMarker :: Lens' ListRoles (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lrMaxItems :: Lens' ListRoles (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Destructuring the Response

listRolesResponse

Arguments

:: Int
lrrsResponseStatus

-> ListRolesResponse

Creates a value of ListRolesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrrsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lrrsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lrrsResponseStatus - -- | The response status code.

lrrsRoles - A list of roles.

data ListRolesResponse

Contains the response to a successful ListRoles request.

See: listRolesResponse smart constructor.

Response Lenses

lrrsMarker :: Lens' ListRolesResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lrrsIsTruncated :: Lens' ListRolesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lrrsResponseStatus :: Lens' ListRolesResponse Int

- | The response status code.

lrrsRoles :: Lens' ListRolesResponse [Role]

A list of roles.