Network.AWS.IAM.ListSSHPublicKeys

Description

Returns information about the SSH public keys associated with the specified IAM user. If there are none, the action returns an empty list.

The SSH public keys returned by this action are used only for authenticating the IAM user to an AWS CodeCommit repository. For more information about using SSH keys to authenticate to an AWS CodeCommit repository, see Set up AWS CodeCommit for SSH Connections in the AWS CodeCommit User Guide .

Although each user is limited to a small number of keys, you can still paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

listSSHPublicKeys :: ListSSHPublicKeys

Creates a value of ListSSHPublicKeys with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lspkUserName - The name of the IAM user to list SSH public keys for. If none is specified, the UserName field is determined implicitly based on the AWS access key used to sign the request. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

lspkMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lspkMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

data ListSSHPublicKeys

See: listSSHPublicKeys smart constructor.

Request Lenses

lspkUserName :: Lens' ListSSHPublicKeys (Maybe Text)

The name of the IAM user to list SSH public keys for. If none is specified, the UserName field is determined implicitly based on the AWS access key used to sign the request. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

lspkMarker :: Lens' ListSSHPublicKeys (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lspkMaxItems :: Lens' ListSSHPublicKeys (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Destructuring the Response

listSSHPublicKeysResponse

Arguments

:: Int
lspkrsResponseStatus

-> ListSSHPublicKeysResponse

Creates a value of ListSSHPublicKeysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lspkrsSSHPublicKeys - A list of the SSH public keys assigned to IAM user.

lspkrsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lspkrsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lspkrsResponseStatus - -- | The response status code.

data ListSSHPublicKeysResponse

Contains the response to a successful ListSSHPublicKeys request.

See: listSSHPublicKeysResponse smart constructor.

Response Lenses

lspkrsSSHPublicKeys :: Lens' ListSSHPublicKeysResponse [SSHPublicKeyMetadata]

A list of the SSH public keys assigned to IAM user.

lspkrsMarker :: Lens' ListSSHPublicKeysResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lspkrsIsTruncated :: Lens' ListSSHPublicKeysResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lspkrsResponseStatus :: Lens' ListSSHPublicKeysResponse Int

- | The response status code.