Network.AWS.IAM.ListServerCertificates

Description

Lists the server certificates stored in IAM that have the specified path prefix. If none exist, the action returns an empty list.

You can paginate the results using the MaxItems and Marker parameters.

For more information about working with server certificates, including a list of AWS services that can use the server certificates that you manage with IAM, go to Working with Server Certificates in the IAM User Guide .

This operation returns paginated results.

Creating a Request

listServerCertificates :: ListServerCertificates

Creates a value of ListServerCertificates with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lscPathPrefix - The path prefix for filtering the results. For example: companyservercerts would get all server certificates for which the path starts with companyservercerts . This parameter is optional. If it is not included, it defaults to a slash (), listing all server certificates. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lscMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lscMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

data ListServerCertificates

See: listServerCertificates smart constructor.

Request Lenses

lscPathPrefix :: Lens' ListServerCertificates (Maybe Text)

The path prefix for filtering the results. For example: companyservercerts would get all server certificates for which the path starts with companyservercerts . This parameter is optional. If it is not included, it defaults to a slash (), listing all server certificates. This paramater allows (per its <http:wikipedia.orgwikiregex regex pattern> ) a string of characters consisting of either a forward slash () by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

lscMarker :: Lens' ListServerCertificates (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lscMaxItems :: Lens' ListServerCertificates (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Destructuring the Response

listServerCertificatesResponse

Arguments

:: Int
lscrsResponseStatus

-> ListServerCertificatesResponse

Creates a value of ListServerCertificatesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lscrsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lscrsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lscrsResponseStatus - -- | The response status code.

lscrsServerCertificateMetadataList - A list of server certificates.

data ListServerCertificatesResponse

Contains the response to a successful ListServerCertificates request.

See: listServerCertificatesResponse smart constructor.

Response Lenses

lscrsMarker :: Lens' ListServerCertificatesResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lscrsIsTruncated :: Lens' ListServerCertificatesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lscrsResponseStatus :: Lens' ListServerCertificatesResponse Int

- | The response status code.

lscrsServerCertificateMetadataList :: Lens' ListServerCertificatesResponse [ServerCertificateMetadata]

A list of server certificates.