Network.AWS.IAM.ListServerCertificates

Description

Lists the server certificates that have the specified path prefix. If none exist, the action returns an empty list.

You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListServerCertificates.html

Request

data ListServerCertificates

Request constructor

listServerCertificates :: ListServerCertificates

ListServerCertificates constructor.

The fields accessible through corresponding lenses are:

lscMarker :: Maybe Text

lscMaxItems :: Maybe Natural

lscPathPrefix :: Maybe Text

Request lenses

lscMarker :: Lens' ListServerCertificates (Maybe Text)

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lscMaxItems :: Lens' ListServerCertificates (Maybe Natural)

Use this only when paginating results to indicate the maximum number of server certificates you want in the response. If there are additional server certificates beyond the maximum you specify, the IsTruncated response element will be set to true. This parameter is optional. If you do not include it, it defaults to 100.

lscPathPrefix :: Lens' ListServerCertificates (Maybe Text)

The path prefix for filtering the results. For example: '/company/servercerts' would get all server certificates for which the path starts with '/company/servercerts'.

This parameter is optional. If it is not included, it defaults to a slash (/), listing all server certificates.

Response

data ListServerCertificatesResponse

Response constructor

listServerCertificatesResponse :: ListServerCertificatesResponse

ListServerCertificatesResponse constructor.

The fields accessible through corresponding lenses are:

lscrIsTruncated :: Maybe Bool

lscrMarker :: Maybe Text

lscrServerCertificateMetadataList :: [ServerCertificateMetadata]

Response lenses

lscrIsTruncated :: Lens' ListServerCertificatesResponse (Maybe Bool)

A flag that indicates whether there are more server certificates to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more server certificates in the list.

lscrMarker :: Lens' ListServerCertificatesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lscrServerCertificateMetadataList :: Lens' ListServerCertificatesResponse [ServerCertificateMetadata]

A list of server certificates.