Network.AWS.IAM.ListSigningCertificates

Description

Returns information about the signing certificates associated with the specified IAM user. If there are none, the action returns an empty list.

Although each user is limited to a small number of signing certificates, you can still paginate the results using the MaxItems and Marker parameters.

If the UserName field is not specified, the user name is determined implicitly based on the AWS access key ID used to sign the request for this API. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

This operation returns paginated results.

Creating a Request

listSigningCertificates :: ListSigningCertificates

Creates a value of ListSigningCertificates with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lUserName - The name of the IAM user whose signing certificates you want to examine. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

lMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

data ListSigningCertificates

See: listSigningCertificates smart constructor.

Request Lenses

lUserName :: Lens' ListSigningCertificates (Maybe Text)

The name of the IAM user whose signing certificates you want to examine. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

lMarker :: Lens' ListSigningCertificates (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lMaxItems :: Lens' ListSigningCertificates (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Destructuring the Response

listSigningCertificatesResponse

Arguments

:: Int
lrsResponseStatus

-> ListSigningCertificatesResponse

Creates a value of ListSigningCertificatesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lrsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lrsResponseStatus - -- | The response status code.

lrsCertificates - A list of the user's signing certificate information.

data ListSigningCertificatesResponse

Contains the response to a successful ListSigningCertificates request.

See: listSigningCertificatesResponse smart constructor.

Response Lenses

lrsMarker :: Lens' ListSigningCertificatesResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lrsIsTruncated :: Lens' ListSigningCertificatesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lrsResponseStatus :: Lens' ListSigningCertificatesResponse Int

- | The response status code.

lrsCertificates :: Lens' ListSigningCertificatesResponse [SigningCertificate]

A list of the user's signing certificate information.