Network.AWS.IAM.ListUserPolicies

Description

Lists the names of the inline policies embedded in the specified IAM user.

An IAM user can also have managed policies attached to it. To list the managed policies that are attached to a user, use ListAttachedUserPolicies . For more information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

You can paginate the results using the MaxItems and Marker parameters. If there are no inline policies embedded with the specified user, the action returns an empty list.

This operation returns paginated results.

Creating a Request

listUserPolicies

Arguments

:: Text
lupUserName

-> ListUserPolicies

Creates a value of ListUserPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lupMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lupMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lupUserName - The name of the user to list policies for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data ListUserPolicies

See: listUserPolicies smart constructor.

Request Lenses

lupMarker :: Lens' ListUserPolicies (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lupMaxItems :: Lens' ListUserPolicies (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

lupUserName :: Lens' ListUserPolicies Text

The name of the user to list policies for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

listUserPoliciesResponse

Arguments

:: Int
luprsResponseStatus

-> ListUserPoliciesResponse

Creates a value of ListUserPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

luprsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

luprsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

luprsResponseStatus - -- | The response status code.

luprsPolicyNames - A list of policy names.

data ListUserPoliciesResponse

Contains the response to a successful ListUserPolicies request.

See: listUserPoliciesResponse smart constructor.

Response Lenses

luprsMarker :: Lens' ListUserPoliciesResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

luprsIsTruncated :: Lens' ListUserPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

luprsResponseStatus :: Lens' ListUserPoliciesResponse Int

- | The response status code.

luprsPolicyNames :: Lens' ListUserPoliciesResponse [Text]

A list of policy names.