Network.AWS.IAM.ListVirtualMFADevices

Description

Lists the virtual MFA devices defined in the AWS account by assignment status. If you do not specify an assignment status, the action returns a list of all virtual MFA devices. Assignment status can be Assigned , Unassigned , or Any .

You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

listVirtualMFADevices :: ListVirtualMFADevices

Creates a value of ListVirtualMFADevices with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lvmdAssignmentStatus - The status (Unassigned or Assigned ) of the devices to list. If you do not specify an AssignmentStatus , the action defaults to Any which lists both assigned and unassigned virtual MFA devices.

lvmdMarker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lvmdMaxItems - (Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

data ListVirtualMFADevices

See: listVirtualMFADevices smart constructor.

Request Lenses

lvmdAssignmentStatus :: Lens' ListVirtualMFADevices (Maybe AssignmentStatusType)

The status (Unassigned or Assigned ) of the devices to list. If you do not specify an AssignmentStatus , the action defaults to Any which lists both assigned and unassigned virtual MFA devices.

lvmdMarker :: Lens' ListVirtualMFADevices (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

lvmdMaxItems :: Lens' ListVirtualMFADevices (Maybe Natural)

(Optional) Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true . If you do not include this parameter, it defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true and Marker contains a value to include in the subsequent call that tells the service where to continue from.

Destructuring the Response

listVirtualMFADevicesResponse

Arguments

:: Int
lvmdrsResponseStatus

-> ListVirtualMFADevicesResponse

Creates a value of ListVirtualMFADevicesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lvmdrsMarker - When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lvmdrsIsTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lvmdrsResponseStatus - -- | The response status code.

lvmdrsVirtualMFADevices - The list of virtual MFA devices in the current account that match the AssignmentStatus value that was passed in the request.

data ListVirtualMFADevicesResponse

Contains the response to a successful ListVirtualMFADevices request.

See: listVirtualMFADevicesResponse smart constructor.

Response Lenses

lvmdrsMarker :: Lens' ListVirtualMFADevicesResponse (Maybe Text)

When IsTruncated is true , this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lvmdrsIsTruncated :: Lens' ListVirtualMFADevicesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all of your results.

lvmdrsResponseStatus :: Lens' ListVirtualMFADevicesResponse Int

- | The response status code.

lvmdrsVirtualMFADevices :: Lens' ListVirtualMFADevicesResponse [VirtualMFADevice]

The list of virtual MFA devices in the current account that match the AssignmentStatus value that was passed in the request.