Network.AWS.IAM.ListVirtualMFADevices

Description

Lists the virtual MFA devices under the AWS account by assignment status. If you do not specify an assignment status, the action returns a list of all virtual MFA devices. Assignment status can be Assigned, Unassigned, or Any.

You can paginate the results using the MaxItems and Marker parameters.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListVirtualMFADevices.html

Request

data ListVirtualMFADevices

Request constructor

listVirtualMFADevices :: ListVirtualMFADevices

ListVirtualMFADevices constructor.

The fields accessible through corresponding lenses are:

lvmfadAssignmentStatus :: Maybe AssignmentStatusType

lvmfadMarker :: Maybe Text

lvmfadMaxItems :: Maybe Natural

Request lenses

lvmfadAssignmentStatus :: Lens' ListVirtualMFADevices (Maybe AssignmentStatusType)

The status (unassigned or assigned) of the devices to list. If you do not specify an AssignmentStatus, the action defaults to Any which lists both assigned and unassigned virtual MFA devices.

lvmfadMarker :: Lens' ListVirtualMFADevices (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lvmfadMaxItems :: Lens' ListVirtualMFADevices (Maybe Natural)

Use this parameter only when paginating results to indicate the maximum number of MFA devices you want in the response. If there are additional MFA devices beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListVirtualMFADevicesResponse

Response constructor

listVirtualMFADevicesResponse :: ListVirtualMFADevicesResponse

ListVirtualMFADevicesResponse constructor.

The fields accessible through corresponding lenses are:

lvmfadrIsTruncated :: Maybe Bool

lvmfadrMarker :: Maybe Text

lvmfadrVirtualMFADevices :: [VirtualMFADevice]

Response lenses

lvmfadrIsTruncated :: Lens' ListVirtualMFADevicesResponse (Maybe Bool)

A flag that indicates whether there are more items to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items the list.

lvmfadrMarker :: Lens' ListVirtualMFADevicesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lvmfadrVirtualMFADevices :: Lens' ListVirtualMFADevicesResponse [VirtualMFADevice]

The list of virtual MFA devices in the current account that match the AssignmentStatus value that was passed in the request.