Network.AWS.IAM.PutGroupPolicy

Description

Adds or updates an inline policy document that is embedded in the specified IAM group.

A user can also have managed policies attached to it. To attach a managed policy to a group, use AttachGroupPolicy . To create a new managed policy, use CreatePolicy . For information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

For information about limits on the number of inline policies that you can embed in a group, see Limitations on IAM Entities in the IAM User Guide .

Creating a Request

putGroupPolicy

Arguments

:: Text
pgpGroupName

-> Text
pgpPolicyName

-> Text
pgpPolicyDocument

-> PutGroupPolicy

Creates a value of PutGroupPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pgpGroupName - The name of the group to associate the policy with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

pgpPolicyName - The name of the policy document. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

pgpPolicyDocument - The policy document. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

data PutGroupPolicy

See: putGroupPolicy smart constructor.

Request Lenses

pgpGroupName :: Lens' PutGroupPolicy Text

The name of the group to associate the policy with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

pgpPolicyName :: Lens' PutGroupPolicy Text

The name of the policy document. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

pgpPolicyDocument :: Lens' PutGroupPolicy Text

The policy document. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

Destructuring the Response

putGroupPolicyResponse :: PutGroupPolicyResponse

Creates a value of PutGroupPolicyResponse with the minimum fields required to make a request.

data PutGroupPolicyResponse

See: putGroupPolicyResponse smart constructor.