Network.AWS.IAM.PutRolePolicy

Description

Adds or updates an inline policy document that is embedded in the specified IAM role.

When you embed an inline policy in a role, the inline policy is used as part of the role's access (permissions) policy. The role's trust policy is created at the same time as the role, using CreateRole . You can update a role's trust policy using UpdateAssumeRolePolicy . For more information about IAM roles, go to Using Roles to Delegate Permissions and Federate Identities .

A role can also have a managed policy attached to it. To attach a managed policy to a role, use AttachRolePolicy . To create a new managed policy, use CreatePolicy . For information about policies, see Managed Policies and Inline Policies in the IAM User Guide .

For information about limits on the number of inline policies that you can embed with a role, see Limitations on IAM Entities in the IAM User Guide .

Creating a Request

putRolePolicy

Arguments

:: Text
prpRoleName

-> Text
prpPolicyName

-> Text
prpPolicyDocument

-> PutRolePolicy

Creates a value of PutRolePolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prpRoleName - The name of the role to associate the policy with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

prpPolicyName - The name of the policy document. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

prpPolicyDocument - The policy document. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

data PutRolePolicy

See: putRolePolicy smart constructor.

Request Lenses

prpRoleName :: Lens' PutRolePolicy Text

The name of the role to associate the policy with. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

prpPolicyName :: Lens' PutRolePolicy Text

The name of the policy document. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

prpPolicyDocument :: Lens' PutRolePolicy Text

The policy document. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

Destructuring the Response

putRolePolicyResponse :: PutRolePolicyResponse

Creates a value of PutRolePolicyResponse with the minimum fields required to make a request.

data PutRolePolicyResponse

See: putRolePolicyResponse smart constructor.