Network.AWS.IAM.RemoveRoleFromInstanceProfile

Description

Removes the specified role from the specified instance profile.

Make sure you do not have any Amazon EC2 instances running with the role you are about to remove from the instance profile. Removing a role from an instance profile that is associated with a running instance will break any applications running on the instance. For more information about roles, go to Working with Roles. For more information about instance profiles, go to About Instance Profiles.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_RemoveRoleFromInstanceProfile.html

Request

data RemoveRoleFromInstanceProfile

Request constructor

removeRoleFromInstanceProfile

Arguments

:: Text
rrfipInstanceProfileName

-> Text
rrfipRoleName

-> RemoveRoleFromInstanceProfile

RemoveRoleFromInstanceProfile constructor.

The fields accessible through corresponding lenses are:

rrfipInstanceProfileName :: Text

rrfipRoleName :: Text

Request lenses

rrfipInstanceProfileName :: Lens' RemoveRoleFromInstanceProfile Text

The name of the instance profile to update.

rrfipRoleName :: Lens' RemoveRoleFromInstanceProfile Text

The name of the role to remove.

Response

data RemoveRoleFromInstanceProfileResponse

Response constructor

removeRoleFromInstanceProfileResponse :: RemoveRoleFromInstanceProfileResponse

RemoveRoleFromInstanceProfileResponse constructor.