Network.AWS.IAM.Types

Description

AWS Identity and Access Management (IAM) enables you to securely control access to AWS services and resources for your users. Using IAM, you can create and manage AWS users and groups and use permissions to allow and deny their access to AWS resources.

Service

data IAM

Supported version (2010-05-08) of the AWS Identity and Access Management service.

Errors

data IAMError

A sum type representing possible errors returned by the IAM service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

CredentialReportExpiredException
The request was rejected because the most recent credential report has expired. To generate a new credential report, use GenerateCredentialReport. For more information about credential report expiration, see Getting Credential Reports in the Using IAM guide.

Fields

_creeMessage :: Maybe Text

CredentialReportNotPresentException
The request was rejected because the credential report does not exist. To generate a credential report, use GenerateCredentialReport.

Fields

_crnpeMessage :: Maybe Text

CredentialReportNotReadyException
The request was rejected because the credential report is still being generated.

Fields

_crnreMessage :: Maybe Text

DeleteConflictException
The request was rejected because it attempted to delete a resource that has attached subordinate entities. The error message describes these entities.

Fields

_dceMessage :: Maybe Text

DuplicateCertificateException
The request was rejected because the same certificate is associated to another user under the account.

Fields

_dce1Message :: Maybe Text

EntityAlreadyExistsException
The request was rejected because it attempted to create a resource that already exists.

Fields

_eaeeMessage :: Maybe Text

EntityTemporarilyUnmodifiableException
The request was rejected because it referenced an entity that is temporarily unmodifiable, such as a user name that was deleted and then recreated. The error indicates that the request is likely to succeed if you try again after waiting several minutes. The error message describes the entity.

Fields

_etueMessage :: Maybe Text

IAMClient HttpException

IAMSerializer String

IAMService String

InvalidAuthenticationCodeException
The request was rejected because the authentication code was not recognized. The error message describes the specific error.

Fields

_iaceMessage :: Maybe Text

InvalidCertificateException
The request was rejected because the certificate is invalid.

Fields

_iceMessage :: Maybe Text

InvalidInputException

Fields

_iieMessage :: Maybe Text

InvalidUserTypeException
The request was rejected because the type of user for the transaction was incorrect.

Fields

_iuteMessage :: Maybe Text

KeyPairMismatchException
The request was rejected because the public key certificate and the private key do not match.

Fields

_kpmeMessage :: Maybe Text

LimitExceededException
The request was rejected because it attempted to create resources beyond the current AWS account limits. The error message describes the limit exceeded.

Fields

_leeMessage :: Maybe Text

MalformedCertificateException
The request was rejected because the certificate was malformed or expired. The error message describes the specific error.

Fields

_mceMessage :: Maybe Text

MalformedPolicyDocumentException
The request was rejected because the policy document was malformed. The error message describes the specific error.

Fields

_mpdeMessage :: Maybe Text

NoSuchEntityException
The request was rejected because it referenced an entity that does not exist. The error message describes the entity.

Fields

_nseeMessage :: Maybe Text

PasswordPolicyViolationException
The request was rejected because the provided password did not meet the requirements imposed by the account password policy.

Fields

_ppveMessage :: Maybe Text

_CredentialReportExpiredException :: Prism' IAMError (Maybe Text)

The request was rejected because the most recent credential report has expired. To generate a new credential report, use GenerateCredentialReport. For more information about credential report expiration, see Getting Credential Reports in the Using IAM guide.

See: CredentialReportExpiredException

_CredentialReportNotPresentException :: Prism' IAMError (Maybe Text)

The request was rejected because the credential report does not exist. To generate a credential report, use GenerateCredentialReport.

See: CredentialReportNotPresentException

_CredentialReportNotReadyException :: Prism' IAMError (Maybe Text)

The request was rejected because the credential report is still being generated.

See: CredentialReportNotReadyException

_DeleteConflictException :: Prism' IAMError (Maybe Text)

The request was rejected because it attempted to delete a resource that has attached subordinate entities. The error message describes these entities.

See: DeleteConflictException

_DuplicateCertificateException :: Prism' IAMError (Maybe Text)

The request was rejected because the same certificate is associated to another user under the account.

See: DuplicateCertificateException

_EntityAlreadyExistsException :: Prism' IAMError (Maybe Text)

The request was rejected because it attempted to create a resource that already exists.

See: EntityAlreadyExistsException

_EntityTemporarilyUnmodifiableException :: Prism' IAMError (Maybe Text)

The request was rejected because it referenced an entity that is temporarily unmodifiable, such as a user name that was deleted and then recreated. The error indicates that the request is likely to succeed if you try again after waiting several minutes. The error message describes the entity.

See: EntityTemporarilyUnmodifiableException

_IAMClient :: Prism' IAMError HttpException

See: IAMClient

_IAMSerializer :: Prism' IAMError String

See: IAMSerializer

_IAMService :: Prism' IAMError String

See: IAMService

_InvalidAuthenticationCodeException :: Prism' IAMError (Maybe Text)

The request was rejected because the authentication code was not recognized. The error message describes the specific error.

See: InvalidAuthenticationCodeException

_InvalidCertificateException :: Prism' IAMError (Maybe Text)

The request was rejected because the certificate is invalid.

See: InvalidCertificateException

_InvalidInputException :: Prism' IAMError (Maybe Text)

See: InvalidInputException

_InvalidUserTypeException :: Prism' IAMError (Maybe Text)

The request was rejected because the type of user for the transaction was incorrect.

See: InvalidUserTypeException

_KeyPairMismatchException :: Prism' IAMError (Maybe Text)

The request was rejected because the public key certificate and the private key do not match.

See: KeyPairMismatchException

_LimitExceededException :: Prism' IAMError (Maybe Text)

The request was rejected because it attempted to create resources beyond the current AWS account limits. The error message describes the limit exceeded.

See: LimitExceededException

_MalformedCertificateException :: Prism' IAMError (Maybe Text)

The request was rejected because the certificate was malformed or expired. The error message describes the specific error.

See: MalformedCertificateException

_MalformedPolicyDocumentException :: Prism' IAMError (Maybe Text)

The request was rejected because the policy document was malformed. The error message describes the specific error.

See: MalformedPolicyDocumentException

_NoSuchEntityException :: Prism' IAMError (Maybe Text)

The request was rejected because it referenced an entity that does not exist. The error message describes the entity.

See: NoSuchEntityException

_PasswordPolicyViolationException :: Prism' IAMError (Maybe Text)

The request was rejected because the provided password did not meet the requirements imposed by the account password policy.

See: PasswordPolicyViolationException

XML

xmlOptions :: Tagged a XMLOptions

AssignmentStatusType

data AssignmentStatusType

Constructors

AssignmentStatusTypeAny
Any

AssignmentStatusTypeAssigned
Assigned

AssignmentStatusTypeUnassigned
Unassigned

ReportFormatType

data ReportFormatType

Constructors

ReportFormatTypeTextCsv
text/csv

ReportStateType

data ReportStateType

Constructors

ReportStateTypeComplete
COMPLETE

ReportStateTypeInprogress
INPROGRESS

ReportStateTypeStarted
STARTED

StatusType

data StatusType

Constructors

StatusTypeActive
Active

StatusTypeInactive
Inactive

SummaryKeyType

data SummaryKeyType

Constructors

SummaryKeyTypeAccessKeysPerUserQuota
AccessKeysPerUserQuota

SummaryKeyTypeAccountMFAEnabled
AccountMFAEnabled

SummaryKeyTypeGroupPolicySizeQuota
GroupPolicySizeQuota

SummaryKeyTypeGroups
Groups

SummaryKeyTypeGroupsPerUserQuota
GroupsPerUserQuota

SummaryKeyTypeGroupsQuota
GroupsQuota

SummaryKeyTypeMFADevices
MFADevices

SummaryKeyTypeMFADevicesInUse
MFADevicesInUse

SummaryKeyTypeServerCertificates
ServerCertificates

SummaryKeyTypeServerCertificatesQuota
ServerCertificatesQuota

SummaryKeyTypeSigningCertificatesPerUserQuota
SigningCertificatesPerUserQuota

SummaryKeyTypeUserPolicySizeQuota
UserPolicySizeQuota

SummaryKeyTypeUsers
Users

SummaryKeyTypeUsersQuota
UsersQuota

AccessKey

data AccessKey

Information about the access key.

accessKey

Arguments

:: Text
akUserName

-> Text
akAccessKeyId

-> StatusType
akStatus

-> Text
akSecretAccessKey

-> AccessKey

Smart constructor for the minimum required fields to construct a valid AccessKey data type.

AccessKey is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

UserName :: Text

AccessKeyId :: Text

Status :: StatusType

SecretAccessKey :: Text

CreateDate :: Maybe ISO8601

akUserName :: Lens' AccessKey Text

Name of the user the key is associated with.

akAccessKeyId :: Lens' AccessKey Text

The ID for this access key.

akStatus :: Lens' AccessKey StatusType

The status of the access key. Active means the key is valid for API calls, while Inactive means it is not.

akSecretAccessKey :: Lens' AccessKey Text

The secret key used to sign requests.

akCreateDate :: Lens' AccessKey (Maybe ISO8601)

The date when the access key was created.

AccessKeyMetadata

data AccessKeyMetadata

The AccessKey data type contains information about an AWS access key, without its secret key. This data type is used as a response element in the action ListAccessKeys.

accessKeyMetadata :: AccessKeyMetadata

Smart constructor for the minimum required fields to construct a valid AccessKeyMetadata data type.

AccessKeyMetadata is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

UserName :: Maybe Text

AccessKeyId :: Maybe Text

Status :: Maybe StatusType

CreateDate :: Maybe ISO8601

akmUserName :: Lens' AccessKeyMetadata (Maybe Text)

Name of the user the key is associated with.

akmAccessKeyId :: Lens' AccessKeyMetadata (Maybe Text)

The ID for this access key.

akmStatus :: Lens' AccessKeyMetadata (Maybe StatusType)

The status of the access key. Active means the key is valid for API calls, while Inactive means it is not.

akmCreateDate :: Lens' AccessKeyMetadata (Maybe ISO8601)

The date when the access key was created.

Group

data Group

Information about the group.

group

Arguments

:: Text
gPath

-> Text
gGroupName

-> Text
gGroupId

-> Text
gArn

-> ISO8601
gCreateDate

-> Group

Smart constructor for the minimum required fields to construct a valid Group data type.

Group is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Path :: Text

GroupName :: Text

GroupId :: Text

Arn :: Text

CreateDate :: ISO8601

gPath :: Lens' Group Text

Path to the group. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

gGroupName :: Lens' Group Text

The name that identifies the group.

gGroupId :: Lens' Group Text

The stable and unique string identifying the group. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

gArn :: Lens' Group Text

The Amazon Resource Name (ARN) specifying the group. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

gCreateDate :: Lens' Group ISO8601

The date when the group was created.

InstanceProfile

data InstanceProfile

Information about the instance profile.

instanceProfile

Arguments

:: Text
ipPath

-> Text
ipInstanceProfileName

-> Text
ipInstanceProfileId

-> Text
ipArn

-> ISO8601
ipCreateDate

-> [Role]
ipRoles

-> InstanceProfile

Smart constructor for the minimum required fields to construct a valid InstanceProfile data type.

InstanceProfile is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Path :: Text

InstanceProfileName :: Text

InstanceProfileId :: Text

Arn :: Text

CreateDate :: ISO8601

Roles :: [Role]

ipPath :: Lens' InstanceProfile Text

Path to the instance profile. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

ipInstanceProfileName :: Lens' InstanceProfile Text

The name identifying the instance profile.

ipInstanceProfileId :: Lens' InstanceProfile Text

The stable and unique string identifying the instance profile. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

ipArn :: Lens' InstanceProfile Text

The Amazon Resource Name (ARN) specifying the instance profile. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

ipCreateDate :: Lens' InstanceProfile ISO8601

The date when the instance profile was created.

ipRoles :: Lens' InstanceProfile [Role]

The role associated with the instance profile.

LoginProfile

data LoginProfile

The user name and password create date.

loginProfile

Arguments

:: Text
lpUserName

-> ISO8601
lpCreateDate

-> LoginProfile

Smart constructor for the minimum required fields to construct a valid LoginProfile data type.

LoginProfile is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

UserName :: Text

CreateDate :: ISO8601

PasswordResetRequired :: Maybe Bool

lpUserName :: Lens' LoginProfile Text

The name of the user, which can be used for signing in to the AWS Management Console.

lpCreateDate :: Lens' LoginProfile ISO8601

The date when the password for the user was created.

lpPasswordResetRequired :: Lens' LoginProfile (Maybe Bool)

Specifies whether the user is required to set a new password on next sign-in.

MFADevice

data MFADevice

The MFADevice data type contains information about an MFA device. This data type is used as a response element in the action ListMFADevices.

mFADevice

Arguments

:: Text
mfadUserName

-> Text
mfadSerialNumber

-> ISO8601
mfadEnableDate

-> MFADevice

Smart constructor for the minimum required fields to construct a valid MFADevice data type.

MFADevice is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

UserName :: Text

SerialNumber :: Text

EnableDate :: ISO8601

mfadUserName :: Lens' MFADevice Text

The user with whom the MFA device is associated.

mfadSerialNumber :: Lens' MFADevice Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

mfadEnableDate :: Lens' MFADevice ISO8601

The date when the MFA device was enabled for the user.

PasswordPolicy

data PasswordPolicy

The PasswordPolicy data type contains information about the account password policy. This data type is used as a response element in the action GetAccountPasswordPolicy.

passwordPolicy :: PasswordPolicy

Smart constructor for the minimum required fields to construct a valid PasswordPolicy data type.

PasswordPolicy is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

MinimumPasswordLength :: Maybe Integer

RequireSymbols :: Maybe Bool

RequireNumbers :: Maybe Bool

RequireUppercaseCharacters :: Maybe Bool

RequireLowercaseCharacters :: Maybe Bool

AllowUsersToChangePassword :: Maybe Bool

ExpirePasswords :: Maybe Bool

MaxPasswordAge :: Maybe Integer

PasswordReusePrevention :: Maybe Integer

HardExpiry :: Maybe Bool

ppMinimumPasswordLength :: Lens' PasswordPolicy (Maybe Integer)

Minimum length to require for IAM user passwords.

ppRequireSymbols :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether to require symbols for IAM user passwords.

ppRequireNumbers :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether to require numbers for IAM user passwords.

ppRequireUppercaseCharacters :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether to require uppercase characters for IAM user passwords.

ppRequireLowercaseCharacters :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether to require lowercase characters for IAM user passwords.

ppAllowUsersToChangePassword :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM users are allowed to change their own password.

ppExpirePasswords :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM users are required to change their password after a specified number of days.

ppMaxPasswordAge :: Lens' PasswordPolicy (Maybe Integer)

The number of days that an IAM user password is valid.

ppPasswordReusePrevention :: Lens' PasswordPolicy (Maybe Integer)

Specifies the number of previous passwords that IAM users are prevented from reusing.

ppHardExpiry :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM users are prevented from setting a new password after their password has expired.

Role

data Role

The Role data type contains information about a role. This data type is used as a response element in the following actions: CreateRole GetRole ListRoles.

role

Arguments

:: Text
rPath

-> Text
rRoleName

-> Text
rRoleId

-> Text
rArn

-> ISO8601
rCreateDate

-> Role

Smart constructor for the minimum required fields to construct a valid Role data type to populate a request.

The fields accessible through corresponding lenses are:

Path :: Text

RoleName :: Text

RoleId :: Text

Arn :: Text

CreateDate :: ISO8601

AssumeRolePolicyDocument :: Maybe Text

rPath :: Lens' Role Text

Path to the role. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

rRoleName :: Lens' Role Text

The name identifying the role.

rRoleId :: Lens' Role Text

The stable and unique string identifying the role. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

rArn :: Lens' Role Text

The Amazon Resource Name (ARN) specifying the role. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

rCreateDate :: Lens' Role ISO8601

The date when the role was created.

rAssumeRolePolicyDocument :: Lens' Role (Maybe Text)

The policy that grants an entity permission to assume the role. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcs/rfc3986.html.

SAMLProviderListEntry

data SAMLProviderListEntry

The list of SAML providers for this account.

sAMLProviderListEntry :: SAMLProviderListEntry

Smart constructor for the minimum required fields to construct a valid SAMLProviderListEntry data type.

SAMLProviderListEntry is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Arn :: Maybe Text

ValidUntil :: Maybe ISO8601

CreateDate :: Maybe ISO8601

samlpleArn :: Lens' SAMLProviderListEntry (Maybe Text)

The Amazon Resource Name (ARN) of the SAML provider.

samlpleValidUntil :: Lens' SAMLProviderListEntry (Maybe ISO8601)

The expiration date and time for the SAML provider.

samlpleCreateDate :: Lens' SAMLProviderListEntry (Maybe ISO8601)

The date and time when the SAML provider was created.

ServerCertificate

data ServerCertificate

Information about the server certificate.

serverCertificate

Arguments

:: ServerCertificateMetadata
scServerCertificateMetadata

-> Text
scCertificateBody

-> ServerCertificate

Smart constructor for the minimum required fields to construct a valid ServerCertificate data type.

ServerCertificate is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ServerCertificateMetadata :: ServerCertificateMetadata

CertificateBody :: Text

CertificateChain :: Maybe Text

scServerCertificateMetadata :: Lens' ServerCertificate ServerCertificateMetadata

The meta information of the server certificate, such as its name, path, ID, and ARN.

scCertificateBody :: Lens' ServerCertificate Text

The contents of the public key certificate.

scCertificateChain :: Lens' ServerCertificate (Maybe Text)

The contents of the public key certificate chain.

ServerCertificateMetadata

data ServerCertificateMetadata

The meta information of the server certificate, such as its name, path, ID, and ARN.

serverCertificateMetadata

Arguments

:: Text
scmPath

-> Text
scmServerCertificateName

-> Text
scmServerCertificateId

-> Text
scmArn

-> ServerCertificateMetadata

Smart constructor for the minimum required fields to construct a valid ServerCertificateMetadata data type.

ServerCertificateMetadata is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Path :: Text

ServerCertificateName :: Text

ServerCertificateId :: Text

Arn :: Text

UploadDate :: Maybe ISO8601

Expiration :: Maybe ISO8601

scmPath :: Lens' ServerCertificateMetadata Text

Path to the server certificate. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

scmServerCertificateName :: Lens' ServerCertificateMetadata Text

The name that identifies the server certificate.

scmServerCertificateId :: Lens' ServerCertificateMetadata Text

The stable and unique string identifying the server certificate. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

scmArn :: Lens' ServerCertificateMetadata Text

The Amazon Resource Name (ARN) specifying the server certificate. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

scmUploadDate :: Lens' ServerCertificateMetadata (Maybe ISO8601)

The date when the server certificate was uploaded.

scmExpiration :: Lens' ServerCertificateMetadata (Maybe ISO8601)

The date on which the certificate is set to expire.

SigningCertificate

data SigningCertificate

The SigningCertificate data type contains information about an X.509 signing certificate. This data type is used as a response element in the actions UploadSigningCertificate and ListSigningCertificates.

signingCertificate

Arguments

:: Text
scrUserName

-> Text
scrCertificateId

-> Text
scrCertificateBody

-> StatusType
scrStatus

-> SigningCertificate

Smart constructor for the minimum required fields to construct a valid SigningCertificate data type.

SigningCertificate is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

UserName :: Text

CertificateId :: Text

CertificateBody :: Text

Status :: StatusType

UploadDate :: Maybe ISO8601

scrUserName :: Lens' SigningCertificate Text

Name of the user the signing certificate is associated with.

scrCertificateId :: Lens' SigningCertificate Text

The ID for the signing certificate.

scrCertificateBody :: Lens' SigningCertificate Text

The contents of the signing certificate.

scrStatus :: Lens' SigningCertificate StatusType

The status of the signing certificate. Active means the key is valid for API calls, while Inactive means it is not.

scrUploadDate :: Lens' SigningCertificate (Maybe ISO8601)

The date when the signing certificate was uploaded.

User

data User

Information about the user.

user

Arguments

:: Text
uPath

-> Text
uUserName

-> Text
uUserId

-> Text
uArn

-> ISO8601
uCreateDate

-> User

Smart constructor for the minimum required fields to construct a valid User data type to populate a request.

The fields accessible through corresponding lenses are:

Path :: Text

UserName :: Text

UserId :: Text

Arn :: Text

CreateDate :: ISO8601

uPath :: Lens' User Text

Path to the user. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide.

uUserName :: Lens' User Text

The name identifying the user.

uUserId :: Lens' User Text

The stable and unique string identifying the user. For more information about IDs, see Identifiers for IAM Entities in the Using IAM guide.

uArn :: Lens' User Text

The Amazon Resource Name (ARN) specifying the user. For more information about ARNs and how to use them in policies, see Identifiers for IAM Entities in the Using IAM guide.

uCreateDate :: Lens' User ISO8601

The date when the user was created.

VirtualMFADevice

data VirtualMFADevice

A newly created virtual MFA device.

virtualMFADevice

Arguments

:: Text
vmfadSerialNumber

-> VirtualMFADevice

Smart constructor for the minimum required fields to construct a valid VirtualMFADevice data type.

VirtualMFADevice is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SerialNumber :: Text

Base32StringSeed :: Maybe ByteString

QRCodePNG :: Maybe ByteString

User :: Maybe User

EnableDate :: Maybe ISO8601

vmfadSerialNumber :: Lens' VirtualMFADevice Text

The serial number associated with VirtualMFADevice.

vmfadBase32StringSeed :: Lens' VirtualMFADevice (Maybe ByteString)

The Base32 seed defined as specified in RFC3548. The Base32StringSeed is Base64-encoded.

vmfadQRCodePNG :: Lens' VirtualMFADevice (Maybe ByteString)

A QR code PNG image that encodes otpauth:/totp$virtualMFADeviceName@$AccountName? secret=$Base32String where $virtualMFADeviceName is one of the create call arguments, AccountName is the user name if set (accountId otherwise), and Base32String is the seed in Base32 format. The Base32String is Base64-encoded.

vmfadUser :: Lens' VirtualMFADevice (Maybe User)

The User data type contains information about a user. This data type is used as a response element in the following actions: CreateUser GetUser ListUsers.

vmfadEnableDate :: Lens' VirtualMFADevice (Maybe ISO8601)