Network.AWS.IAM.UpdateAccountPasswordPolicy

Description

Updates the password policy settings for the AWS account.

For more information about using a password policy, see Managing an IAM Password Policy in the IAM User Guide .

Creating a Request

updateAccountPasswordPolicy :: UpdateAccountPasswordPolicy

Creates a value of UpdateAccountPasswordPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uappMinimumPasswordLength - The minimum number of characters allowed in an IAM user password. Default value: 6

uappRequireNumbers - Specifies whether IAM user passwords must contain at least one numeric character (0 to 9). Default value: false

uappPasswordReusePrevention - Specifies the number of previous passwords that IAM users are prevented from reusing. The default value of 0 means IAM users are not prevented from reusing previous passwords. Default value: 0

uappRequireLowercaseCharacters - Specifies whether IAM user passwords must contain at least one lowercase character from the ISO basic Latin alphabet (a to z). Default value: false

uappMaxPasswordAge - The number of days that an IAM user password is valid. The default value of 0 means IAM user passwords never expire. Default value: 0

uappHardExpiry - Prevents IAM users from setting a new password after their password has expired. Default value: false

uappRequireSymbols - Specifies whether IAM user passwords must contain at least one of the following non-alphanumeric characters: ! @ # $ % ^ &amp; * ( ) _ + - = [ ] { } | ' Default value: false

uappRequireUppercaseCharacters - Specifies whether IAM user passwords must contain at least one uppercase character from the ISO basic Latin alphabet (A to Z). Default value: false

uappAllowUsersToChangePassword - Allows all IAM users in your account to use the AWS Management Console to change their own passwords. For more information, see Letting IAM Users Change Their Own Passwords in the IAM User Guide . Default value: false

data UpdateAccountPasswordPolicy

See: updateAccountPasswordPolicy smart constructor.

Request Lenses

uappMinimumPasswordLength :: Lens' UpdateAccountPasswordPolicy (Maybe Natural)

The minimum number of characters allowed in an IAM user password. Default value: 6

uappRequireNumbers :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one numeric character (0 to 9). Default value: false

uappPasswordReusePrevention :: Lens' UpdateAccountPasswordPolicy (Maybe Natural)

Specifies the number of previous passwords that IAM users are prevented from reusing. The default value of 0 means IAM users are not prevented from reusing previous passwords. Default value: 0

uappRequireLowercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one lowercase character from the ISO basic Latin alphabet (a to z). Default value: false

uappMaxPasswordAge :: Lens' UpdateAccountPasswordPolicy (Maybe Natural)

The number of days that an IAM user password is valid. The default value of 0 means IAM user passwords never expire. Default value: 0

uappHardExpiry :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Prevents IAM users from setting a new password after their password has expired. Default value: false

uappRequireSymbols :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one of the following non-alphanumeric characters: ! @ # $ % ^ &amp; * ( ) _ + - = [ ] { } | ' Default value: false

uappRequireUppercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one uppercase character from the ISO basic Latin alphabet (A to Z). Default value: false

uappAllowUsersToChangePassword :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Allows all IAM users in your account to use the AWS Management Console to change their own passwords. For more information, see Letting IAM Users Change Their Own Passwords in the IAM User Guide . Default value: false

Destructuring the Response

updateAccountPasswordPolicyResponse :: UpdateAccountPasswordPolicyResponse

Creates a value of UpdateAccountPasswordPolicyResponse with the minimum fields required to make a request.

data UpdateAccountPasswordPolicyResponse

See: updateAccountPasswordPolicyResponse smart constructor.