Network.AWS.IAM.UpdateAccountPasswordPolicy

Description

Updates the password policy settings for the AWS account.

This action does not support partial updates. No parameters are required, but if you do not specify a parameter, that parameter's value reverts to its default value. See the Request Parameters section for each parameter's default value.

For more information about using a password policy, see Managing an IAMPassword Policy in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_UpdateAccountPasswordPolicy.html

Request

data UpdateAccountPasswordPolicy

Request constructor

updateAccountPasswordPolicy :: UpdateAccountPasswordPolicy

UpdateAccountPasswordPolicy constructor.

The fields accessible through corresponding lenses are:

uappAllowUsersToChangePassword :: Maybe Bool

uappHardExpiry :: Maybe Bool

uappMaxPasswordAge :: Maybe Natural

uappMinimumPasswordLength :: Maybe Natural

uappPasswordReusePrevention :: Maybe Natural

uappRequireLowercaseCharacters :: Maybe Bool

uappRequireNumbers :: Maybe Bool

uappRequireSymbols :: Maybe Bool

uappRequireUppercaseCharacters :: Maybe Bool

Request lenses

uappAllowUsersToChangePassword :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Allows all IAM users in your account to use the AWS Management Console to change their own passwords. For more information, see Letting IAM UsersChange Their Own Passwords in the Using IAM guide.

Default value: false

uappHardExpiry :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Prevents IAM users from setting a new password after their password has expired.

Default value: false

uappMaxPasswordAge :: Lens' UpdateAccountPasswordPolicy (Maybe Natural)

The number of days that an IAM user password is valid. The default value of 0 means IAM user passwords never expire.

Default value: 0

uappMinimumPasswordLength :: Lens' UpdateAccountPasswordPolicy (Maybe Natural)

The minimum number of characters allowed in an IAM user password.

Default value: 6

uappPasswordReusePrevention :: Lens' UpdateAccountPasswordPolicy (Maybe Natural)

Specifies the number of previous passwords that IAM users are prevented from reusing. The default value of 0 means IAM users are not prevented from reusing previous passwords.

Default value: 0

uappRequireLowercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one lowercase character from the ISO basic Latin alphabet (a to z).

Default value: false

uappRequireNumbers :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).

Default value: false

uappRequireSymbols :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one of the following non-alphanumeric characters:

! @ # $ % ^ &amp; * ( ) _ + - = [ ] { } | '

Default value: false

uappRequireUppercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one uppercase character from the ISO basic Latin alphabet (A to Z).

Default value: false

Response

data UpdateAccountPasswordPolicyResponse

Response constructor

updateAccountPasswordPolicyResponse :: UpdateAccountPasswordPolicyResponse

UpdateAccountPasswordPolicyResponse constructor.