Network.AWS.IAM.UpdateGroup

Description

Updates the name and/or the path of the specified IAM group.

Important: You should understand the implications of changing a group's path or name. For more information, see Renaming Users and Groups in the IAM User Guide .

Creating a Request

updateGroup

Arguments

:: Text
ugGroupName

-> UpdateGroup

Creates a value of UpdateGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ugNewGroupName - New name for the IAM group. Only include this if changing the group's name. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

ugNewPath - New path for the IAM group. Only include this if changing the group's path. This paramater allows (per its regex pattern ) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

ugGroupName - Name of the IAM group to update. If you're changing the name of the group, this is the original name. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data UpdateGroup

See: updateGroup smart constructor.

Request Lenses

ugNewGroupName :: Lens' UpdateGroup (Maybe Text)

New name for the IAM group. Only include this if changing the group's name. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

ugNewPath :: Lens' UpdateGroup (Maybe Text)

New path for the IAM group. Only include this if changing the group's path. This paramater allows (per its regex pattern ) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

ugGroupName :: Lens' UpdateGroup Text

Name of the IAM group to update. If you're changing the name of the group, this is the original name. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

updateGroupResponse :: UpdateGroupResponse

Creates a value of UpdateGroupResponse with the minimum fields required to make a request.

data UpdateGroupResponse

See: updateGroupResponse smart constructor.