Network.AWS.IAM.UpdateSSHPublicKey

Description

Sets the status of an IAM user's SSH public key to active or inactive. SSH public keys that are inactive cannot be used for authentication. This action can be used to disable a user's SSH public key as part of a key rotation work flow.

The SSH public key affected by this action is used only for authenticating the associated IAM user to an AWS CodeCommit repository. For more information about using SSH keys to authenticate to an AWS CodeCommit repository, see Set up AWS CodeCommit for SSH Connections in the AWS CodeCommit User Guide .

Creating a Request

updateSSHPublicKey

Arguments

:: Text
uspkUserName

-> Text
uspkSSHPublicKeyId

-> StatusType
uspkStatus

-> UpdateSSHPublicKey

Creates a value of UpdateSSHPublicKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uspkUserName - The name of the IAM user associated with the SSH public key. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uspkSSHPublicKeyId - The unique identifier for the SSH public key. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

uspkStatus - The status to assign to the SSH public key. Active means the key can be used for authentication with an AWS CodeCommit repository. Inactive means the key cannot be used.

data UpdateSSHPublicKey

See: updateSSHPublicKey smart constructor.

Request Lenses

uspkUserName :: Lens' UpdateSSHPublicKey Text

The name of the IAM user associated with the SSH public key. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uspkSSHPublicKeyId :: Lens' UpdateSSHPublicKey Text

The unique identifier for the SSH public key. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

uspkStatus :: Lens' UpdateSSHPublicKey StatusType

The status to assign to the SSH public key. Active means the key can be used for authentication with an AWS CodeCommit repository. Inactive means the key cannot be used.

Destructuring the Response

updateSSHPublicKeyResponse :: UpdateSSHPublicKeyResponse

Creates a value of UpdateSSHPublicKeyResponse with the minimum fields required to make a request.

data UpdateSSHPublicKeyResponse

See: updateSSHPublicKeyResponse smart constructor.