Network.AWS.IAM.UpdateServerCertificate

Description

Updates the name and/or the path of the specified server certificate stored in IAM.

For more information about working with server certificates, including a list of AWS services that can use the server certificates that you manage with IAM, go to Working with Server Certificates in the IAM User Guide .

Important: You should understand the implications of changing a server certificate's path or name. For more information, see Renaming a Server Certificate in the IAM User Guide .

Creating a Request

updateServerCertificate

Arguments

:: Text
uServerCertificateName

-> UpdateServerCertificate

Creates a value of UpdateServerCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uNewServerCertificateName - The new name for the server certificate. Include this only if you are updating the server certificate's name. The name of the certificate cannot contain any spaces. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uNewPath - The new path for the server certificate. Include this only if you are updating the server certificate's path. This paramater allows (per its regex pattern ) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

uServerCertificateName - The name of the server certificate that you want to update. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data UpdateServerCertificate

See: updateServerCertificate smart constructor.

Request Lenses

uNewServerCertificateName :: Lens' UpdateServerCertificate (Maybe Text)

The new name for the server certificate. Include this only if you are updating the server certificate's name. The name of the certificate cannot contain any spaces. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uNewPath :: Lens' UpdateServerCertificate (Maybe Text)

The new path for the server certificate. Include this only if you are updating the server certificate's path. This paramater allows (per its regex pattern ) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

uServerCertificateName :: Lens' UpdateServerCertificate Text

The name of the server certificate that you want to update. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

updateServerCertificateResponse :: UpdateServerCertificateResponse

Creates a value of UpdateServerCertificateResponse with the minimum fields required to make a request.

data UpdateServerCertificateResponse

See: updateServerCertificateResponse smart constructor.