Network.AWS.IAM.UpdateServerCertificate

Description

Updates the name and/or the path of the specified server certificate.

You should understand the implications of changing a server certificate's path or name. For more information, see Managing Server Certificates in the Using IAM guide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_UpdateServerCertificate.html

Request

data UpdateServerCertificate

Request constructor

updateServerCertificate

Arguments

:: Text
usc1ServerCertificateName

-> UpdateServerCertificate

UpdateServerCertificate constructor.

The fields accessible through corresponding lenses are:

usc1NewPath :: Maybe Text

usc1NewServerCertificateName :: Maybe Text

usc1ServerCertificateName :: Text

Request lenses

usc1NewPath :: Lens' UpdateServerCertificate (Maybe Text)

The new path for the server certificate. Include this only if you are updating the server certificate's path.

usc1NewServerCertificateName :: Lens' UpdateServerCertificate (Maybe Text)

The new name for the server certificate. Include this only if you are updating the server certificate's name.

usc1ServerCertificateName :: Lens' UpdateServerCertificate Text

The name of the server certificate that you want to update.

Response

data UpdateServerCertificateResponse

Response constructor

updateServerCertificateResponse :: UpdateServerCertificateResponse

UpdateServerCertificateResponse constructor.