Network.AWS.IAM.UpdateSigningCertificate

Description

Changes the status of the specified user signing certificate from active to disabled, or vice versa. This action can be used to disable an IAM user's signing certificate as part of a certificate rotation work flow.

If the UserName field is not specified, the UserName is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

Creating a Request

updateSigningCertificate

Arguments

:: Text
uscCertificateId

-> StatusType
uscStatus

-> UpdateSigningCertificate

Creates a value of UpdateSigningCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uscUserName - The name of the IAM user the signing certificate belongs to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uscCertificateId - The ID of the signing certificate you want to update. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

uscStatus - The status you want to assign to the certificate. Active means the certificate can be used for API calls to AWS, while Inactive means the certificate cannot be used.

data UpdateSigningCertificate

See: updateSigningCertificate smart constructor.

Request Lenses

uscUserName :: Lens' UpdateSigningCertificate (Maybe Text)

The name of the IAM user the signing certificate belongs to. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uscCertificateId :: Lens' UpdateSigningCertificate Text

The ID of the signing certificate you want to update. This parameter allows (per its regex pattern ) a string of characters that can consist of any upper or lowercased letter or digit.

uscStatus :: Lens' UpdateSigningCertificate StatusType

The status you want to assign to the certificate. Active means the certificate can be used for API calls to AWS, while Inactive means the certificate cannot be used.

Destructuring the Response

updateSigningCertificateResponse :: UpdateSigningCertificateResponse

Creates a value of UpdateSigningCertificateResponse with the minimum fields required to make a request.

data UpdateSigningCertificateResponse

See: updateSigningCertificateResponse smart constructor.