Network.AWS.IAM.UpdateUser

Description

Updates the name and/or the path of the specified IAM user.

Important: You should understand the implications of changing an IAM user's path or name. For more information, see Renaming an IAM User and Renaming an IAM Group in the IAM User Guide .

Creating a Request

updateUser

Arguments

:: Text
uuUserName

-> UpdateUser

Creates a value of UpdateUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uuNewUserName - New name for the user. Include this parameter only if you're changing the user's name. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uuNewPath - New path for the IAM user. Include this parameter only if you're changing the user's path. This paramater allows (per its regex pattern ) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

uuUserName - Name of the user to update. If you're changing the name of the user, this is the original user name. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

data UpdateUser

See: updateUser smart constructor.

Request Lenses

uuNewUserName :: Lens' UpdateUser (Maybe Text)

New name for the user. Include this parameter only if you're changing the user's name. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uuNewPath :: Lens' UpdateUser (Maybe Text)

New path for the IAM user. Include this parameter only if you're changing the user's path. This paramater allows (per its regex pattern ) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes, containing any ASCII character from the ! (u0021) thru the DEL character (u007F), including most punctuation characters, digits, and upper and lowercased letters.

uuUserName :: Lens' UpdateUser Text

Name of the user to update. If you're changing the name of the user, this is the original user name. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

Destructuring the Response

updateUserResponse :: UpdateUserResponse

Creates a value of UpdateUserResponse with the minimum fields required to make a request.

data UpdateUserResponse

See: updateUserResponse smart constructor.