Network.AWS.IAM.UpdateUser

Description

Updates the name and/or the path of the specified user.

You should understand the implications of changing a user's path or name. For more information, see Renaming Users and Groups in the Using IAM guide. To change a user name the requester must have appropriate permissions on both the source object and the target object. For example, to change Bob to Robert, the entity making the request must have permission on Bob and Robert, or must have permission on all (*). For more information about permissions, see Permissions and Policies.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_UpdateUser.html

Request

data UpdateUser

Request constructor

updateUser

Arguments

:: Text
uuUserName

-> UpdateUser

UpdateUser constructor.

The fields accessible through corresponding lenses are:

uuNewPath :: Maybe Text

uuNewUserName :: Maybe Text

uuUserName :: Text

Request lenses

uuNewPath :: Lens' UpdateUser (Maybe Text)

New path for the user. Include this parameter only if you're changing the user's path.

uuNewUserName :: Lens' UpdateUser (Maybe Text)

New name for the user. Include this parameter only if you're changing the user's name.

uuUserName :: Lens' UpdateUser Text

Name of the user to update. If you're changing the name of the user, this is the original user name.

Response

data UpdateUserResponse

Response constructor

updateUserResponse :: UpdateUserResponse

UpdateUserResponse constructor.