Network.AWS.IAM.UploadSigningCertificate

Description

Uploads an X.509 signing certificate and associates it with the specified IAM user. Some AWS services use X.509 signing certificates to validate requests that are signed with a corresponding private key. When you upload the certificate, its default status is Active .

If the UserName field is not specified, the IAM user name is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

Creating a Request

uploadSigningCertificate

Arguments

:: Text
uplCertificateBody

-> UploadSigningCertificate

Creates a value of UploadSigningCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uplUserName - The name of the user the signing certificate is for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uplCertificateBody - The contents of the signing certificate. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

data UploadSigningCertificate

See: uploadSigningCertificate smart constructor.

Request Lenses

uplUserName :: Lens' UploadSigningCertificate (Maybe Text)

The name of the user the signing certificate is for. This parameter allows (per its regex pattern ) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@-

uplCertificateBody :: Lens' UploadSigningCertificate Text

The contents of the signing certificate. The regex pattern used to validate this parameter is a string of characters consisting of any printable ASCII character ranging from the space character (u0020) through end of the ASCII character range as well as the printable characters in the Basic Latin and Latin-1 Supplement character set (through u00FF). It also includes the special characters tab (u0009), line feed (u000A), and carriage return (u000D).

Destructuring the Response

uploadSigningCertificateResponse

Arguments

:: Int
uscrsResponseStatus

-> SigningCertificate
uscrsCertificate

-> UploadSigningCertificateResponse

Creates a value of UploadSigningCertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uscrsResponseStatus - -- | The response status code.

uscrsCertificate - Information about the certificate.

data UploadSigningCertificateResponse

Contains the response to a successful UploadSigningCertificate request.

See: uploadSigningCertificateResponse smart constructor.

Response Lenses

uscrsResponseStatus :: Lens' UploadSigningCertificateResponse Int

- | The response status code.

uscrsCertificate :: Lens' UploadSigningCertificateResponse SigningCertificate

Information about the certificate.