Network.AWS.IAM.UploadSigningCertificate

Description

Uploads an X.509 signing certificate and associates it with the specified user. Some AWS services use X.509 signing certificates to validate requests that are signed with a corresponding private key. When you upload the certificate, its default status is Active.

If the UserName field is not specified, the user name is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, you can use this action to manage root credentials even if the AWS account has no associated users.

Because the body of a X.509 certificate can be large, you should use POST rather than GET when calling UploadSigningCertificate. For information about setting up signatures and authorization through the API, go to Signing AWSAPI Requests in the AWS General Reference. For general information about using the Query API with IAM, go to Making Query Requests in the Using IAMguide.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_UploadSigningCertificate.html

Request

data UploadSigningCertificate

Request constructor

uploadSigningCertificate

Arguments

:: Text
usc1CertificateBody

-> UploadSigningCertificate

UploadSigningCertificate constructor.

The fields accessible through corresponding lenses are:

usc1CertificateBody :: Text

usc1UserName :: Maybe Text

Request lenses

usc1CertificateBody :: Lens' UploadSigningCertificate Text

The contents of the signing certificate.

usc1UserName :: Lens' UploadSigningCertificate (Maybe Text)

The name of the user the signing certificate is for.

Response

data UploadSigningCertificateResponse

Response constructor

uploadSigningCertificateResponse

Arguments

:: SigningCertificate
uscrCertificate

-> UploadSigningCertificateResponse

UploadSigningCertificateResponse constructor.

The fields accessible through corresponding lenses are:

uscrCertificate :: SigningCertificate

Response lenses

uscrCertificate :: Lens' UploadSigningCertificateResponse SigningCertificate

Information about the certificate.