Network.AWS.IAM.V2010_05_08.CreateInstanceProfile

Description

Creates a new instance profile. For information about instance profiles, go to About Instance Profiles. For information about the number of instance profiles you can create, see Limitations on IAM Entities in the Using IAM guide. https:/iam.amazonaws.com ?Action=CreateInstanceProfile &InstanceProfileName=Webserver &Path=application_abccomponent_xyz/ &Version=2010-05-08 &AUTHPARAMS AIPAD5ARO2C5EXAMPLE3G Webserver application_abccomponent_xyz/ arn:aws:iam::123456789012:instance-profileapplication_abccomponent_xyz/Webserver 2012-05-09T16:11:10.222Z 974142ee-99f1-11e1-a4c3-27EXAMPLE804.

Request

data CreateInstanceProfile

Request constructor

mkCreateInstanceProfile

Arguments

:: Text
cipInstanceProfileName

-> CreateInstanceProfile

Smart constructor for the minimum required parameters to construct a valid CreateInstanceProfile request.

Request lenses

cipInstanceProfileName :: Lens' CreateInstanceProfile Text

Name of the instance profile to create.

cipPath :: Lens' CreateInstanceProfile (Maybe Text)

The path to the instance profile. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide. This parameter is optional. If it is not included, it defaults to a slash (/).

Response

data CreateInstanceProfileResponse

Contains the result of a successful invocation of the CreateInstanceProfile action.

Response lenses

ciprsInstanceProfile :: Lens' CreateInstanceProfileResponse InstanceProfile

Information about the instance profile.