Network.AWS.IAM.V2010_05_08.GetAccountSummary

Description

Retrieves account level information about account entity usage and IAM quotas. For information about limitations on IAM entities, see Limitations on IAM Entities in the Using IAM guide. https:/iam.amazonaws.com ?Action=GetAccountSummary &Version=2010-05-08 &AUTHPARAMS Groups 31 GroupsQuota 50 UsersQuota 150 Users 35 GroupPolicySizeQuota 10240 AccessKeysPerUserQuota 2 GroupsPerUserQuota 10 UserPolicySizeQuota 10240 SigningCertificatesPerUserQuota 2 ServerCertificates 0 ServerCertificatesQuota 10 AccountMFAEnabled 0 MFADevicesInUse 10 MFADevices 20 f1e38443-f1ad-11df-b1ef-a9265EXAMPLE.

Documentation

data GetAccountSummary

Constructors

GetAccountSummary

data GetAccountSummaryResponse

Constructors

GetAccountSummaryResponse

Fields

_gasrSummaryMap :: Map SummaryKeyType Integer

A set of key value pairs containing account-level information. SummaryMap contains the following keys: AccessKeysPerUserQuota - Maximum number of access keys that can be created per user AccountMFAEnabled - 1 if the root account has an MFA device assigned to it, 0 otherwise AssumeRolePolicySizeQuota - Maximum allowed size for assume role policy documents (in kilobytes) GroupPolicySizeQuota - Maximum allowed size for Group policy documents (in kilobytes) Groups - Number of Groups for the AWS account GroupsPerUserQuota - Maximum number of groups a user can belong to GroupsQuota - Maximum groups allowed for the AWS account InstanceProfiles - Number of instance profiles for the AWS account InstanceProfilesQuota - Maximum instance profiles allowed for the AWS account MFADevices - Number of MFA devices, either assigned or unassigned MFADevicesInUse - Number of MFA devices that have been assigned to an IAM user or to the root account RolePolicySizeQuota - Maximum allowed size for role policy documents (in kilobytes) Roles - Number of roles for the AWS account RolesQuota - Maximum roles allowed for the AWS account ServerCertificates - Number of server certificates for the AWS account ServerCertificatesQuota - Maximum server certificates allowed for the AWS account SigningCertificatesPerUserQuota - Maximum number of X509 certificates allowed for a user UserPolicySizeQuota - Maximum allowed size for user policy documents (in kilobytes) Users - Number of users for the AWS account UsersQuota - Maximum users allowed for the AWS account.

gasrSummaryMap :: Iso' GetAccountSummaryResponse (Map SummaryKeyType Integer)