Network.AWS.IAM.V2010_05_08.GetGroupPolicy

Description

Retrieves the specified policy document for the specified group. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcs/rfc3986.html. https:/iam.amazonaws.com ?Action=GetGroupPolicy &GroupName=Admins &PolicyName=AdminRoot &AUTHPARAMS Admins AdminRoot {Version:"2012-10-17",Statement:[{Effect:Allow,Action:"*",Resource:"*"}]} 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data GetGroupPolicy

Request constructor

mkGetGroupPolicy

Arguments

:: Text
ggpGroupName

-> Text
ggpPolicyName

-> GetGroupPolicy

Smart constructor for the minimum required parameters to construct a valid GetGroupPolicy request.

Request lenses

ggpGroupName :: Lens' GetGroupPolicy Text

Name of the group the policy is associated with.

ggpPolicyName :: Lens' GetGroupPolicy Text

Name of the policy document to get.

Response

data GetGroupPolicyResponse

Contains the result of a successful invocation of the GetGroupPolicy action.

Response lenses

ggprGroupName :: Lens' GetGroupPolicyResponse Text

The group the policy is associated with.

ggprPolicyName :: Lens' GetGroupPolicyResponse Text

The name of the policy.

ggprPolicyDocument :: Lens' GetGroupPolicyResponse Text

The policy document.