Network.AWS.IAM.V2010_05_08.GetUserPolicy

Description

Retrieves the specified policy document for the specified user. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcs/rfc3986.html. https:/iam.amazonaws.com ?Action=GetUserPolicy &UserName=Bob &PolicyName=AllAccessPolicy &AUTHPARAMS Bob AllAccessPolicy {Version:"2012-10-17",Statement:[{Effect:Allow,Action:"*",Resource:"*"}]} 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data GetUserPolicy

Request constructor

mkGetUserPolicy

Arguments

:: Text
gupUserName

-> Text
gupPolicyName

-> GetUserPolicy

Smart constructor for the minimum required parameters to construct a valid GetUserPolicy request.

Request lenses

gupUserName :: Lens' GetUserPolicy Text

Name of the user who the policy is associated with.

gupPolicyName :: Lens' GetUserPolicy Text

Name of the policy document to get.

Response

data GetUserPolicyResponse

Contains the result of a successful invocation of the GetUserPolicy action.

Response lenses

guprUserName :: Lens' GetUserPolicyResponse Text

The user the policy is associated with.

guprPolicyName :: Lens' GetUserPolicyResponse Text

The name of the policy.

guprPolicyDocument :: Lens' GetUserPolicyResponse Text

The policy document.