Network.AWS.IAM.V2010_05_08.ListAccessKeys

Description

Returns information about the access key IDs associated with the specified user. If there are none, the action returns an empty list. Although each user is limited to a small number of keys, you can still paginate the results using the MaxItems and Marker parameters. If the UserName field is not specified, the UserName is determined implicitly based on the AWS access key ID used to sign the request. Because this action works for access keys under the AWS account, this API can be used to manage root credentials even if the AWS account has no associated users. To ensure the security of your AWS account, the secret access key is accessible only during key and user creation. https:/iam.amazonaws.com ?Action=ListAccessKeys &UserName=Bob &Version=2010-05-08 &AUTHPARAMS Bob Bob AKIAIOSFODNN7EXAMPLE Active Bob AKIAI44QH8DHBEXAMPLE Inactive false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Request

data ListAccessKeys

Request constructor

mkListAccessKeys :: ListAccessKeys

Smart constructor for the minimum required parameters to construct a valid ListAccessKeys request.

Request lenses

lakUserName :: Lens' ListAccessKeys (Maybe Text)

Name of the user.

lakMarker :: Lens' ListAccessKeys (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lakMaxItems :: Lens' ListAccessKeys (Maybe Integer)

Use this parameter only when paginating results to indicate the maximum number of keys you want in the response. If there are additional keys beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListAccessKeysResponse

Contains the result of a successful invocation of the ListAccessKeys action.

Response lenses

lakrAccessKeyMetadata :: Lens' ListAccessKeysResponse [AccessKeyMetadata]

A list of access key metadata.

lakrIsTruncated :: Lens' ListAccessKeysResponse Bool

A flag that indicates whether there are more keys to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more keys in the list.

lakrMarker :: Lens' ListAccessKeysResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.