Network.AWS.IAM.V2010_05_08.ListInstanceProfiles

Description

Lists the instance profiles that have the specified path prefix. If there are none, the action returns an empty list. For more information about instance profiles, go to About Instance Profiles. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListInstanceProfiles &MaxItems=100 &PathPrefix=application_abc &Version=2010-05-08 &AUTHPARAMS false AIPACIFN4OZXG7EXAMPLE Database application_abccomponent_xyz/ arn:aws:iam::123456789012:instance-profileapplication_abccomponent_xyz/Database 2012-05-09T16:27:03Z AIPACZLSXM2EYYEXAMPLE Webserver application_abccomponent_xyz/ arn:aws:iam::123456789012:instance-profileapplication_abccomponent_xyz/Webserver 2012-05-09T16:27:11Z fd74fa8d-99f3-11e1-a4c3-27EXAMPLE804.

Request

data ListInstanceProfiles

Request constructor

mkListInstanceProfiles :: ListInstanceProfiles

Smart constructor for the minimum required parameters to construct a valid ListInstanceProfiles request.

Request lenses

lipPathPrefix :: Lens' ListInstanceProfiles (Maybe Text)

The path prefix for filtering the results. For example: application_abccomponent_xyz/, which would get all instance profiles whose path starts with application_abccomponent_xyz/. This parameter is optional. If it is not included, it defaults to a slash (/), listing all instance profiles.

lipMarker :: Lens' ListInstanceProfiles (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lipMaxItems :: Lens' ListInstanceProfiles (Maybe Integer)

Use this parameter only when paginating results to indicate the maximum number of user names you want in the response. If there are additional user names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListInstanceProfilesResponse

Contains the result of a successful invocation of the ListInstanceProfiles action.

Response constructor

mkListInstanceProfilesResponse

Arguments

:: [InstanceProfile]
liprInstanceProfiles

-> Bool
liprIsTruncated

-> ListInstanceProfilesResponse

Smart constructor for the minimum required parameters to construct a valid ListInstanceProfilesResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

liprInstanceProfiles :: Lens' ListInstanceProfilesResponse [InstanceProfile]

A list of instance profiles.

liprIsTruncated :: Lens' ListInstanceProfilesResponse Bool

A flag that indicates whether there are more instance profiles to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more instance profiles in the list.

liprMarker :: Lens' ListInstanceProfilesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.