Network.AWS.IAM.V2010_05_08.ListRolePolicies

Description

Lists the names of the policies associated with the specified role. If there are none, the action returns an empty list. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListRolePolicies &MaxItems=100 &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS CloudwatchPutMetricPolicy S3AccessPolicy false 8c7e1816-99f0-11e1-a4c3-27EXAMPLE804.

Request

data ListRolePolicies

Request constructor

mkListRolePolicies

Arguments

:: Text
lrpRoleName

-> ListRolePolicies

Smart constructor for the minimum required parameters to construct a valid ListRolePolicies request.

Request lenses

lrpRoleName :: Lens' ListRolePolicies Text

The name of the role to list policies for.

lrpMarker :: Lens' ListRolePolicies (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

lrpMaxItems :: Lens' ListRolePolicies (Maybe Integer)

Use this parameter only when paginating results to indicate the maximum number of user names you want in the response. If there are additional user names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

Response

data ListRolePoliciesResponse

Contains the result of a successful invocation of the ListRolePolicies action.

Response lenses

lrprPolicyNames :: Lens' ListRolePoliciesResponse [Text]

A list of policy names.

lrprIsTruncated :: Lens' ListRolePoliciesResponse Bool

A flag that indicates whether there are more policy names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policy names in the list.

lrprMarker :: Lens' ListRolePoliciesResponse (Maybe Text)

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.