Network.AWS.IAM.V2010_05_08.ListUsers

Description

Lists the users that have the specified path prefix. If there are none, the action returns an empty list. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListUsers &PathPrefix=division_abcsubdivision_xyzproduct_1234engineering/ &Version=2010-05-08 &AUTHPARAMS division_abcsubdivision_xyzengineering Andrew AID2MAB8DPLSRHEXAMPLE arn:aws:iam::123456789012:user division_abcsubdivision_xyzengineeringAndrew division_abcsubdivision_xyzengineering Jackie AIDIODR4TAW7CSEXAMPLE arn:aws:iam::123456789012:user division_abcsubdivision_xyzengineeringJackie false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

listUsers :: ListUsers

Minimum specification for a ListUsers request.

data ListUsers

Constructors

ListUsers

Fields

_lurMarker :: Maybe Text

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

_lurMaxItems :: Maybe Integer

Use this parameter only when paginating results to indicate the maximum number of user names you want in the response. If there are additional user names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

_lurPathPrefix :: Maybe Text

The path prefix for filtering the results. For example: division_abcsubdivision_xyz/, which would get all user names whose path starts with division_abcsubdivision_xyz/. This parameter is optional. If it is not included, it defaults to a slash (/), listing all user names.

lurPathPrefix :: Lens' ListUsers (Maybe Text)

lurMaxItems :: Lens' ListUsers (Maybe Integer)

lurMarker :: Lens' ListUsers (Maybe Text)

data ListUsersResponse

Constructors

ListUsersResponse

Fields

_lusIsTruncated :: Bool

A flag that indicates whether there are more user names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more users in the list.

_lusUsers :: [User]

A list of users.

_lusMarker :: Maybe Text

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lusUsers :: Lens' ListUsersResponse [User]

lusMarker :: Lens' ListUsersResponse (Maybe Text)

lusIsTruncated :: Lens' ListUsersResponse Bool