Network.AWS.IAM.V2010_05_08.UpdateAssumeRolePolicy

Description

Updates the policy that grants an entity permission to assume a role. For more information about roles, go to Working with Roles. https:/iam.amazonaws.com ?Action=UpdateAssumeRolePolicy &PolicyDocument={Version:"2012-10-17",Statement:[{Effect:Allow,Principal:{Service:["ec2.amazonaws.com"]},Action:["sts:AssumeRole"]}]} &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS 309c1671-99ed-11e1-a4c3-270EXAMPLE04.

Request

data UpdateAssumeRolePolicy

Request constructor

mkUpdateAssumeRolePolicy

Arguments

:: Text
uarpRoleName

-> Text
uarpPolicyDocument

-> UpdateAssumeRolePolicy

Smart constructor for the minimum required parameters to construct a valid UpdateAssumeRolePolicy request.

Request lenses

uarpRoleName :: Lens' UpdateAssumeRolePolicy Text

Name of the role to update.

uarpPolicyDocument :: Lens' UpdateAssumeRolePolicy Text

The policy that grants an entity permission to assume the role.

Response

data UpdateAssumeRolePolicyResponse

Response constructor

mkUpdateAssumeRolePolicyResponse :: UpdateAssumeRolePolicyResponse

Smart constructor for the minimum required parameters to construct a valid UpdateAssumeRolePolicyResponse response.

This constructor is provided for convenience and testing purposes.